test(stack): simplify managed contract scaffolding
