+++ packages/stack/src/managed-stack-contract.integration.test.ts
@@ -14,7 +14,6 @@ import { createStack } from "./node.ts";
 import {
   managedNativeServiceMatrix,
   managedStackContractFixtures,
-  type ManagedStackContractFact,
   type ManagedStackContractScenario,
   validateManagedStackContractFixtures,
 } from "./testing.ts";
@@ -25,3821 +24,108 @@ describe("managed stack acceptance contract", () => {
     expect(validateManagedStackContractFixtures(managedStackContractFixtures)).toEqual([]);
   });
 
-  it("rejects contract edits that make IDs, effects, and projections disagree", () => {
-    const scenario: ManagedStackContractScenario | undefined = managedStackContractFixtures.find(
-      ({ id }) => id === "identity.return-to-branch-reuses-stack",
-    );
-    if (scenario === undefined) {
-      throw new Error("identity.return-to-branch-reuses-stack fixture is required");
-    }
-    if (
-      scenario.expected.selection === undefined ||
-      scenario.expected.output.human === undefined ||
-      scenario.expected.output.json === undefined
-    ) {
-      throw new Error("identity.return-to-branch-reuses-stack must select and project a stack");
-    }
-
-    const missingStartWrite = {
-      ...scenario,
-      expected: { ...scenario.expected, writes: [] },
+  it("lints structural, referential, effect, and projection mistakes", () => {
+    const findScenario = (id: string): ManagedStackContractScenario => {
+      const scenario: ManagedStackContractScenario | undefined = managedStackContractFixtures.find(
+        (candidate) => candidate.id === id,
+      );
+      if (scenario === undefined) {
+        throw new Error(`${id} fixture is required`);
+      }
+      return scenario;
     };
-    expect(validateManagedStackContractFixtures([missingStartWrite])).toContain(
-      `${scenario.id}: start runtime effect requires a matching state write`,
-    );
 
-    const undeclaredSelection = {
-      ...scenario,
-      expected: {
-        ...scenario.expected,
-        selection: { ...scenario.expected.selection, stackId: "stack-undeclared" },
-      },
-    };
-    expect(validateManagedStackContractFixtures([undeclaredSelection])).toContain(
-      `${scenario.id}: selection references undeclared ID stack-undeclared`,
-    );
-
-    const independentBranchesScenario: ManagedStackContractScenario | undefined =
-      managedStackContractFixtures.find(
-        ({ id }) => id === "identity.same-commit-different-branches-are-independent",
-      );
+    const reuse = findScenario("identity.return-to-branch-reuses-stack");
+    const portConflict = findScenario("ports.explicit-port-conflict-fails");
+    const readOnly = findScenario("identity.branch-copy-read-only-does-not-write");
     if (
-      independentBranchesScenario === undefined ||
-      independentBranchesScenario.expected.selection === undefined
+      reuse.expected.selection === undefined ||
+      reuse.expected.output.json === undefined ||
+      portConflict.expected.error === undefined ||
+      portConflict.expected.output.json === undefined
     ) {
-      throw new Error(
-        "identity.same-commit-different-branches-are-independent selection is required",
-      );
+      throw new Error("lint examples require selected and structured fixture outputs");
     }
-    const selectionWithWrongContext = {
-      ...independentBranchesScenario,
-      expected: {
-        ...independentBranchesScenario.expected,
-        selection: {
-          ...independentBranchesScenario.expected.selection,
-          contextId: "context-main",
-        },
-        output: {
-          ...independentBranchesScenario.expected.output,
-          api: {
-            ...independentBranchesScenario.expected.output.api,
-            contextId: "context-main",
-          },
-        },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([selectionWithWrongContext])).toContain(
-      `${independentBranchesScenario.id}: selected stack stack-feat-default belongs to context context-feat, not context-main`,
-    );
 
-    const selectionWithWrongName = {
-      ...independentBranchesScenario,
-      expected: {
-        ...independentBranchesScenario.expected,
-        selection: {
-          ...independentBranchesScenario.expected.selection,
-          stackName: "review",
-        },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([selectionWithWrongName])).toContain(
-      `${independentBranchesScenario.id}: selected stack stack-feat-default is named default, not review`,
-    );
-
-    const linkedWorktreeScenario: ManagedStackContractScenario | undefined =
-      managedStackContractFixtures.find(
-        ({ id }) => id === "identity.linked-worktrees-share-project-not-checkout",
-      );
-    if (
-      linkedWorktreeScenario?.expected.selection === undefined ||
-      linkedWorktreeScenario.expected.output.api === undefined
-    ) {
-      throw new Error("identity.linked-worktrees-share-project-not-checkout fixture is required");
-    }
-    const siblingCheckoutSelection = {
-      ...linkedWorktreeScenario,
-      expected: {
-        ...linkedWorktreeScenario.expected,
-        selection: { ...linkedWorktreeScenario.expected.selection, checkoutId: "checkout-a" },
-        output: {
-          ...linkedWorktreeScenario.expected.output,
-          api: { ...linkedWorktreeScenario.expected.output.api, checkoutId: "checkout-a" },
-        },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([siblingCheckoutSelection])).toContain(
-      `${linkedWorktreeScenario.id}: selection must use checkout checkout-b for worktree-b`,
-    );
-
-    const namedStackScenario: ManagedStackContractScenario | undefined =
-      managedStackContractFixtures.find(
-        ({ id }) => id === "identity.named-stacks-are-context-scoped",
-      );
-    if (namedStackScenario === undefined || namedStackScenario.when.interface !== "cli") {
-      throw new Error("identity.named-stacks-are-context-scoped fixture is required");
-    }
-    const actionSelectingDefaultStack = {
-      ...namedStackScenario,
-      when: {
-        ...namedStackScenario.when,
-        argv: namedStackScenario.when.argv.map((arg) => (arg === "review" ? "default" : arg)),
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([actionSelectingDefaultStack])).toContain(
-      `${namedStackScenario.id}: explicit stack name default disagrees with selected stack review`,
-    );
-
-    const undeclaredWrite = {
-      ...scenario,
-      expected: {
-        ...scenario.expected,
-        writes: [{ target: "runtime-state", operation: "start", id: "stack-undeclared" }],
-        runtimeEffects: [{ operation: "start", stackId: "stack-undeclared" }],
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([undeclaredWrite])).toContain(
-      `${scenario.id}: runtime-state start references undeclared ID stack-undeclared`,
-    );
-
-    const divergentProjection = {
-      ...scenario,
-      expected: {
-        ...scenario.expected,
-        output: {
-          ...scenario.expected.output,
-          json: { ...scenario.expected.output.json, outcome: "create" },
-        },
-      },
-    };
-    expect(validateManagedStackContractFixtures([divergentProjection])).toContain(
-      `${scenario.id}: projected outcome disagrees with the managed result`,
-    );
-
-    const divergentHumanProjection = {
-      ...scenario,
-      expected: {
-        ...scenario.expected,
-        output: {
-          ...scenario.expected.output,
-          human: {
-            ...scenario.expected.output.human,
-            fields: { ...scenario.expected.output.human.fields, stackId: "stack-other" },
-          },
-        },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([divergentHumanProjection])).toContain(
-      `${scenario.id}: projected stackId disagrees with the managed result`,
-    );
-
-    const existingTarget: ManagedStackContractFact = {
-      kind: "managed-target",
-      stackId: "stack-main-default",
-      exists: true,
-    };
-    const ambiguousExistingStart = {
-      ...scenario,
-      given: [...scenario.given.filter(({ kind }) => kind !== "stack"), existingTarget],
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([ambiguousExistingStart])).toContain(
-      `${scenario.id}: starting existing stack stack-main-default requires an explicit stopped lifecycle`,
-    );
-
-    const trackedMarkerScenario = managedStackContractFixtures.find(
-      ({ id }) => id === "identity.fresh-clone-ignores-tracked-marker",
-    );
-    if (trackedMarkerScenario === undefined) {
-      throw new Error("identity.fresh-clone-ignores-tracked-marker fixture is required");
-    }
-    const identityMarkerWrite = {
-      target: "identity-marker",
-      operation: "create",
-      id: "marker-project-a",
-      storage: "project-local-untracked",
-      workspacePath: "checkout-a",
-      projectId: "project-a",
-      checkoutId: "checkout-a",
-      contextId: "context-main",
-    } satisfies ManagedStackContractScenario["expected"]["writes"][number];
-    const trackedMarkerMutation = {
-      ...trackedMarkerScenario,
-      expected: {
-        ...trackedMarkerScenario.expected,
-        writes: [...trackedMarkerScenario.expected.writes, identityMarkerWrite],
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([trackedMarkerMutation])).toContain(
-      `${trackedMarkerScenario.id}: a tracked identity marker must remain untouched`,
-    );
-
-    const gitWorkspaceScenario: ManagedStackContractScenario | undefined =
-      managedStackContractFixtures.find(
-        ({ id }) => id === "identity.fresh-clone-creates-project-and-checkout",
-      );
-    if (gitWorkspaceScenario === undefined) {
-      throw new Error("identity.fresh-clone-creates-project-and-checkout fixture is required");
-    }
-    const gitWorkspaceMarkerMutation = {
-      ...gitWorkspaceScenario,
-      expected: {
-        ...gitWorkspaceScenario.expected,
-        writes: [...gitWorkspaceScenario.expected.writes, identityMarkerWrite],
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([gitWorkspaceMarkerMutation])).toContain(
-      `${gitWorkspaceScenario.id}: Git workspace identity must use Git-local metadata`,
-    );
-
-    if (
-      gitWorkspaceScenario.expected.selection === undefined ||
-      gitWorkspaceScenario.expected.output.json === undefined
-    ) {
-      throw new Error("identity.fresh-clone-creates-project-and-checkout selection is required");
-    }
-    const selectionUsingAbsentProject = {
-      ...gitWorkspaceScenario,
-      expected: {
-        ...gitWorkspaceScenario.expected,
-        selection: { ...gitWorkspaceScenario.expected.selection, projectId: "project-a" },
-        output: {
-          ...gitWorkspaceScenario.expected.output,
-          json: { ...gitWorkspaceScenario.expected.output.json, project_id: "project-a" },
-        },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([selectionUsingAbsentProject])).toContain(
-      `${gitWorkspaceScenario.id}: selection references undeclared ID project-a`,
-    );
-
-    const absentLegacyScenario = managedStackContractFixtures.find(
-      ({ id }) => id === "bootstrap.absent-legacy-starts-fresh",
-    );
-    if (absentLegacyScenario === undefined) {
-      throw new Error("bootstrap.absent-legacy-starts-fresh fixture is required");
-    }
-    const unpublishedManagedState = {
-      ...absentLegacyScenario,
-      expected: {
-        ...absentLegacyScenario.expected,
-        writes: absentLegacyScenario.expected.writes.filter(({ target }) => target !== "registry"),
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([unpublishedManagedState])).toContain(
-      `${absentLegacyScenario.id}: managed-state create requires registry publication`,
-    );
-
-    const publishedWithoutState = {
-      ...absentLegacyScenario,
-      expected: {
-        ...absentLegacyScenario.expected,
-        writes: absentLegacyScenario.expected.writes.filter(
-          ({ target }) => target !== "managed-state",
-        ),
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([publishedWithoutState])).toContain(
-      `${absentLegacyScenario.id}: registry publication requires managed-state creation or copy`,
-    );
-
-    const targetlessStateWrites = {
-      ...absentLegacyScenario,
-      expected: {
-        ...absentLegacyScenario.expected,
-        writes: absentLegacyScenario.expected.writes.map((write) =>
-          write.target === "managed-state" || write.target === "registry"
-            ? { ...write, id: "" }
-            : write,
-        ),
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([targetlessStateWrites])).toContain(
-      `${absentLegacyScenario.id}: managed-state write requires a target ID`,
-    );
-
-    const targetlessRuntimeEffect = {
-      ...absentLegacyScenario,
-      expected: {
-        ...absentLegacyScenario.expected,
-        runtimeEffects: absentLegacyScenario.expected.runtimeEffects.map((effect) => ({
-          ...effect,
-          stackId: "",
-        })),
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([targetlessRuntimeEffect])).toContain(
-      `${absentLegacyScenario.id}: start runtime effect requires a stack ID`,
-    );
-
-    const stoppedStackScenario = managedStackContractFixtures.find(
-      ({ id }) => id === "reclamation.default-stop-preserves-data",
-    );
-    if (stoppedStackScenario === undefined) {
-      throw new Error("reclamation.default-stop-preserves-data fixture is required");
-    }
-    const missingStopEffect = {
-      ...stoppedStackScenario,
-      expected: { ...stoppedStackScenario.expected, runtimeEffects: [] },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([missingStopEffect])).toContain(
-      `${stoppedStackScenario.id}: runtime-state update requires a matching runtime effect`,
-    );
-
-    const folderToGitScenario = managedStackContractFixtures.find(
-      ({ id }) => id === "identity.folder-to-git-exact-claim-preserves-identity",
-    );
-    if (folderToGitScenario === undefined) {
-      throw new Error("identity.folder-to-git-exact-claim-preserves-identity fixture is required");
-    }
-    const incompleteGitIdentity = {
-      ...folderToGitScenario,
-      expected: {
-        ...folderToGitScenario.expected,
-        writes: folderToGitScenario.expected.writes.filter(({ id }) => id !== "project-a"),
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([incompleteGitIdentity])).toContain(
-      `${folderToGitScenario.id}: folder-to-Git identity project-a must be persisted in Git-local metadata`,
-    );
-
-    const incorrectlyScopedGitIdentity = {
-      ...folderToGitScenario,
-      expected: {
-        ...folderToGitScenario.expected,
-        writes: folderToGitScenario.expected.writes.map((write) =>
-          write.target === "git-config" && write.id === "project-a"
-            ? { ...write, scope: "worktree" }
-            : write,
-        ),
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([incorrectlyScopedGitIdentity])).toContain(
-      `${folderToGitScenario.id}: Git identity project-a must use common config scope`,
-    );
-
-    const qualificationScenario = managedStackContractFixtures.find(
-      ({ id }) => id === "native-qualification.all-services-qualify-platform",
-    );
-    if (qualificationScenario === undefined) {
-      throw new Error("native-qualification.all-services-qualify-platform fixture is required");
-    }
-    const incompleteQualification = {
-      ...qualificationScenario,
-      given: qualificationScenario.given.map((fact) =>
-        fact.kind === "native-qualification"
-          ? { ...fact, qualifiedServices: fact.qualifiedServices.slice(1) }
-          : fact,
-      ),
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([incompleteQualification])).toContain(
-      `${qualificationScenario.id}: native qualification omits service postgres`,
-    );
-
-    const qualificationForDifferentPlatform = {
-      ...qualificationScenario,
-      given: qualificationScenario.given.map((fact) =>
-        fact.kind === "native-qualification" ? { ...fact, platform: "linux-amd64" } : fact,
-      ),
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([qualificationForDifferentPlatform])).toContain(
-      `${qualificationScenario.id}: native qualification platform must match the preflight action`,
-    );
-
-    const qualificationForUnknownPlatform = {
-      ...qualificationScenario,
-      given: qualificationScenario.given.map((fact) =>
-        fact.kind === "native-qualification" ? { ...fact, platform: "solaris-sparc" } : fact,
-      ),
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([qualificationForUnknownPlatform])).toContain(
-      `${qualificationScenario.id}: native qualification uses unknown platform solaris-sparc`,
-    );
-
-    const failedQualificationScenario: ManagedStackContractScenario | undefined =
-      managedStackContractFixtures.find(
-        ({ id }) => id === "native-qualification.one-service-failure-disables-platform",
-      );
-    if (failedQualificationScenario === undefined) {
-      throw new Error(
-        "native-qualification.one-service-failure-disables-platform fixture is required",
-      );
-    }
-    const qualificationPartitionsContradictResult = {
-      ...failedQualificationScenario,
-      given: failedQualificationScenario.given.map((fact) =>
-        fact.kind === "native-qualification"
-          ? {
-              ...fact,
-              qualifiedServices: managedNativeServiceMatrix.services.map(([service]) => service),
-              failedServices: [],
-            }
-          : fact,
-      ),
-    } satisfies ManagedStackContractScenario;
-    expect(
-      validateManagedStackContractFixtures([qualificationPartitionsContradictResult]),
-    ).toContain(
-      `${failedQualificationScenario.id}: native preflight decision must match its qualification partitions`,
-    );
-
-    const statusScenario: ManagedStackContractScenario | undefined =
-      managedStackContractFixtures.find(
-        ({ id }) => id === "identity.symlink-alias-reuses-checkout",
-      );
-    if (statusScenario === undefined || statusScenario.expected.output.json === undefined) {
-      throw new Error("identity.symlink-alias-reuses-checkout JSON fixture is required");
-    }
-    const mutatingStatus = {
-      ...statusScenario,
-      expected: {
-        ...statusScenario.expected,
-        outcome: "reuse",
-        output: {
-          ...statusScenario.expected.output,
-          json: { ...statusScenario.expected.output.json, outcome: "reuse" },
-        },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([mutatingStatus])).toContain(
-      `${statusScenario.id}: successful status operations must report`,
-    );
-
-    const stateWritingStatus = {
-      ...statusScenario,
-      expected: {
-        ...statusScenario.expected,
-        writes: [{ target: "registry", operation: "update", id: "checkout-a" }],
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([stateWritingStatus])).toContain(
-      `${statusScenario.id}: status operations must not mutate state`,
-    );
-
-    const apiStatusScenario: ManagedStackContractScenario | undefined =
-      managedStackContractFixtures.find(
-        ({ id }) => id === "identity.same-checkout-branch-and-name-reuses-stack",
-      );
-    if (apiStatusScenario === undefined || apiStatusScenario.expected.output.api === undefined) {
-      throw new Error(
-        "identity.same-checkout-branch-and-name-reuses-stack API fixture is required",
-      );
-    }
-    const apiStatusReturningReuse = {
-      ...apiStatusScenario,
-      expected: {
-        ...apiStatusScenario.expected,
-        outcome: "reuse",
-        output: {
-          ...apiStatusScenario.expected.output,
-          api: { ...apiStatusScenario.expected.output.api, outcome: "reuse" },
-        },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([apiStatusReturningReuse])).toContain(
-      `${apiStatusScenario.id}: successful status operations must report`,
-    );
-
-    const bareWorktreeScenario = managedStackContractFixtures.find(
-      ({ id }) => id === "identity.bare-repository-linked-worktrees-share-project",
-    );
-    if (bareWorktreeScenario === undefined) {
-      throw new Error(
-        "identity.bare-repository-linked-worktrees-share-project fixture is required",
-      );
-    }
-    const siblingGitStateOnly = {
-      ...bareWorktreeScenario,
-      given: bareWorktreeScenario.given.map((fact) =>
-        fact.kind === "git-state" ? { ...fact, workspacePath: "worktree-a" } : fact,
-      ),
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([siblingGitStateOnly])).toContain(
-      `${bareWorktreeScenario.id}: resolving worktree worktree-b requires its Git state`,
-    );
-
-    const recreatedCheckoutIdentity = {
-      ...bareWorktreeScenario,
-      expected: {
-        ...bareWorktreeScenario.expected,
-        writes: [
-          ...bareWorktreeScenario.expected.writes,
+    const cases: ReadonlyArray<{
+      readonly fixtures: ReadonlyArray<ManagedStackContractScenario>;
+      readonly expectedError: string;
+    }> = [
+      {
+        fixtures: [
           {
-            target: "git-config",
-            operation: "create",
-            id: "checkout-b",
-            scope: "worktree",
+            ...reuse,
+            expected: { ...reuse.expected, writes: [] },
           },
         ],
+        expectedError: `${reuse.id}: start runtime effect requires a matching state write`,
       },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([recreatedCheckoutIdentity])).toContain(
-      `${bareWorktreeScenario.id}: Git identity checkout-b is already declared`,
-    );
-
-    const unavailableRuntimeScenario = managedStackContractFixtures.find(
-      ({ id }) => id === "runtime.missing-persisted-prerequisite-fails",
-    );
-    if (unavailableRuntimeScenario === undefined) {
-      throw new Error("runtime.missing-persisted-prerequisite-fails fixture is required");
-    }
-    const ambiguousRuntimeFailure = {
-      ...unavailableRuntimeScenario,
-      given: unavailableRuntimeScenario.given.filter(({ kind }) => kind !== "stack"),
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([ambiguousRuntimeFailure])).toContain(
-      `${unavailableRuntimeScenario.id}: persisted runtime failure for stack-main-default requires an explicit stopped lifecycle`,
-    );
-
-    const portabilityScenario: ManagedStackContractScenario | undefined =
-      managedStackContractFixtures.find(
-        ({ id }) => id === "api-boundary.managed-surface-is-node-and-bun-portable",
-      );
-    if (portabilityScenario === undefined || portabilityScenario.when.interface !== "managed-api") {
-      throw new Error("api-boundary.managed-surface-is-node-and-bun-portable fixture is required");
-    }
-    const emptyRuntimeMatrix = {
-      ...portabilityScenario,
-      when: {
-        ...portabilityScenario.when,
-        input: { ...portabilityScenario.when.input, runtimes: [] },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([apiStatusScenario, emptyRuntimeMatrix])).toContain(
-      `${portabilityScenario.id}: portable contract must declare its runtimes`,
-    );
-
-    const duplicateRuntimeMatrix = {
-      ...portabilityScenario,
-      when: {
-        ...portabilityScenario.when,
-        input: { ...portabilityScenario.when.input, runtimes: ["node", "node"] },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(
-      validateManagedStackContractFixtures([apiStatusScenario, duplicateRuntimeMatrix]),
-    ).toContain(`${portabilityScenario.id}: portable contract runtimes must be unique`);
-
-    const runtimeMatrixMissingFact = {
-      ...portabilityScenario,
-      when: {
-        ...portabilityScenario.when,
-        input: { ...portabilityScenario.when.input, runtimes: ["node"] },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(
-      validateManagedStackContractFixtures([apiStatusScenario, runtimeMatrixMissingFact]),
-    ).toContain(`${portabilityScenario.id}: portable runtimes must match declared runtime facts`);
-
-    const divergentPortableResult = {
-      ...portabilityScenario,
-      expected: {
-        ...portabilityScenario.expected,
-        output: {
-          ...portabilityScenario.expected.output,
-          api: {
-            node: { outcome: "reuse", stackId: "stack-main-default" },
-            bun: { outcome: "report", stackId: "stack-main-default" },
-            equal: true,
-          },
-        },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(
-      validateManagedStackContractFixtures([apiStatusScenario, divergentPortableResult]),
-    ).toContain(
-      `${portabilityScenario.id}: portable node outcome must match ${apiStatusScenario.id}`,
-    );
-
-    const divergentPortableIdentity = {
-      ...portabilityScenario,
-      expected: {
-        ...portabilityScenario.expected,
-        output: {
-          ...portabilityScenario.expected.output,
-          api: {
-            node: { outcome: "report", stackId: "stack-main-default" },
-            bun: { outcome: "report", stackId: "stack-other" },
-            equal: true,
-          },
-        },
-      },
-    } satisfies ManagedStackContractScenario;
-    const divergentPortableIdentityErrors = validateManagedStackContractFixtures([
-      apiStatusScenario,
-      divergentPortableIdentity,
-    ]);
-    expect(divergentPortableIdentityErrors).toContain(
-      `${portabilityScenario.id}: portable bun stackId must match ${apiStatusScenario.id}`,
-    );
-    expect(divergentPortableIdentityErrors).toContain(
-      `${portabilityScenario.id}: portable runtime decisions must be identical`,
-    );
-
-    const siblingPortScenario: ManagedStackContractScenario | undefined =
-      managedStackContractFixtures.find(
-        ({ id }) => id === "ports.explicit-port-conflict-with-sibling-fails",
-      );
-    if (siblingPortScenario?.expected.selection === undefined) {
-      throw new Error("ports.explicit-port-conflict-with-sibling-fails selection is required");
-    }
-    const targetOwnsConflictingPort = {
-      ...siblingPortScenario,
-      expected: {
-        ...siblingPortScenario.expected,
-        selection: {
-          ...siblingPortScenario.expected.selection,
-          stackId: "stack-main-default",
-        },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([targetOwnsConflictingPort])).toContain(
-      `${siblingPortScenario.id}: managed sibling port owner must differ from the selected target`,
-    );
-
-    const siblingJsonOutput = siblingPortScenario.expected.output.json;
-    if (siblingJsonOutput === undefined) {
-      throw new Error("ports.explicit-port-conflict-with-sibling-fails JSON fixture is required");
-    }
-    const projectedWrongPortOwner = {
-      ...siblingPortScenario,
-      expected: {
-        ...siblingPortScenario.expected,
-        output: {
-          ...siblingPortScenario.expected.output,
-          json: { ...siblingJsonOutput, owner_stack_id: "stack-other" },
-        },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([projectedWrongPortOwner])).toContain(
-      `${siblingPortScenario.id}: projected managed port owner must match stack-main-default`,
-    );
-
-    const siblingAllocationScenario: ManagedStackContractScenario | undefined =
-      managedStackContractFixtures.find(
-        ({ id }) => id === "ports.sibling-targets-allocate-independent-ports",
-      );
-    if (
-      siblingAllocationScenario === undefined ||
-      siblingAllocationScenario.expected.output.api === undefined
-    ) {
-      throw new Error("ports.sibling-targets-allocate-independent-ports API fixture is required");
-    }
-    const siblingAllocationCollision = {
-      ...siblingAllocationScenario,
-      expected: {
-        ...siblingAllocationScenario.expected,
-        output: {
-          ...siblingAllocationScenario.expected.output,
-          api: {
-            ...siblingAllocationScenario.expected.output.api,
-            ports: { api: 55421, db: 55424 },
-          },
-        },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([siblingAllocationCollision])).toContain(
-      `${siblingAllocationScenario.id}: allocated port 55421 conflicts with a sibling target`,
-    );
-
-    const duplicateSiblingAllocation = {
-      ...siblingAllocationScenario,
-      expected: {
-        ...siblingAllocationScenario.expected,
-        output: {
-          ...siblingAllocationScenario.expected.output,
-          api: {
-            ...siblingAllocationScenario.expected.output.api,
-            ports: { api: 55424, db: 55424 },
-          },
-        },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([duplicateSiblingAllocation])).toContain(
-      `${siblingAllocationScenario.id}: allocated port 55424 is assigned more than once`,
-    );
-
-    const exactPortScenario: ManagedStackContractScenario | undefined =
-      managedStackContractFixtures.find(({ id }) => id === "ports.explicit-free-port-is-used");
-    if (exactPortScenario === undefined || exactPortScenario.when.interface !== "managed-api") {
-      throw new Error("ports.explicit-free-port-is-used fixture is required");
-    }
-    const actionRequestingDifferentExactPort = {
-      ...exactPortScenario,
-      when: {
-        ...exactPortScenario.when,
-        input: {
-          ...exactPortScenario.when.input,
-          portIntents: { "api.port": { intent: "exact", port: 54322 } },
-        },
-      },
-    } satisfies ManagedStackContractScenario;
-    const exactPortErrors = validateManagedStackContractFixtures([
-      actionRequestingDifferentExactPort,
-    ]);
-    expect(exactPortErrors).toContain(
-      `${exactPortScenario.id}: exact port request api.port must match its config fact`,
-    );
-
-    const stickyPortScenario: ManagedStackContractScenario | undefined =
-      managedStackContractFixtures.find(
-        ({ id }) => id === "ports.later-sticky-port-collision-fails",
-      );
-    if (stickyPortScenario === undefined) {
-      throw new Error("ports.later-sticky-port-collision-fails fixture is required");
-    }
-    const stickyPortWithoutStoppedTarget = {
-      ...stickyPortScenario,
-      given: stickyPortScenario.given.filter(({ kind }) => kind !== "stack"),
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([stickyPortWithoutStoppedTarget])).toContain(
-      `${stickyPortScenario.id}: sticky port conflict requires a stopped selected stack`,
-    );
-
-    const portUpdateScenario = managedStackContractFixtures.find(
-      ({ id }) => id === "ports.config-change-on-stopped-stack-applies",
-    );
-    if (portUpdateScenario === undefined) {
-      throw new Error("ports.config-change-on-stopped-stack-applies fixture is required");
-    }
-    const unboundPortUpdate = {
-      ...portUpdateScenario,
-      expected: { ...portUpdateScenario.expected, selection: undefined },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([unboundPortUpdate])).toContain(
-      `${portUpdateScenario.id}: contextual CLI stack result requires a selected target`,
-    );
-
-    const runtimeConflictScenario: ManagedStackContractScenario | undefined =
-      managedStackContractFixtures.find(
-        ({ id }) => id === "runtime.persisted-runtime-conflict-fails",
-      );
-    if (runtimeConflictScenario === undefined) {
-      throw new Error("runtime.persisted-runtime-conflict-fails fixture is required");
-    }
-    const unrelatedPersistedRuntime = {
-      ...runtimeConflictScenario,
-      given: runtimeConflictScenario.given.map((fact) =>
-        fact.kind === "persisted-runtime" ? { ...fact, stackId: "stack-other" } : fact,
-      ),
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([unrelatedPersistedRuntime])).toContain(
-      `${runtimeConflictScenario.id}: persisted runtime must belong to the selected target`,
-    );
-
-    const freshCloneScenario = managedStackContractFixtures.find(
-      ({ id }) => id === "identity.fresh-clone-creates-project-and-checkout",
-    );
-    if (freshCloneScenario === undefined) {
-      throw new Error("identity.fresh-clone-creates-project-and-checkout fixture is required");
-    }
-    const cloneWithoutGitState = {
-      ...freshCloneScenario,
-      given: freshCloneScenario.given.filter(({ kind }) => kind !== "git-state"),
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([cloneWithoutGitState])).toContain(
-      `${freshCloneScenario.id}: creating a Git context requires Git state for the workspace`,
-    );
-
-    const refReplacementScenario = managedStackContractFixtures.find(
-      ({ id }) => id === "identity.manual-ref-replacement-orphans-context",
-    );
-    if (refReplacementScenario === undefined) {
-      throw new Error("identity.manual-ref-replacement-orphans-context fixture is required");
-    }
-    const refReplacementWithoutGitState = {
-      ...refReplacementScenario,
-      given: refReplacementScenario.given.filter((fact) => fact.kind !== "git-state"),
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([refReplacementWithoutGitState])).toContain(
-      `${refReplacementScenario.id}: creating a Git context requires Git state for the workspace`,
-    );
-
-    const detachedCommitScenario = managedStackContractFixtures.find(
-      ({ id }) => id === "identity.detached-commits-reuse-checkout-context",
-    );
-    if (detachedCommitScenario === undefined) {
-      throw new Error("identity.detached-commits-reuse-checkout-context fixture is required");
-    }
-    const detachedReuseWithoutTransition = {
-      ...detachedCommitScenario,
-      given: detachedCommitScenario.given.filter(
-        (fact) => fact.kind !== "identity-transition" || fact.operation !== "detached-commit",
-      ),
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([detachedReuseWithoutTransition])).toContain(
-      `${detachedCommitScenario.id}: detached reuse must declare the commit transition`,
-    );
-
-    const retryScenario = managedStackContractFixtures.find(
-      ({ id }) => id === "bootstrap.retry-after-failed-copy-succeeds",
-    );
-    if (retryScenario === undefined) {
-      throw new Error("bootstrap.retry-after-failed-copy-succeeds fixture is required");
-    }
-    const retryWithoutRollback = {
-      ...retryScenario,
-      given: retryScenario.given.filter(({ kind }) => kind !== "operation-result"),
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([retryWithoutRollback])).toContain(
-      `${retryScenario.id}: bootstrap retry requires a rolled-back prior attempt`,
-    );
-
-    for (const scenarioId of [
-      "credentials.configured-values-are-authoritative",
-      "credentials.explicit-change-applies-after-stop",
-      "credentials.omitted-values-use-stable-defaults",
-      "credentials.compatible-legacy-auth-is-retained",
-    ]) {
-      const credentialPersistenceScenario: ManagedStackContractScenario | undefined =
-        managedStackContractFixtures.find(({ id }) => id === scenarioId);
-      if (credentialPersistenceScenario === undefined) {
-        throw new Error(`${scenarioId} fixture is required`);
-      }
-      const globallyPersistedPlaintext = {
-        ...credentialPersistenceScenario,
-        expected: {
-          ...credentialPersistenceScenario.expected,
-          details: {
-            ...credentialPersistenceScenario.expected.details,
-            plaintext_secrets_in_global_state: true,
-          },
-        },
-      } satisfies ManagedStackContractScenario;
-      expect(validateManagedStackContractFixtures([globallyPersistedPlaintext])).toContain(
-        `${credentialPersistenceScenario.id}: credential persistence must not expose plaintext globally`,
-      );
-    }
-
-    const pruneScenario = managedStackContractFixtures.find(
-      ({ id }) => id === "reclamation.prune-removes-metadata-only",
-    );
-    if (pruneScenario === undefined) {
-      throw new Error("reclamation.prune-removes-metadata-only fixture is required");
-    }
-    const pruneWithoutMutableData = {
-      ...pruneScenario,
-      given: pruneScenario.given.filter(({ kind }) => kind !== "stack"),
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([pruneWithoutMutableData])).toContain(
-      `${pruneScenario.id}: data-preserving prune must declare mutable stack data`,
-    );
-
-    const branchDeletionScenario = managedStackContractFixtures.find(
-      ({ id }) => id === "reclamation.branch-delete-does-not-delete-data",
-    );
-    if (branchDeletionScenario === undefined) {
-      throw new Error("reclamation.branch-delete-does-not-delete-data fixture is required");
-    }
-    const deletionWithUnboundBranch = {
-      ...branchDeletionScenario,
-      given: branchDeletionScenario.given.map((fact) =>
-        fact.kind === "branch" && fact.name === "feat-a"
-          ? { ...fact, contextId: "context-other" }
-          : fact,
-      ),
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([deletionWithUnboundBranch])).toContain(
-      `${branchDeletionScenario.id}: branch deletion must bind its branch to an affected managed stack`,
-    );
-
-    const deletionWithoutCheckoutGitState = {
-      ...branchDeletionScenario,
-      given: branchDeletionScenario.given.filter(
-        (fact) => fact.kind !== "checkout" && fact.kind !== "git-state",
-      ),
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([deletionWithoutCheckoutGitState])).toContain(
-      `${branchDeletionScenario.id}: branch deletion must declare checkout Git state`,
-    );
-
-    const deletionWithoutPreservationResult = {
-      ...branchDeletionScenario,
-      expected: { ...branchDeletionScenario.expected, details: undefined },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([deletionWithoutPreservationResult])).toContain(
-      `${branchDeletionScenario.id}: branch deletion must preserve and orphan managed stack data`,
-    );
-
-    const deleteScenario: ManagedStackContractScenario | undefined =
-      managedStackContractFixtures.find(({ id }) => id === "reclamation.delete-orphan-by-stack-id");
-    if (deleteScenario === undefined || deleteScenario.when.interface !== "cli") {
-      throw new Error("reclamation.delete-orphan-by-stack-id fixture is required");
-    }
-    const deleteWithMismatchedActionTarget = {
-      ...deleteScenario,
-      when: {
-        ...deleteScenario.when,
-        argv: deleteScenario.when.argv.map((arg) => (arg === "stack-orphan" ? "stack-other" : arg)),
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([deleteWithMismatchedActionTarget])).toContain(
-      `${deleteScenario.id}: explicit action target stack-other disagrees with expected stack stack-orphan`,
-    );
-
-    const failedCopyScenario: ManagedStackContractScenario | undefined =
-      managedStackContractFixtures.find(({ id }) => id === "bootstrap.failed-copy-rolls-back");
-    if (failedCopyScenario === undefined || failedCopyScenario.when.interface !== "managed-api") {
-      throw new Error("bootstrap.failed-copy-rolls-back managed API fixture is required");
-    }
-    const failedCopyWithMismatchedActionTarget = {
-      ...failedCopyScenario,
-      when: {
-        ...failedCopyScenario.when,
-        input: { ...failedCopyScenario.when.input, stackId: "stack-other" },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([failedCopyWithMismatchedActionTarget])).toContain(
-      `${failedCopyScenario.id}: explicit action target stack-other disagrees with expected stack stack-main-default`,
-    );
-
-    const runtimeMetadataOnlyDelete = {
-      ...deleteScenario,
-      expected: {
-        ...deleteScenario.expected,
-        writes: deleteScenario.expected.writes.filter((write) => write.target !== "managed-state"),
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([runtimeMetadataOnlyDelete])).toContain(
-      `${deleteScenario.id}: delete runtime effect requires a matching state write`,
-    );
-
-    const deleteWithoutStop = {
-      ...deleteScenario,
-      expected: {
-        ...deleteScenario.expected,
-        runtimeEffects: deleteScenario.expected.runtimeEffects.filter(
-          ({ operation }) => operation !== "stop",
-        ),
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([deleteWithoutStop])).toContain(
-      `${deleteScenario.id}: runtime-state delete requires a matching runtime effect`,
-    );
-
-    const deleteWithoutTombstone = {
-      ...deleteScenario,
-      expected: {
-        ...deleteScenario.expected,
-        writes: deleteScenario.expected.writes.filter(
-          (write) => write.target !== "registry" || write.operation !== "tombstone",
-        ),
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([deleteWithoutTombstone])).toContain(
-      `${deleteScenario.id}: managed-state deletion requires a registry tombstone`,
-    );
-
-    const stickyReuseScenario = managedStackContractFixtures.find(
-      ({ id }) => id === "ports.sticky-ports-reuse-on-return",
-    );
-    if (stickyReuseScenario === undefined) {
-      throw new Error("ports.sticky-ports-reuse-on-return fixture is required");
-    }
-    const unboundStickyReuse = {
-      ...stickyReuseScenario,
-      expected: { ...stickyReuseScenario.expected, selection: undefined },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([unboundStickyReuse])).toContain(
-      `${stickyReuseScenario.id}: sticky port reuse requires a selected target`,
-    );
-
-    const repositoryContractScenario: ManagedStackContractScenario | undefined =
-      managedStackContractFixtures.find(
-        ({ id }) => id === "api-boundary.repository-contract-is-storage-agnostic",
-      );
-    if (
-      repositoryContractScenario === undefined ||
-      repositoryContractScenario.when.interface !== "managed-api"
-    ) {
-      throw new Error("api-boundary.repository-contract-is-storage-agnostic fixture is required");
-    }
-    const repositoryApiOutput = repositoryContractScenario.expected.output.api;
-    if (repositoryApiOutput === undefined) {
-      throw new Error("repository contract API output is required");
-    }
-    const emptyRepositoryMatrix = {
-      ...repositoryContractScenario,
-      when: {
-        ...repositoryContractScenario.when,
-        input: { ...repositoryContractScenario.when.input, adapters: [] },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([scenario, emptyRepositoryMatrix])).toContain(
-      `${repositoryContractScenario.id}: repository contract must declare its adapters`,
-    );
-
-    const duplicateRepositoryMatrix = {
-      ...repositoryContractScenario,
-      when: {
-        ...repositoryContractScenario.when,
-        input: {
-          ...repositoryContractScenario.when.input,
-          adapters: ["in-memory", "in-memory"],
-        },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([scenario, duplicateRepositoryMatrix])).toContain(
-      `${repositoryContractScenario.id}: repository contract adapters must be unique`,
-    );
-
-    const repositoryMatrixMissingFact = {
-      ...repositoryContractScenario,
-      when: {
-        ...repositoryContractScenario.when,
-        input: { ...repositoryContractScenario.when.input, adapters: ["in-memory"] },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([scenario, repositoryMatrixMissingFact])).toContain(
-      `${repositoryContractScenario.id}: repository adapters must match declared repository facts`,
-    );
-
-    const unknownRepositoryReference = {
-      ...repositoryContractScenario,
-      when: {
-        ...repositoryContractScenario.when,
-        input: { ...repositoryContractScenario.when.input, scenarioId: "identity.unknown" },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([unknownRepositoryReference])).toContain(
-      `${repositoryContractScenario.id}: repository contract must reference a declared scenario`,
-    );
-
-    const staleRepositoryOutcome = {
-      ...repositoryContractScenario,
-      expected: {
-        ...repositoryContractScenario.expected,
-        output: {
-          ...repositoryContractScenario.expected.output,
-          api: {
-            ...repositoryApiOutput,
-            "in-memory": { outcome: "create", stackId: "stack-main-default" },
-          },
-        },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([scenario, staleRepositoryOutcome])).toContain(
-      `${repositoryContractScenario.id}: repository in-memory outcome must match ${scenario.id}`,
-    );
-
-    const divergentRepositoryIdentity = {
-      ...repositoryContractScenario,
-      expected: {
-        ...repositoryContractScenario.expected,
-        output: {
-          ...repositoryContractScenario.expected.output,
-          api: {
-            ...repositoryApiOutput,
-            "persistent-adapter": { outcome: "reuse", stackId: "stack-other" },
-          },
-        },
-      },
-    } satisfies ManagedStackContractScenario;
-    const divergentRepositoryErrors = validateManagedStackContractFixtures([
-      scenario,
-      divergentRepositoryIdentity,
-    ]);
-    expect(divergentRepositoryErrors).toContain(
-      `${repositoryContractScenario.id}: repository persistent-adapter stackId must match ${scenario.id}`,
-    );
-    expect(divergentRepositoryErrors).toContain(
-      `${repositoryContractScenario.id}: repository adapter decisions must be identical`,
-    );
-
-    const invalidNameScenario: ManagedStackContractScenario | undefined =
-      managedStackContractFixtures.find(
-        ({ id }) => id === "identity.invalid-stack-name-uppercase-underscore-fails",
-      );
-    const invalidNameHumanOutput = invalidNameScenario?.expected.output.human;
-    if (invalidNameScenario === undefined || invalidNameHumanOutput === undefined) {
-      throw new Error("invalid stack name fixture with human recovery is required");
-    }
-    const divergentHumanRecovery = {
-      ...invalidNameScenario,
-      expected: {
-        ...invalidNameScenario.expected,
-        output: {
-          ...invalidNameScenario.expected.output,
-          human: { ...invalidNameHumanOutput, recovery: ["Try again"] },
-        },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([divergentHumanRecovery])).toContain(
-      `${invalidNameScenario.id}: human recovery disagrees with the managed result`,
-    );
-
-    const divergentJsonRecovery = {
-      ...invalidNameScenario,
-      expected: {
-        ...invalidNameScenario.expected,
-        output: {
-          ...invalidNameScenario.expected.output,
-          json: { ...invalidNameScenario.expected.output.json, recovery: ["Try again"] },
-        },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([divergentJsonRecovery])).toContain(
-      `${invalidNameScenario.id}: JSON recovery disagrees with the managed result`,
-    );
-
-    const invalidNameJson = invalidNameScenario.expected.output.json;
-    if (invalidNameJson === undefined) {
-      throw new Error("invalid stack name JSON fixture is required");
-    }
-    const { code: omittedCode, ...jsonWithoutCode } = invalidNameJson;
-    expect(omittedCode).toBe("INVALID_STACK_NAME");
-    const jsonProjectionWithoutCode = {
-      ...invalidNameScenario,
-      expected: {
-        ...invalidNameScenario.expected,
-        output: { ...invalidNameScenario.expected.output, json: jsonWithoutCode },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([jsonProjectionWithoutCode])).toContain(
-      `${invalidNameScenario.id}: JSON projection requires a code`,
-    );
-
-    const credentialDefaultsScenario: ManagedStackContractScenario | undefined =
-      managedStackContractFixtures.find(
-        ({ id }) => id === "credentials.omitted-values-use-stable-defaults",
-      );
-    const credentialDefaultsJson = credentialDefaultsScenario?.expected.output.json;
-    if (credentialDefaultsScenario === undefined || credentialDefaultsJson === undefined) {
-      throw new Error("credentials.omitted-values-use-stable-defaults JSON fixture is required");
-    }
-    const { outcome: omittedOutcome, ...jsonWithoutOutcome } = credentialDefaultsJson;
-    expect(omittedOutcome).toBe("create");
-    const credentialProjectionWithoutOutcome = {
-      ...credentialDefaultsScenario,
-      expected: {
-        ...credentialDefaultsScenario.expected,
-        output: { ...credentialDefaultsScenario.expected.output, json: jsonWithoutOutcome },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([credentialProjectionWithoutOutcome])).toContain(
-      `${credentialDefaultsScenario.id}: JSON projection requires an outcome`,
-    );
-
-    const newBranchScenario = managedStackContractFixtures.find(
-      ({ id }) => id === "identity.new-branch-first-start-creates-stack",
-    );
-    if (newBranchScenario === undefined) {
-      throw new Error("identity.new-branch-first-start-creates-stack fixture is required");
-    }
-    const contextOwnedByDifferentBranch = {
-      ...newBranchScenario,
-      expected: {
-        ...newBranchScenario.expected,
-        writes: newBranchScenario.expected.writes.map((write) =>
-          write.target === "git-config" && write.id === "context-feat-a"
-            ? { ...write, owner: "main" }
-            : write,
-        ),
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([contextOwnedByDifferentBranch])).toContain(
-      `${newBranchScenario.id}: Git context context-feat-a must belong to branch feat-a`,
-    );
-
-    const recreatedBranchScenario = managedStackContractFixtures.find(
-      ({ id }) => id === "identity.branch-delete-recreate-creates-context",
-    );
-    if (recreatedBranchScenario === undefined) {
-      throw new Error("identity.branch-delete-recreate-creates-context fixture is required");
-    }
-    const recreatedBranchWithoutGitState = {
-      ...recreatedBranchScenario,
-      given: recreatedBranchScenario.given.filter((fact) => fact.kind !== "git-state"),
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([recreatedBranchWithoutGitState])).toContain(
-      `${recreatedBranchScenario.id}: creating a Git context requires Git state for the workspace`,
-    );
-
-    const firstOrdinaryFolderStart = managedStackContractFixtures.find(
-      ({ id }) => id === "identity.non-git-folder-first-start-persists-identity",
-    );
-    if (firstOrdinaryFolderStart === undefined) {
-      throw new Error("identity.non-git-folder-first-start-persists-identity fixture is required");
-    }
-    const ordinaryFolderWithoutMarkerWrite = {
-      ...firstOrdinaryFolderStart,
-      expected: {
-        ...firstOrdinaryFolderStart.expected,
-        writes: firstOrdinaryFolderStart.expected.writes.filter(
-          (write) => write.target !== "identity-marker",
-        ),
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([ordinaryFolderWithoutMarkerWrite])).toContain(
-      `${firstOrdinaryFolderStart.id}: ordinary-folder creation must persist its identity marker`,
-    );
-
-    const laterOrdinaryFolderStart = managedStackContractFixtures.find(
-      ({ id }) => id === "identity.non-git-folder-recovers-persisted-identity",
-    );
-    if (laterOrdinaryFolderStart === undefined) {
-      throw new Error("identity.non-git-folder-recovers-persisted-identity fixture is required");
-    }
-    const ordinaryFolderWithoutMarkerFact = {
-      ...laterOrdinaryFolderStart,
-      given: laterOrdinaryFolderStart.given.filter((fact) => fact.kind !== "identity-marker"),
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([ordinaryFolderWithoutMarkerFact])).toContain(
-      `${laterOrdinaryFolderStart.id}: ordinary-folder reuse must resolve its identity marker`,
-    );
-
-    const runtimeCreationScenario = managedStackContractFixtures.find(
-      ({ id }) => id === "runtime.explicit-api-overrides-auto",
-    );
-    if (runtimeCreationScenario === undefined) {
-      throw new Error("runtime.explicit-api-overrides-auto fixture is required");
-    }
-    const runtimeCreationWithoutAbsentTarget = {
-      ...runtimeCreationScenario,
-      given: runtimeCreationScenario.given.filter((fact) => fact.kind !== "managed-target"),
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([runtimeCreationWithoutAbsentTarget])).toContain(
-      `${runtimeCreationScenario.id}: managed creation must declare absent target stack-main-default`,
-    );
-
-    const runtimeCreationWithoutLegacyState = {
-      ...runtimeCreationScenario,
-      given: runtimeCreationScenario.given.filter((fact) => fact.kind !== "legacy-state"),
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([runtimeCreationWithoutLegacyState])).toContain(
-      `${runtimeCreationScenario.id}: managed creation must declare legacy state absent or incompatible`,
-    );
-    if (runtimeCreationScenario.when.interface !== "managed-api") {
-      throw new Error("runtime.explicit-api-overrides-auto managed API fixture is required");
-    }
-    const runtimeActionDisagreesWithRequest = {
-      ...runtimeCreationScenario,
-      when: {
-        ...runtimeCreationScenario.when,
-        input: { ...runtimeCreationScenario.when.input, runtime: "docker" },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([runtimeActionDisagreesWithRequest])).toContain(
-      `${runtimeCreationScenario.id}: explicit runtime docker must match its managed-api request fact`,
-    );
-
-    const credentialCreationScenario = managedStackContractFixtures.find(
-      ({ id }) => id === "credentials.configured-values-are-authoritative",
-    );
-    if (credentialCreationScenario === undefined) {
-      throw new Error("credentials.configured-values-are-authoritative fixture is required");
-    }
-    const credentialCreationWithoutAbsentTarget = {
-      ...credentialCreationScenario,
-      given: credentialCreationScenario.given.filter((fact) => fact.kind !== "managed-target"),
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([credentialCreationWithoutAbsentTarget])).toContain(
-      `${credentialCreationScenario.id}: managed creation must declare absent target stack-main-default`,
-    );
-
-    const credentialCreationWithoutLegacyState = {
-      ...credentialCreationScenario,
-      given: credentialCreationScenario.given.filter((fact) => fact.kind !== "legacy-state"),
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([credentialCreationWithoutLegacyState])).toContain(
-      `${credentialCreationScenario.id}: managed creation must declare legacy state absent or incompatible`,
-    );
-    if (credentialCreationScenario.when.interface !== "managed-api") {
-      throw new Error(
-        "credentials.configured-values-are-authoritative managed API fixture is required",
-      );
-    }
-    const credentialActionUsingDifferentReference = {
-      ...credentialCreationScenario,
-      when: {
-        ...credentialCreationScenario.when,
-        input: { ...credentialCreationScenario.when.input, auth: "configured-auth-v2" },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(
-      validateManagedStackContractFixtures([credentialActionUsingDifferentReference]),
-    ).toContain(
-      `${credentialCreationScenario.id}: configured credential input configured-auth-v2 must match persisted references`,
-    );
-
-    const portCreationScenario = managedStackContractFixtures.find(
-      ({ id }) => id === "ports.new-target-allocates-and-persists-omitted-ports",
-    );
-    if (portCreationScenario === undefined) {
-      throw new Error("ports.new-target-allocates-and-persists-omitted-ports fixture is required");
-    }
-    const portCreationWithoutLegacyState = {
-      ...portCreationScenario,
-      given: portCreationScenario.given.filter((fact) => fact.kind !== "legacy-state"),
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([portCreationWithoutLegacyState])).toContain(
-      `${portCreationScenario.id}: managed creation must declare legacy state absent or incompatible`,
-    );
-    const portCreationWithCopyableLegacyState = {
-      ...portCreationScenario,
-      given: portCreationScenario.given.map((fact) =>
-        fact.kind === "legacy-state"
-          ? {
-              ...fact,
-              lifecycle: "stopped",
-              database: "compatible",
-              storage: "compatible",
-              credentials: "compatible",
-            }
-          : fact,
-      ),
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([portCreationWithCopyableLegacyState])).toContain(
-      `${portCreationScenario.id}: managed creation must declare legacy state absent or incompatible`,
-    );
-
-    const concurrencyScenario: ManagedStackContractScenario | undefined =
-      managedStackContractFixtures.find(
-        ({ id }) => id === "identity.concurrent-create-publishes-once",
-      );
-    if (concurrencyScenario === undefined || concurrencyScenario.when.interface !== "managed-api") {
-      throw new Error("identity.concurrent-create-publishes-once fixture is required");
-    }
-    const singleContenderAction = {
-      ...concurrencyScenario,
-      when: {
-        ...concurrencyScenario.when,
-        input: { ...concurrencyScenario.when.input, contenders: 1 },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([singleContenderAction])).toContain(
-      `${concurrencyScenario.id}: concurrent action contenders must match the declared race of 2`,
-    );
-
-    const concurrencyForDifferentTarget = {
-      ...concurrencyScenario,
-      when: {
-        ...concurrencyScenario.when,
-        input: { ...concurrencyScenario.when.input, stackName: "review" },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([concurrencyForDifferentTarget])).toContain(
-      `${concurrencyScenario.id}: concurrent action target must match context-feat/default`,
-    );
-
-    const incompleteConcurrentResults = {
-      ...concurrencyScenario,
-      expected: {
-        ...concurrencyScenario.expected,
-        details: { ...concurrencyScenario.expected.details, contender_results: ["create"] },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([incompleteConcurrentResults])).toContain(
-      `${concurrencyScenario.id}: concurrent details results must cover 2 contenders`,
-    );
-
-    const duplicateConcurrentCreation = {
-      ...concurrencyScenario,
-      expected: {
-        ...concurrencyScenario.expected,
-        details: {
-          ...concurrencyScenario.expected.details,
-          contender_results: ["create", "create"],
-        },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([duplicateConcurrentCreation])).toContain(
-      `${concurrencyScenario.id}: concurrent race must create once and reuse thereafter`,
-    );
-
-    const isolatedStateRootScenario: ManagedStackContractScenario | undefined =
-      managedStackContractFixtures.find(
-        ({ id }) => id === "api-boundary.managed-api-accepts-isolated-state-root",
-      );
-    if (
-      isolatedStateRootScenario === undefined ||
-      isolatedStateRootScenario.when.interface !== "managed-api"
-    ) {
-      throw new Error("api-boundary.managed-api-accepts-isolated-state-root fixture is required");
-    }
-    const actionUsingDifferentStateRoot = {
-      ...isolatedStateRootScenario,
-      when: {
-        ...isolatedStateRootScenario.when,
-        input: { ...isolatedStateRootScenario.when.input, stateRoot: "/tmp/other-root" },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([actionUsingDifferentStateRoot])).toContain(
-      `${isolatedStateRootScenario.id}: isolated state root input must match its options and observed boundary`,
-    );
-
-    const isolatedResolutionWithoutAbsentTarget = {
-      ...isolatedStateRootScenario,
-      given: isolatedStateRootScenario.given.filter((fact) => fact.kind !== "managed-target"),
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([isolatedResolutionWithoutAbsentTarget])).toContain(
-      `${isolatedStateRootScenario.id}: managed creation must declare absent target stack-main-default`,
-    );
-
-    const isolatedResolutionWithoutIdentityClaims = {
-      ...isolatedStateRootScenario,
-      given: isolatedStateRootScenario.given.filter((fact) => fact.kind !== "identity-claim"),
-    } satisfies ManagedStackContractScenario;
-    expect(
-      validateManagedStackContractFixtures([isolatedResolutionWithoutIdentityClaims]),
-    ).toContain(
-      `${isolatedStateRootScenario.id}: creating Git identity project-a requires an absent project claim`,
-    );
-
-    const bootstrapCopyScenario: ManagedStackContractScenario | undefined =
-      managedStackContractFixtures.find(
-        ({ id }) => id === "bootstrap.first-start-copies-compatible-legacy-state",
-      );
-    if (bootstrapCopyScenario === undefined) {
-      throw new Error("bootstrap.first-start-copies-compatible-legacy-state fixture is required");
-    }
-    const copyFromRunningLegacyState = {
-      ...bootstrapCopyScenario,
-      given: bootstrapCopyScenario.given.map((fact) =>
-        fact.kind === "legacy-state" ? { ...fact, lifecycle: "running" } : fact,
-      ),
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([copyFromRunningLegacyState])).toContain(
-      `${bootstrapCopyScenario.id}: bootstrap copy requires absent target stack-main-default and compatible stopped legacy state`,
-    );
-  });
-
-  it("enforces diagnostic codes and non-mutating report and error outcomes", () => {
-    const readOnlyScenario: ManagedStackContractScenario | undefined =
-      managedStackContractFixtures.find(
-        ({ id }) => id === "identity.branch-copy-read-only-does-not-write",
-      );
-    const ambiguousScenario: ManagedStackContractScenario | undefined =
-      managedStackContractFixtures.find(
-        ({ id }) => id === "identity.branch-copy-ambiguous-read-only",
-      );
-    if (readOnlyScenario === undefined || ambiguousScenario?.expected.error === undefined) {
-      throw new Error("branch-copy read-only fixtures are required");
-    }
-
-    const mutatingReport = {
-      ...readOnlyScenario,
-      expected: {
-        ...readOnlyScenario.expected,
-        writes: [{ target: "registry", operation: "update", id: "context-main" }],
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([mutatingReport])).toContain(
-      `${readOnlyScenario.id}: report outcome must not mutate state`,
-    );
-
-    const mutatingError = {
-      ...ambiguousScenario,
-      expected: {
-        ...ambiguousScenario.expected,
-        writes: [{ target: "managed-state", operation: "update", id: "context-main" }],
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([mutatingError])).toContain(
-      `${ambiguousScenario.id}: error outcome must not mutate state outside rollback cleanup`,
-    );
-
-    const lowerCaseCode = {
-      ...ambiguousScenario,
-      expected: {
-        ...ambiguousScenario.expected,
-        error: { ...ambiguousScenario.expected.error, code: "ambiguous_context_owner" },
-        output: {
-          ...ambiguousScenario.expected.output,
-          json: {
-            ...ambiguousScenario.expected.output.json,
-            code: "ambiguous_context_owner",
-          },
-        },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([lowerCaseCode])).toContain(
-      `${ambiguousScenario.id}: diagnostic code ambiguous_context_owner must use SCREAMING_SNAKE_CASE`,
-    );
-
-    const ambiguityWithoutClaim = {
-      ...ambiguousScenario,
-      given: ambiguousScenario.given.filter((fact) => fact.kind !== "identity-claim"),
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([ambiguityWithoutClaim])).toContain(
-      `${ambiguousScenario.id}: ambiguous context must bind at least two claiming branches to its projections`,
-    );
-
-    const ambiguityWithoutTransition = {
-      ...ambiguousScenario,
-      given: ambiguousScenario.given.filter(
-        (fact) => fact.kind !== "identity-transition" || fact.operation !== "branch-copy",
-      ),
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([ambiguityWithoutTransition])).toContain(
-      `${ambiguousScenario.id}: ambiguous context must bind at least two claiming branches to its projections`,
-    );
-  });
-
-  it("binds public action inputs and state facts to observable results", () => {
-    const findScenario = (id: string): ManagedStackContractScenario => {
-      const scenario = managedStackContractFixtures.find((candidate) => candidate.id === id);
-      if (scenario === undefined) {
-        throw new Error(`${id} fixture is required`);
-      }
-      return scenario;
-    };
-
-    const stackNamesScenario = findScenario("identity.valid-stack-names-resolve-deterministically");
-    if (stackNamesScenario.when.interface !== "managed-api") {
-      throw new Error("stack-name resolution must use the managed API");
-    }
-    const missingRequestedStackName = {
-      ...stackNamesScenario,
-      when: {
-        ...stackNamesScenario.when,
-        input: { ...stackNamesScenario.when.input, stackNames: ["default"] },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([missingRequestedStackName])).toContain(
-      `${stackNamesScenario.id}: requested stack names must match their fact and projected results`,
-    );
-
-    const automaticPortsScenario = findScenario(
-      "ports.new-target-allocates-and-persists-omitted-ports",
-    );
-    if (automaticPortsScenario.when.interface !== "managed-api") {
-      throw new Error("automatic port allocation must use the managed API");
-    }
-    const missingAutomaticPortRequest = {
-      ...automaticPortsScenario,
-      when: {
-        ...automaticPortsScenario.when,
-        input: {
-          ...automaticPortsScenario.when.input,
-          portIntents: { "api.port": "automatic" },
-        },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([missingAutomaticPortRequest])).toContain(
-      `${automaticPortsScenario.id}: requested port keys must match their config facts`,
-    );
-    const projectedAutomaticIntentDisagrees = {
-      ...automaticPortsScenario,
-      expected: {
-        ...automaticPortsScenario.expected,
-        output: {
-          ...automaticPortsScenario.expected.output,
-          api: {
-            ...automaticPortsScenario.expected.output.api,
-            intents: { api: "automatic", db: "exact" },
-          },
-        },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([projectedAutomaticIntentDisagrees])).toContain(
-      `${automaticPortsScenario.id}: automatic port request db.port must match its fact and projected allocation`,
-    );
-
-    const autoRuntimeScenario = findScenario("runtime.auto-prefers-docker");
-    const unavailablePreferredRuntime = {
-      ...autoRuntimeScenario,
-      given: autoRuntimeScenario.given.map((fact) =>
-        fact.kind === "runtime-availability" && fact.runtime === "docker"
-          ? { ...fact, available: false }
-          : fact,
-      ),
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([unavailablePreferredRuntime])).toContain(
-      `${autoRuntimeScenario.id}: automatic runtime failure must bind both unavailability reasons`,
-    );
-
-    const injectedRepositoryScenario = findScenario(
-      "api-boundary.managed-api-accepts-injected-repository",
-    );
-    if (injectedRepositoryScenario.when.interface !== "managed-api") {
-      throw new Error("injected repository boundary must use the managed API");
-    }
-    const differentInjectedRepository = {
-      ...injectedRepositoryScenario,
-      when: {
-        ...injectedRepositoryScenario.when,
-        input: { ...injectedRepositoryScenario.when.input, repository: "other-repository" },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([differentInjectedRepository])).toContain(
-      `${injectedRepositoryScenario.id}: injected repository and state root must match the observed managed service`,
-    );
-
-    const nativePreflightScenario = findScenario(
-      "native-qualification.all-services-qualify-platform",
-    );
-    if (nativePreflightScenario.when.interface !== "managed-api") {
-      throw new Error("native preflight must use the managed API");
-    }
-    const preflightWithoutPlatform = {
-      ...nativePreflightScenario,
-      when: { ...nativePreflightScenario.when, input: {} },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([preflightWithoutPlatform])).toContain(
-      `${nativePreflightScenario.id}: native qualification platform must match the preflight action`,
-    );
-
-    const persistedRuntimeConflictScenario = findScenario(
-      "runtime.persisted-runtime-conflict-fails",
-    );
-    const matchingPersistedRuntime = {
-      ...persistedRuntimeConflictScenario,
-      given: persistedRuntimeConflictScenario.given.map((fact) =>
-        fact.kind === "persisted-runtime" ? { ...fact, runtime: "native" } : fact,
-      ),
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([matchingPersistedRuntime])).toContain(
-      `${persistedRuntimeConflictScenario.id}: persisted runtime conflict must bind persisted and requested values`,
-    );
-
-    const destructiveStopScenario = findScenario("reclamation.delete-orphan-by-stack-id");
-    if (destructiveStopScenario.when.interface !== "cli") {
-      throw new Error("destructive orphan deletion must use the CLI");
-    }
-    const destructiveStopWithoutFlag = {
-      ...destructiveStopScenario,
-      when: {
-        ...destructiveStopScenario.when,
-        argv: destructiveStopScenario.when.argv.filter((arg) => arg !== "--no-backup"),
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([destructiveStopWithoutFlag])).toContain(
-      `${destructiveStopScenario.id}: destructive stop requires --no-backup`,
-    );
-
-    const failedCopyScenario = findScenario("bootstrap.failed-copy-rolls-back");
-    if (failedCopyScenario.when.interface !== "managed-api") {
-      throw new Error("failed bootstrap copy must use the managed API");
-    }
-    const rollbackWithoutInjectedFailure = {
-      ...failedCopyScenario,
-      when: {
-        ...failedCopyScenario.when,
-        input: { ...failedCopyScenario.when.input, injectCopyFailure: false },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([rollbackWithoutInjectedFailure])).toContain(
-      `${failedCopyScenario.id}: bootstrap rollback requires enabled copy-failure injection`,
-    );
-
-    const branchSelectionScenario = findScenario(
-      "identity.same-commit-different-branches-are-independent",
-    );
-    if (branchSelectionScenario.when.interface !== "managed-api") {
-      throw new Error("branch selection must use the managed API");
-    }
-    const mismatchedActiveGitBranch = {
-      ...branchSelectionScenario,
-      given: branchSelectionScenario.given.map((fact) =>
-        fact.kind === "git-state" ? { ...fact, branch: "main" } : fact,
-      ),
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([mismatchedActiveGitBranch])).toContain(
-      `${branchSelectionScenario.id}: selected context must match the active Git branch and checked-out branch fact`,
-    );
-
-    const ordinaryFolderScenario = findScenario(
-      "identity.non-git-folder-first-start-persists-identity",
-    );
-    const markerWrittenToDifferentWorkspace = {
-      ...ordinaryFolderScenario,
-      expected: {
-        ...ordinaryFolderScenario.expected,
-        writes: ordinaryFolderScenario.expected.writes.map((write) =>
-          write.target === "identity-marker"
-            ? { ...write, workspacePath: "/work/other-project" }
-            : write,
-        ),
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([markerWrittenToDifferentWorkspace])).toContain(
-      `${ordinaryFolderScenario.id}: ordinary-folder creation must persist its identity marker`,
-    );
-
-    const repositoryEqualityScenario = findScenario(
-      "api-boundary.repository-contract-is-storage-agnostic",
-    );
-    const falseRepositoryEqualityFlag = {
-      ...repositoryEqualityScenario,
-      expected: {
-        ...repositoryEqualityScenario.expected,
-        output: {
-          ...repositoryEqualityScenario.expected.output,
-          api: { ...repositoryEqualityScenario.expected.output.api, equal: false },
-        },
-      },
-    } satisfies ManagedStackContractScenario;
-    const referencedRepositoryScenario = findScenario("identity.return-to-branch-reuses-stack");
-    expect(
-      validateManagedStackContractFixtures([
-        falseRepositoryEqualityFlag,
-        referencedRepositoryScenario,
-      ]),
-    ).toContain(
-      `${repositoryEqualityScenario.id}: repository equality flags must match compared decisions`,
-    );
-
-    const portableEqualityScenario = findScenario(
-      "api-boundary.managed-surface-is-node-and-bun-portable",
-    );
-    const falsePortableEqualityFlag = {
-      ...portableEqualityScenario,
-      expected: {
-        ...portableEqualityScenario.expected,
-        output: {
-          ...portableEqualityScenario.expected.output,
-          api: { ...portableEqualityScenario.expected.output.api, equal: false },
-        },
-      },
-    } satisfies ManagedStackContractScenario;
-    const referencedPortableScenario = findScenario(
-      "identity.same-checkout-branch-and-name-reuses-stack",
-    );
-    expect(
-      validateManagedStackContractFixtures([falsePortableEqualityFlag, referencedPortableScenario]),
-    ).toContain(
-      `${portableEqualityScenario.id}: portable equality flags must match compared results`,
-    );
-
-    const actionFromUndeclaredWorkspace = {
-      ...branchSelectionScenario,
-      when: {
-        ...branchSelectionScenario.when,
-        input: { ...branchSelectionScenario.when.input, cwd: "undeclared-workspace" },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([actionFromUndeclaredWorkspace])).toContain(
-      `${branchSelectionScenario.id}: managed action cwd must match a declared workspace`,
-    );
-    const actionWithUnknownOperation = {
-      ...branchSelectionScenario,
-      when: {
-        ...branchSelectionScenario.when,
-        input: { ...branchSelectionScenario.when.input, operation: "unknown" },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([actionWithUnknownOperation])).toContain(
-      `${branchSelectionScenario.id}: resolveStack operation must be start or status`,
-    );
-
-    const portResolutionScenario = findScenario(
-      "ports.exact-default-value-differs-from-omitted-default",
-    );
-    if (portResolutionScenario.when.interface !== "managed-api") {
-      throw new Error("port-intent resolution must use the managed API");
-    }
-    const changedExplicitConfigPort = {
-      ...portResolutionScenario,
-      when: {
-        ...portResolutionScenario.when,
-        input: {
-          ...portResolutionScenario.when.input,
-          config: { "api.port": 54322 },
-        },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([changedExplicitConfigPort])).toContain(
-      `${portResolutionScenario.id}: resolved port api.port must match its input and fact`,
-    );
-  });
-
-  it("binds failure and lifecycle decisions to their declared preconditions", () => {
-    const findScenario = (id: string): ManagedStackContractScenario => {
-      const scenario = managedStackContractFixtures.find((candidate) => candidate.id === id);
-      if (scenario === undefined) {
-        throw new Error(`${id} fixture is required`);
-      }
-      return scenario;
-    };
-
-    const managedStartScenario = findScenario("ports.explicit-free-port-is-used");
-    if (managedStartScenario.when.interface !== "managed-api") {
-      throw new Error("explicit managed port start must use the managed API");
-    }
-    const actionUsingWrongManagedMethod = {
-      ...managedStartScenario,
-      when: { ...managedStartScenario.when, method: "stopStack" },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([actionUsingWrongManagedMethod])).toContain(
-      `${managedStartScenario.id}: managed action must use a declared public method`,
-    );
-
-    const runtimeConflictScenario = findScenario("runtime.explicit-and-config-conflict-fails");
-    const matchingRuntimeRequests = {
-      ...runtimeConflictScenario,
-      given: runtimeConflictScenario.given.map((fact) =>
-        fact.kind === "runtime-request" && fact.source === "config"
-          ? { ...fact, runtime: "docker" }
-          : fact,
-      ),
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([matchingRuntimeRequests])).toContain(
-      `${runtimeConflictScenario.id}: runtime conflict must bind different explicit and configured runtimes`,
-    );
-
-    const exactPortConflictScenario = findScenario("ports.explicit-port-conflict-fails");
-    const differentConfiguredConflictPort = {
-      ...exactPortConflictScenario,
-      given: exactPortConflictScenario.given.map((fact) =>
-        fact.kind === "config-port" ? { ...fact, value: 54322 } : fact,
-      ),
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([differentConfiguredConflictPort])).toContain(
-      `${exactPortConflictScenario.id}: exact port conflict must bind config, occupancy, and projections`,
-    );
-
-    const changedCredentialsScenario = findScenario(
-      "credentials.explicit-change-applies-after-stop",
-    );
-    const unchangedCredentialUpdate = {
-      ...changedCredentialsScenario,
-      given: changedCredentialsScenario.given.map((fact) =>
-        fact.kind === "credential-state"
-          ? { ...fact, valuesId: fact.previousValuesId ?? fact.valuesId }
-          : fact,
-      ),
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([unchangedCredentialUpdate])).toContain(
-      `${changedCredentialsScenario.id}: credential change requires different old and new values`,
-    );
-
-    const decodedDefaultsScenario = findScenario(
-      "ports.exact-default-value-differs-from-omitted-default",
-    );
-    if (decodedDefaultsScenario.when.interface !== "managed-api") {
-      throw new Error("decoded defaults must use the managed API");
-    }
-    const missingDecodedDefault = {
-      ...decodedDefaultsScenario,
-      when: {
-        ...decodedDefaultsScenario.when,
-        input: {
-          ...decodedDefaultsScenario.when.input,
-          decodedDefaults: { "api.port": 54321 },
-        },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([missingDecodedDefault])).toContain(
-      `${decodedDefaultsScenario.id}: decoded default keys must cover resolved port facts`,
-    );
-
-    const existingTargetScenario = findScenario("bootstrap.existing-managed-target-ignores-legacy");
-    const absentExistingTarget = {
-      ...existingTargetScenario,
-      given: existingTargetScenario.given.map((fact) =>
-        fact.kind === "managed-target" ? { ...fact, exists: false } : fact,
-      ),
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([absentExistingTarget])).toContain(
-      `${existingTargetScenario.id}: absent managed target stack-main-default contradicts an existing stack`,
-    );
-
-    const duplicateClaimScenario = findScenario("identity.copied-checkout-reports-duplicate-claim");
-    const exactDuplicateClaim = {
-      ...duplicateClaimScenario,
-      given: duplicateClaimScenario.given.map((fact) =>
-        fact.kind === "identity-claim" ? { ...fact, status: "exact" } : fact,
-      ),
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([exactDuplicateClaim])).toContain(
-      `${duplicateClaimScenario.id}: duplicate checkout error must bind both conflicting live paths`,
-    );
-
-    const inaccessiblePathScenario = findScenario("identity.inaccessible-previous-path-fails");
-    const missingPreviousPath = {
-      ...inaccessiblePathScenario,
-      given: inaccessiblePathScenario.given.map((fact) =>
-        fact.kind === "workspace" ? { ...fact, previousPathAccess: "missing" } : fact,
-      ),
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([missingPreviousPath])).toContain(
-      `${inaccessiblePathScenario.id}: inaccessible checkout error must bind path access and ambiguous claim`,
-    );
-
-    const stopScenario = findScenario("reclamation.default-stop-preserves-data");
-    const stopAlreadyStoppedStack = {
-      ...stopScenario,
-      given: stopScenario.given.map((fact) =>
-        fact.kind === "stack" ? { ...fact, lifecycle: "stopped" } : fact,
-      ),
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([stopAlreadyStoppedStack])).toContain(
-      `${stopScenario.id}: stopping stack stack-main-default requires a running lifecycle`,
-    );
-
-    const pruneScenario = findScenario("reclamation.prune-removes-metadata-only");
-    const pruneActiveRecord = {
-      ...pruneScenario,
-      given: pruneScenario.given.map((fact) =>
-        fact.kind === "managed-record" ? { ...fact, status: "active" } : fact,
-      ),
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([pruneActiveRecord])).toContain(
-      `${pruneScenario.id}: prune may delete only orphaned registry metadata`,
-    );
-    const pruneNonOrphanedStack = {
-      ...pruneScenario,
-      given: pruneScenario.given.map((fact) =>
-        fact.kind === "stack" ? { ...fact, orphaned: false } : fact,
-      ),
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([pruneNonOrphanedStack])).toContain(
-      `${pruneScenario.id}: prune may delete only orphaned registry metadata`,
-    );
-
-    const strictRuntimeScenario = findScenario("runtime.explicit-runtime-is-strict");
-    const requestedRuntimeIsAvailable = {
-      ...strictRuntimeScenario,
-      given: strictRuntimeScenario.given.map((fact) =>
-        fact.kind === "runtime-availability" && fact.runtime === "docker"
-          ? { ...fact, available: true }
-          : fact,
-      ),
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([requestedRuntimeIsAvailable])).toContain(
-      `${strictRuntimeScenario.id}: explicit runtime error must bind an unavailable requested runtime`,
-    );
-    const differentUnavailableReason = {
-      ...strictRuntimeScenario,
-      given: strictRuntimeScenario.given.map((fact) =>
-        fact.kind === "runtime-availability" && fact.runtime === "docker"
-          ? { ...fact, reason: "socket unavailable" }
-          : fact,
-      ),
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([differentUnavailableReason])).toContain(
-      `${strictRuntimeScenario.id}: explicit runtime error must bind an unavailable requested runtime`,
-    );
-
-    const unsupportedPlatformScenario = findScenario(
-      "native-qualification.unsupported-platform-fails-preflight",
-    );
-    const supportedPlatformReportedUnsupported = {
-      ...unsupportedPlatformScenario,
-      given: unsupportedPlatformScenario.given.map((fact) =>
-        fact.kind === "native-qualification" ? { ...fact, platform: "darwin-arm64" } : fact,
-      ),
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([supportedPlatformReportedUnsupported])).toContain(
-      `${unsupportedPlatformScenario.id}: unsupported native error must bind an unsupported platform`,
-    );
-  });
-
-  it("rejects lifecycle, ownership, and comparison results with contradictory evidence", () => {
-    const findScenario = (id: string): ManagedStackContractScenario => {
-      const scenario = managedStackContractFixtures.find((candidate) => candidate.id === id);
-      if (scenario === undefined) {
-        throw new Error(`${id} fixture is required`);
-      }
-      return scenario;
-    };
-
-    const runningLegacyScenario = findScenario(
-      "bootstrap.running-legacy-source-fails-without-mutation",
-    );
-    const stoppedLegacyReportedRunning = {
-      ...runningLegacyScenario,
-      given: runningLegacyScenario.given.map((fact) =>
-        fact.kind === "legacy-state" ? { ...fact, lifecycle: "stopped" } : fact,
-      ),
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([stoppedLegacyReportedRunning])).toContain(
-      `${runningLegacyScenario.id}: running legacy error requires a running source and absent target`,
-    );
-
-    const directStackScenario = findScenario("api-boundary.direct-create-stack-is-ephemeral");
-    const explicitRootsReportedTemporary = {
-      ...directStackScenario,
-      given: directStackScenario.given.map((fact) =>
-        fact.kind === "direct-stack-options" ? { ...fact, stackRoot: "explicit" } : fact,
-      ),
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([explicitRootsReportedTemporary])).toContain(
-      `${directStackScenario.id}: direct stack root inputs must agree with temporary-state behavior`,
-    );
-
-    const failedCopyScenario = findScenario("bootstrap.failed-copy-rolls-back");
-    const rollbackAgainstExistingTarget = {
-      ...failedCopyScenario,
-      given: failedCopyScenario.given.map((fact) =>
-        fact.kind === "managed-target" ? { ...fact, exists: true } : fact,
-      ),
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([rollbackAgainstExistingTarget])).toContain(
-      `${failedCopyScenario.id}: bootstrap rollback requires failure injection against an absent target`,
-    );
-
-    const stickyCollisionScenario = findScenario("ports.later-sticky-port-collision-fails");
-    const unrelatedStickyAssignment = {
-      ...stickyCollisionScenario,
-      given: stickyCollisionScenario.given.map((fact) =>
-        fact.kind === "port-assignment" ? { ...fact, port: 55422 } : fact,
-      ),
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([unrelatedStickyAssignment])).toContain(
-      `${stickyCollisionScenario.id}: sticky port conflict must bind assignment, occupancy, and projections`,
-    );
-
-    const exactPortChangeScenario = findScenario("ports.config-change-on-stopped-stack-applies");
-    const ignoredExactPortChange = {
-      ...exactPortChangeScenario,
-      given: exactPortChangeScenario.given.map((fact) =>
-        fact.kind === "config-port" ? { ...fact, value: 55322 } : fact,
-      ),
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([ignoredExactPortChange])).toContain(
-      `${exactPortChangeScenario.id}: exact port change must bind previous assignment and requested value`,
-    );
-    const automaticIntentReportedAsExactChange = {
-      ...exactPortChangeScenario,
-      given: exactPortChangeScenario.given.map((fact) =>
-        fact.kind === "config-port" ? { ...fact, intent: "automatic" } : fact,
-      ),
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([automaticIntentReportedAsExactChange])).toContain(
-      `${exactPortChangeScenario.id}: exact port change must bind previous assignment and requested value`,
-    );
-
-    const explicitRuntimeScenario = findScenario("runtime.explicit-api-overrides-auto");
-    const unavailableExplicitRuntimeStarts = {
-      ...explicitRuntimeScenario,
-      given: explicitRuntimeScenario.given.map((fact) =>
-        fact.kind === "runtime-availability" && fact.runtime === "native"
-          ? { ...fact, available: false }
-          : fact,
-      ),
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([unavailableExplicitRuntimeStarts])).toContain(
-      `${explicitRuntimeScenario.id}: successful explicit runtime requires matching availability`,
-    );
-
-    const orphanDeletionScenario = findScenario("reclamation.delete-orphan-by-stack-id");
-    const activeStackReportedOrphaned = {
-      ...orphanDeletionScenario,
-      given: orphanDeletionScenario.given.map((fact) =>
-        fact.kind === "stack" ? { ...fact, orphaned: false } : fact,
-      ),
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([activeStackReportedOrphaned])).toContain(
-      `${orphanDeletionScenario.id}: global orphan deletion requires an orphaned target`,
-    );
-
-    const legacyCredentialsScenario = findScenario(
-      "credentials.compatible-legacy-auth-is-retained",
-    );
-    const differentLegacyCredentialReference = {
-      ...legacyCredentialsScenario,
-      given: legacyCredentialsScenario.given.map((fact) =>
-        fact.kind === "credential-state" ? { ...fact, valuesId: "legacy-auth-v2" } : fact,
-      ),
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([differentLegacyCredentialReference])).toContain(
-      `${legacyCredentialsScenario.id}: copied legacy credentials must bind their persisted reference`,
-    );
-
-    const credentialDriftScenario = findScenario("credentials.running-change-reports-drift");
-    const stoppedStackReportedRunningCredentialDrift = {
-      ...credentialDriftScenario,
-      given: credentialDriftScenario.given.map((fact) =>
-        fact.kind === "stack" ? { ...fact, lifecycle: "stopped" } : fact,
-      ),
-    } satisfies ManagedStackContractScenario;
-    expect(
-      validateManagedStackContractFixtures([stoppedStackReportedRunningCredentialDrift]),
-    ).toContain(
-      `${credentialDriftScenario.id}: credential drift report requires a running selected stack`,
-    );
-
-    const idempotentDeletionScenario = findScenario("reclamation.delete-repeat-is-idempotent");
-    const activeRecordReportedDeleted = {
-      ...idempotentDeletionScenario,
-      given: idempotentDeletionScenario.given.map((fact) =>
-        fact.kind === "managed-record" ? { ...fact, status: "active" } : fact,
-      ),
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([activeRecordReportedDeleted])).toContain(
-      `${idempotentDeletionScenario.id}: idempotent deletion requires a tombstoned target`,
-    );
-
-    const folderReuseScenario = findScenario(
-      "identity.folder-to-git-exact-claim-preserves-identity",
-    );
-    const ambiguousProjectClaimReused = {
-      ...folderReuseScenario,
-      given: folderReuseScenario.given.map((fact) =>
-        fact.kind === "identity-claim" && fact.scope === "project"
-          ? { ...fact, status: "ambiguous" }
-          : fact,
-      ),
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([ambiguousProjectClaimReused])).toContain(
-      `${folderReuseScenario.id}: folder-to-Git reuse requires exact project and checkout claims`,
-    );
-
-    const repositoryMatrixScenario = findScenario(
-      "api-boundary.repository-contract-is-storage-agnostic",
-    );
-    const repositoryMatrixChangesRuntime = {
-      ...repositoryMatrixScenario,
-      given: repositoryMatrixScenario.given.map((fact) =>
-        fact.kind === "managed-api-options" && fact.repository === "in-memory"
-          ? { ...fact, runtime: "bun" }
-          : fact,
-      ),
-    } satisfies ManagedStackContractScenario;
-    expect(
-      validateManagedStackContractFixtures([
-        repositoryMatrixChangesRuntime,
-        findScenario("identity.return-to-branch-reuses-stack"),
-      ]),
-    ).toContain(
-      `${repositoryMatrixScenario.id}: repository comparison must hold runtime and state root constant`,
-    );
-
-    const portableMatrixScenario = findScenario(
-      "api-boundary.managed-surface-is-node-and-bun-portable",
-    );
-    const portableMatrixChangesRepository = {
-      ...portableMatrixScenario,
-      given: portableMatrixScenario.given.map((fact) =>
-        fact.kind === "managed-api-options" && fact.runtime === "bun"
-          ? { ...fact, repository: "persistent-adapter" }
-          : fact,
-      ),
-    } satisfies ManagedStackContractScenario;
-    expect(
-      validateManagedStackContractFixtures([
-        portableMatrixChangesRepository,
-        findScenario("identity.same-checkout-branch-and-name-reuses-stack"),
-      ]),
-    ).toContain(
-      `${portableMatrixScenario.id}: portable comparison must hold repository and state root constant`,
-    );
-  });
-
-  it("rejects observable results whose identity, runtime, port, and credential evidence disagrees", () => {
-    const findScenario = (id: string): ManagedStackContractScenario => {
-      const scenario = managedStackContractFixtures.find((candidate) => candidate.id === id);
-      if (scenario === undefined) {
-        throw new Error(`${id} fixture is required`);
-      }
-      return scenario;
-    };
-
-    const copiedBranchScenario = findScenario(
-      "identity.branch-copy-known-owner-creates-context-on-mutation",
-    );
-    const copiedBranchWithoutOriginal = {
-      ...copiedBranchScenario,
-      given: copiedBranchScenario.given.map((fact) =>
-        fact.kind === "identity-transition" && fact.operation === "branch-copy"
-          ? { ...fact, originalExists: false }
-          : fact,
-      ),
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([copiedBranchWithoutOriginal])).toContain(
-      `${copiedBranchScenario.id}: copied branch transition must match live original and checked-out branch facts`,
-    );
-
-    const configRuntimeScenario = findScenario("runtime.config-overrides-default-auto");
-    const ignoredConfigRuntime = {
-      ...configRuntimeScenario,
-      given: configRuntimeScenario.given.map((fact) =>
-        fact.kind === "runtime-request" && fact.source === "config"
-          ? { ...fact, runtime: "docker" }
-          : fact,
-      ),
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([ignoredConfigRuntime])).toContain(
-      `${configRuntimeScenario.id}: effective runtime request must match availability and successful projections`,
-    );
-
-    const runtimeDriftScenario = findScenario("runtime.status-reports-one-stack-wide-runtime");
-    const inventedRuntimeDrift = {
-      ...runtimeDriftScenario,
-      given: runtimeDriftScenario.given.map((fact) =>
-        fact.kind === "persisted-runtime" ? { ...fact, runtime: "native" } : fact,
-      ),
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([inventedRuntimeDrift])).toContain(
-      `${runtimeDriftScenario.id}: runtime drift must bind running stack, persisted runtime, config, and projections`,
-    );
-
-    const failedCopyScenario = findScenario("bootstrap.failed-copy-rolls-back");
-    const runningLegacyCopyRolledBack = {
-      ...failedCopyScenario,
-      given: failedCopyScenario.given.map((fact) =>
-        fact.kind === "legacy-state" ? { ...fact, lifecycle: "running" } : fact,
-      ),
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([runningLegacyCopyRolledBack])).toContain(
-      `${failedCopyScenario.id}: bootstrap rollback requires a compatible stopped legacy source`,
-    );
-
-    const runningLegacyScenario = findScenario(
-      "bootstrap.running-legacy-source-fails-without-mutation",
-    );
-    const existingManagedTargetReadsLegacy = {
-      ...runningLegacyScenario,
-      given: runningLegacyScenario.given.map((fact) =>
-        fact.kind === "managed-target" ? { ...fact, exists: true } : fact,
-      ),
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([existingManagedTargetReadsLegacy])).toContain(
-      `${runningLegacyScenario.id}: running legacy error requires a running source and absent target`,
-    );
-
-    const runningLegacyPortScenario = findScenario(
-      "ports.running-legacy-source-fails-before-allocation",
-    );
-    const unrelatedLegacyPortConflict = {
-      ...runningLegacyPortScenario,
-      given: runningLegacyPortScenario.given.map((fact) =>
-        fact.kind === "config-port" ? { ...fact, value: 54322 } : fact,
-      ),
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([unrelatedLegacyPortConflict])).toContain(
-      `${runningLegacyPortScenario.id}: running legacy port failure must bind config, occupancy, and projections`,
-    );
-
-    const defaultCredentialsScenario = findScenario(
-      "credentials.omitted-values-use-stable-defaults",
-    );
-    const unrelatedDefaultCredentials = {
-      ...defaultCredentialsScenario,
-      given: defaultCredentialsScenario.given.map((fact) =>
-        fact.kind === "credential-state" ? { ...fact, valuesId: "other-local-defaults" } : fact,
-      ),
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([unrelatedDefaultCredentials])).toContain(
-      `${defaultCredentialsScenario.id}: persisted credential reference and source must match declared values`,
-    );
-
-    const persistedCredentialsScenario = findScenario(
-      "credentials.unchanged-values-survive-restart",
-    );
-    const unrelatedPersistedCredentials = {
-      ...persistedCredentialsScenario,
-      given: persistedCredentialsScenario.given.map((fact) =>
-        fact.kind === "credential-state" ? { ...fact, valuesId: "other-persisted-values" } : fact,
-      ),
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([unrelatedPersistedCredentials])).toContain(
-      `${persistedCredentialsScenario.id}: persisted credential reference and source must match declared values`,
-    );
-
-    const stickyPortScenario = findScenario("ports.sticky-ports-reuse-on-return");
-    const exactAssignmentReportedSticky = {
-      ...stickyPortScenario,
-      given: stickyPortScenario.given.map((fact) =>
-        fact.kind === "port-assignment" ? { ...fact, intent: "exact" } : fact,
-      ),
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([exactAssignmentReportedSticky])).toContain(
-      `${stickyPortScenario.id}: sticky port reuse must bind automatic config, assignment, and projections`,
-    );
-
-    const removedExactPortScenario = findScenario(
-      "ports.removing-exact-key-keeps-current-port-sticky",
-    );
-    const unrelatedExactPortReportedSticky = {
-      ...removedExactPortScenario,
-      given: removedExactPortScenario.given.map((fact) =>
-        fact.kind === "config-port" ? { ...fact, previousValue: 54322 } : fact,
-      ),
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([unrelatedExactPortReportedSticky])).toContain(
-      `${removedExactPortScenario.id}: exact-to-automatic port transition must preserve the previous assignment`,
-    );
-
-    const unregisteredScenario = findScenario(
-      "identity.read-only-unregistered-checkout-does-not-write",
-    );
-    const registeredCheckoutReportedAbsent = {
-      ...unregisteredScenario,
-      given: unregisteredScenario.given.map((fact) =>
-        fact.kind === "identity-claim" ? { ...fact, status: "exact" } : fact,
-      ),
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([registeredCheckoutReportedAbsent])).toContain(
-      `${unregisteredScenario.id}: unregistered status requires an absent checkout claim`,
-    );
-
-    const markerRecoveryScenario = findScenario(
-      "identity.non-git-folder-recovers-persisted-identity",
-    );
-    const gitWorkspaceTrustsLocalMarker = {
-      ...markerRecoveryScenario,
-      given: markerRecoveryScenario.given.map((fact) =>
-        fact.kind === "workspace" ? { ...fact, mode: "git" } : fact,
-      ),
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([gitWorkspaceTrustsLocalMarker])).toContain(
-      `${markerRecoveryScenario.id}: local identity marker recovery requires an ordinary folder`,
-    );
-
-    const persistedRuntimeScenario = findScenario("runtime.persisted-runtime-reused-for-auto");
-    const siblingRuntimeReused = {
-      ...persistedRuntimeScenario,
-      given: persistedRuntimeScenario.given.map((fact) =>
-        fact.kind === "persisted-runtime" ? { ...fact, stackId: "stack-other-default" } : fact,
-      ),
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([siblingRuntimeReused])).toContain(
-      `${persistedRuntimeScenario.id}: automatic runtime must resolve from persisted state or declared availability`,
-    );
-
-    const isolatedRootScenario = findScenario(
-      "api-boundary.managed-api-accepts-isolated-state-root",
-    );
-    const defaultOptionsUseIsolatedRoot = {
-      ...isolatedRootScenario,
-      given: isolatedRootScenario.given.map((fact) =>
-        fact.kind === "managed-api-options" ? { ...fact, stateRoot: "default" } : fact,
-      ),
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([defaultOptionsUseIsolatedRoot])).toContain(
-      `${isolatedRootScenario.id}: isolated state root input must match its options and observed boundary`,
-    );
-
-    const automaticRuntimeFailureScenario = findScenario(
-      "runtime.auto-fails-when-neither-runtime-is-available",
-    );
-    const unrelatedAutomaticFailureReason = {
-      ...automaticRuntimeFailureScenario,
-      given: automaticRuntimeFailureScenario.given.map((fact) =>
-        fact.kind === "runtime-availability" && fact.runtime === "docker"
-          ? { ...fact, reason: "unrelated failure" }
-          : fact,
-      ),
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([unrelatedAutomaticFailureReason])).toContain(
-      `${automaticRuntimeFailureScenario.id}: automatic runtime failure must bind both unavailability reasons`,
-    );
-
-    const stackNamesScenario = findScenario("identity.valid-stack-names-resolve-deterministically");
-    const wrongNamedStackResult = {
-      ...stackNamesScenario,
-      expected: {
-        ...stackNamesScenario.expected,
-        output: {
-          ...stackNamesScenario.expected.output,
-          api: {
-            ...stackNamesScenario.expected.output.api,
-            default: { contextId: "context-feat", stackId: "stack-unrelated" },
-          },
-        },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([wrongNamedStackResult])).toContain(
-      `${stackNamesScenario.id}: resolved stack name default must bind its context and stack ID`,
-    );
-
-    const projectedManagedResultScenario = findScenario(
-      "api-boundary.cli-projects-shared-managed-results",
-    );
-    const tombstonedRecordReportedRunning = {
-      ...projectedManagedResultScenario,
-      given: projectedManagedResultScenario.given.map((fact) =>
-        fact.kind === "managed-record" ? { ...fact, status: "tombstoned" } : fact,
-      ),
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([tombstonedRecordReportedRunning])).toContain(
-      `${projectedManagedResultScenario.id}: projected managed status requires an active running record and persisted runtime`,
-    );
-    const stoppedStackReportedRunning = {
-      ...projectedManagedResultScenario,
-      given: projectedManagedResultScenario.given.map((fact) =>
-        fact.kind === "stack" ? { ...fact, lifecycle: "stopped" } : fact,
-      ),
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([stoppedStackReportedRunning])).toContain(
-      `${projectedManagedResultScenario.id}: projected managed status requires an active running record and persisted runtime`,
-    );
-
-    const persistedRuntimeAvailabilityScenario = findScenario(
-      "runtime.persisted-runtime-reused-for-auto",
-    );
-    const persistedRuntimeWithoutAvailability = {
-      ...persistedRuntimeAvailabilityScenario,
-      given: persistedRuntimeAvailabilityScenario.given.map((fact) =>
-        fact.kind === "runtime-availability" && fact.runtime === "native"
-          ? { ...fact, runtime: "docker" }
-          : fact,
-      ),
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([persistedRuntimeWithoutAvailability])).toContain(
-      `${persistedRuntimeAvailabilityScenario.id}: persisted automatic runtime requires matching availability evidence`,
-    );
-
-    const qualificationScenario = findScenario(
-      "native-qualification.one-service-failure-disables-platform",
-    );
-    if (qualificationScenario.when.interface !== "managed-api") {
-      throw new Error("native qualification fixture must use the managed API");
-    }
-    const unsupportedPlatformReportedAsUnqualified = {
-      ...qualificationScenario,
-      given: qualificationScenario.given.map((fact) =>
-        fact.kind === "native-qualification" ? { ...fact, platform: "darwin-x64" } : fact,
-      ),
-      when: {
-        ...qualificationScenario.when,
-        input: { ...qualificationScenario.when.input, platform: "darwin-x64" },
-      },
-      expected: {
-        ...qualificationScenario.expected,
-        output: {
-          ...qualificationScenario.expected.output,
-          api: { ...qualificationScenario.expected.output.api, platform: "darwin-x64" },
-        },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(
-      validateManagedStackContractFixtures([unsupportedPlatformReportedAsUnqualified]),
-    ).toContain(
-      `${qualificationScenario.id}: unsupported native platform must use the dedicated preflight error`,
-    );
-
-    const renameScenario = findScenario("identity.branch-rename-preserves-context");
-    const renameToUnrelatedBranch = {
-      ...renameScenario,
-      given: renameScenario.given.map((fact) =>
-        fact.kind === "identity-transition" && fact.operation === "branch-rename"
-          ? { ...fact, to: "unrelated" }
-          : fact,
-      ),
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([renameToUnrelatedBranch])).toContain(
-      `${renameScenario.id}: branch rename must match the checked-out branch and updated context owner`,
-    );
-
-    const repositoryScenario = findScenario("api-boundary.repository-contract-is-storage-agnostic");
-    const referencedRepositoryScenario = findScenario("identity.return-to-branch-reuses-stack");
-    const adaptersAgreeOnWrongContext = {
-      ...repositoryScenario,
-      expected: {
-        ...repositoryScenario.expected,
-        output: {
-          ...repositoryScenario.expected.output,
-          api: {
-            ...repositoryScenario.expected.output.api,
-            "in-memory": {
-              outcome: "reuse",
-              projectId: "project-a",
-              checkoutId: "checkout-a",
-              contextId: "context-unrelated",
-              stackId: "stack-main-default",
-              stackName: "default",
-            },
-            "persistent-adapter": {
-              outcome: "reuse",
-              projectId: "project-a",
-              checkoutId: "checkout-a",
-              contextId: "context-unrelated",
-              stackId: "stack-main-default",
-              stackName: "default",
-            },
-          },
-        },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(
-      validateManagedStackContractFixtures([
-        referencedRepositoryScenario,
-        adaptersAgreeOnWrongContext,
-      ]),
-    ).toContain(
-      `${repositoryScenario.id}: repository in-memory decision must completely match ${referencedRepositoryScenario.id}`,
-    );
-
-    const ambiguousContextScenario = findScenario("identity.branch-copy-ambiguous-read-only");
-    const independentBranchesReportedAmbiguous = {
-      ...ambiguousContextScenario,
-      given: ambiguousContextScenario.given.map((fact) =>
-        fact.kind === "branch" && fact.name === "main"
-          ? { ...fact, contextId: "context-independent" }
-          : fact,
-      ),
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([independentBranchesReportedAmbiguous])).toContain(
-      `${ambiguousContextScenario.id}: ambiguous context must bind at least two claiming branches to its projections`,
-    );
-
-    const invalidStackNameScenario = findScenario(
-      "identity.invalid-stack-name-uppercase-underscore-fails",
-    );
-    if (invalidStackNameScenario.when.interface !== "cli") {
-      throw new Error("invalid stack name fixture must use the CLI");
-    }
-    const validStackNameRejected = {
-      ...invalidStackNameScenario,
-      given: invalidStackNameScenario.given.map((fact) =>
-        fact.kind === "stack-names" ? { ...fact, names: ["review"] } : fact,
-      ),
-      when: {
-        ...invalidStackNameScenario.when,
-        argv: invalidStackNameScenario.when.argv.map((argument) =>
-          argument === "Feature_A" ? "review" : argument,
-        ),
-      },
-      expected: {
-        ...invalidStackNameScenario.expected,
-        output: {
-          ...invalidStackNameScenario.expected.output,
-          human: {
-            ...invalidStackNameScenario.expected.output.human!,
-            fields: { stack: "review" },
-          },
-          json: { ...invalidStackNameScenario.expected.output.json, stack_name: "review" },
-        },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([validStackNameRejected])).toContain(
-      `${invalidStackNameScenario.id}: invalid stack name error must bind a requested name outside the supported grammar`,
-    );
-
-    const automaticPortScenario = findScenario(
-      "ports.new-target-allocates-and-persists-omitted-ports",
-    );
-    const duplicateAutomaticPort = {
-      ...automaticPortScenario,
-      expected: {
-        ...automaticPortScenario.expected,
-        output: {
-          ...automaticPortScenario.expected.output,
-          api: {
-            ...automaticPortScenario.expected.output.api,
-            ports: { api: 55421, db: 55421 },
-          },
-        },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([duplicateAutomaticPort])).toContain(
-      `${automaticPortScenario.id}: allocated port 55421 is assigned more than once`,
-    );
-
-    const refReplacementScenario = findScenario("identity.manual-ref-replacement-orphans-context");
-    const refReplacementTargetsUnobservedCommit = {
-      ...refReplacementScenario,
-      given: refReplacementScenario.given.map((fact) =>
-        fact.kind === "identity-transition" && fact.operation === "ref-replacement"
-          ? { ...fact, to: "commit-unrelated" }
-          : fact,
-      ),
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([refReplacementTargetsUnobservedCommit])).toContain(
-      `${refReplacementScenario.id}: ref replacement target must match the action workspace commit`,
-    );
-
-    const selectorConflictScenario = findScenario(
-      "reclamation.selectors-stack-and-stack-id-conflict",
-    );
-    if (selectorConflictScenario.when.interface !== "cli") {
-      throw new Error("selector conflict fixture must use the CLI");
-    }
-    const singleSelectorReportedAsConflict = {
-      ...selectorConflictScenario,
-      when: {
-        ...selectorConflictScenario.when,
-        argv: ["stop", "--experimental", "--stack", "review"],
-      },
-      expected: {
-        ...selectorConflictScenario.expected,
-        output: {
-          ...selectorConflictScenario.expected.output,
-          human: {
-            ...selectorConflictScenario.expected.output.human!,
-            fields: { selectors: "--stack" },
-          },
-          json: { ...selectorConflictScenario.expected.output.json, selectors: ["--stack"] },
-        },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([singleSelectorReportedAsConflict])).toContain(
-      `${selectorConflictScenario.id}: selector conflict must bind at least two requested selection modes`,
-    );
-
-    const pruneScenario = findScenario("reclamation.prune-removes-metadata-only");
-    const pruneReportsDifferentRecord = {
-      ...pruneScenario,
-      expected: {
-        ...pruneScenario.expected,
-        output: {
-          ...pruneScenario.expected.output,
-          json: { ...pruneScenario.expected.output.json, pruned_records: ["stack-other"] },
-        },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([pruneReportsDifferentRecord])).toContain(
-      `${pruneScenario.id}: prune projections must match deleted registry records`,
-    );
-
-    const portableDecisionScenario = findScenario(
-      "api-boundary.managed-surface-is-node-and-bun-portable",
-    );
-    const referencedPortableDecisionScenario = findScenario(
-      "identity.same-checkout-branch-and-name-reuses-stack",
-    );
-    const portableRuntimesAgreeOnWrongContext = {
-      ...portableDecisionScenario,
-      expected: {
-        ...portableDecisionScenario.expected,
-        output: {
-          ...portableDecisionScenario.expected.output,
-          api: {
-            ...portableDecisionScenario.expected.output.api,
-            node: {
-              outcome: "report",
-              projectId: "project-a",
-              checkoutId: "checkout-a",
-              contextId: "context-unrelated",
-              stackId: "stack-main-default",
-              stackName: "default",
-            },
-            bun: {
-              outcome: "report",
-              projectId: "project-a",
-              checkoutId: "checkout-a",
-              contextId: "context-unrelated",
-              stackId: "stack-main-default",
-              stackName: "default",
-            },
-          },
-        },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(
-      validateManagedStackContractFixtures([
-        referencedPortableDecisionScenario,
-        portableRuntimesAgreeOnWrongContext,
-      ]),
-    ).toContain(
-      `${portableDecisionScenario.id}: portable node decision must completely match ${referencedPortableDecisionScenario.id}`,
-    );
-
-    const persistedRuntimeFailureScenario = findScenario(
-      "runtime.missing-persisted-prerequisite-fails",
-    );
-    const persistedRuntimeReportsUnrelatedReason = {
-      ...persistedRuntimeFailureScenario,
-      given: persistedRuntimeFailureScenario.given.map((fact) =>
-        fact.kind === "runtime-availability" && fact.runtime === "native"
-          ? { ...fact, reason: "unrelated failure" }
-          : fact,
-      ),
-    } satisfies ManagedStackContractScenario;
-    expect(
-      validateManagedStackContractFixtures([persistedRuntimeReportsUnrelatedReason]),
-    ).toContain(
-      `${persistedRuntimeFailureScenario.id}: unavailable persisted runtime must fail without switching`,
-    );
-
-    const reducedNativeGraph = {
-      ...qualificationScenario,
-      expected: {
-        ...qualificationScenario.expected,
-        output: {
-          ...qualificationScenario.expected.output,
-          api: {
-            ...qualificationScenario.expected.output.api,
-            availableServices: ["postgres"],
-          },
-        },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([reducedNativeGraph])).toContain(
-      `${qualificationScenario.id}: failed native qualification must expose no reduced service graph`,
-    );
-
-    const automaticNativeScenario = findScenario("runtime.auto-selects-fully-qualified-native");
-    const automaticNativeWithoutDockerEvidence = {
-      ...automaticNativeScenario,
-      given: automaticNativeScenario.given.filter(
-        (fact) => fact.kind !== "runtime-availability" || fact.runtime !== "docker",
-      ),
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([automaticNativeWithoutDockerEvidence])).toContain(
-      `${automaticNativeScenario.id}: automatic native fallback requires explicit Docker unavailability`,
-    );
-
-    if (qualificationScenario.expected.error === undefined) {
-      throw new Error("failed native qualification fixture must project an error");
-    }
-    const qualificationUsesUnrelatedError = {
-      ...qualificationScenario,
-      expected: {
-        ...qualificationScenario.expected,
-        error: { ...qualificationScenario.expected.error, code: "UNRELATED_ERROR" },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([qualificationUsesUnrelatedError])).toContain(
-      `${qualificationScenario.id}: supported unqualified native platform must use NATIVE_PLATFORM_NOT_QUALIFIED`,
-    );
-
-    const persistedRuntimeWithoutProvenance = {
-      ...persistedRuntimeAvailabilityScenario,
-      expected: {
-        ...persistedRuntimeAvailabilityScenario.expected,
-        output: {
-          ...persistedRuntimeAvailabilityScenario.expected.output,
-          json: { ...persistedRuntimeAvailabilityScenario.expected.output.json, persisted: false },
-        },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([persistedRuntimeWithoutProvenance])).toContain(
-      `${persistedRuntimeAvailabilityScenario.id}: automatic persisted runtime reuse must report persisted provenance`,
-    );
-
-    const siblingPortScenario = findScenario("ports.sibling-targets-allocate-independent-ports");
-    const omittedSiblingAssignment = {
-      ...siblingPortScenario,
-      given: siblingPortScenario.given.filter(
-        (fact) => fact.kind !== "port-assignment" || fact.stackId !== "stack-main-default",
-      ),
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([omittedSiblingAssignment])).toContain(
-      `${siblingPortScenario.id}: sibling allocation must bind all avoided stack IDs`,
-    );
-
-    const nonDestructiveStopScenario = findScenario("reclamation.default-stop-preserves-data");
-    const nonDestructiveStopReportsDataLoss = {
-      ...nonDestructiveStopScenario,
-      expected: {
-        ...nonDestructiveStopScenario.expected,
-        details: {
-          ...nonDestructiveStopScenario.expected.details,
-          data_preserved: false,
-          registry_record_preserved: false,
-        },
-        output: {
-          ...nonDestructiveStopScenario.expected.output,
-          human: {
-            ...nonDestructiveStopScenario.expected.output.human!,
-            fields: { dataPreserved: "false" },
-          },
-          json: {
-            ...nonDestructiveStopScenario.expected.output.json,
-            data_preserved: false,
-          },
-        },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([nonDestructiveStopReportsDataLoss])).toContain(
-      `${nonDestructiveStopScenario.id}: non-destructive stop must report preserved data and registry`,
-    );
-
-    const existingManagedTargetScenario = findScenario(
-      "bootstrap.existing-managed-target-ignores-legacy",
-    );
-    const existingTargetReportsBootstrap = {
-      ...existingManagedTargetScenario,
-      expected: {
-        ...existingManagedTargetScenario.expected,
-        details: { ...existingManagedTargetScenario.expected.details, legacy_state_read: true },
-        output: {
-          ...existingManagedTargetScenario.expected.output,
-          api: {
-            ...existingManagedTargetScenario.expected.output.api,
-            bootstrap: "copied",
-            legacyStateRead: true,
-          },
-        },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([existingTargetReportsBootstrap])).toContain(
-      `${existingManagedTargetScenario.id}: existing managed target must not report legacy bootstrap`,
-    );
-
-    const checkoutRebindScenario = findScenario("identity.missing-previous-path-rebinds-checkout");
-    const checkoutRebindWithoutRegistryUpdate = {
-      ...checkoutRebindScenario,
-      expected: {
-        ...checkoutRebindScenario.expected,
-        writes: checkoutRebindScenario.expected.writes.filter(
-          (write) => write.target !== "registry",
-        ),
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([checkoutRebindWithoutRegistryUpdate])).toContain(
-      `${checkoutRebindScenario.id}: automatic checkout rebind must persist the checkout registry update`,
-    );
-
-    const exactPortChangeScenario = findScenario("ports.config-change-on-stopped-stack-applies");
-    const exactPortChangeWithoutPersistence = {
-      ...exactPortChangeScenario,
-      expected: {
-        ...exactPortChangeScenario.expected,
-        writes: exactPortChangeScenario.expected.writes.filter(
-          (write) => write.target !== "managed-state",
-        ),
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([exactPortChangeWithoutPersistence])).toContain(
-      `${exactPortChangeScenario.id}: exact port change must persist assignment before runtime start`,
-    );
-
-    const engineScopedStopScenario = findScenario("reclamation.stop-is-engine-scoped");
-    const engineScopedStopWithoutRunningLegacy = {
-      ...engineScopedStopScenario,
-      given: engineScopedStopScenario.given.map((fact) =>
-        fact.kind === "legacy-state" ? { ...fact, lifecycle: "stopped" } : fact,
-      ),
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([engineScopedStopWithoutRunningLegacy])).toContain(
-      `${engineScopedStopScenario.id}: engine-scoped stop requires a simultaneously running legacy stack`,
-    );
-
-    const bootstrapCopyScenario = findScenario(
-      "bootstrap.first-start-copies-compatible-legacy-state",
-    );
-    const bootstrapCopyReportsLegacyMutation = {
-      ...bootstrapCopyScenario,
-      expected: {
-        ...bootstrapCopyScenario.expected,
-        details: { ...bootstrapCopyScenario.expected.details, legacy_state_mutated: true },
-        output: {
-          ...bootstrapCopyScenario.expected.output,
-          json: {
-            ...bootstrapCopyScenario.expected.output.json,
-            legacy_state_mutated: true,
-          },
-        },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([bootstrapCopyReportsLegacyMutation])).toContain(
-      `${bootstrapCopyScenario.id}: bootstrap copy must not mutate legacy state`,
-    );
-
-    const pruneReportsMutableDataDeletion = {
-      ...pruneScenario,
-      expected: {
-        ...pruneScenario.expected,
-        details: { ...pruneScenario.expected.details, mutable_data_deleted: true },
-        output: {
-          ...pruneScenario.expected.output,
-          human: {
-            ...pruneScenario.expected.output.human!,
-            fields: { dataDeleted: "true" },
-          },
-          json: { ...pruneScenario.expected.output.json, mutable_data_deleted: true },
-        },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([pruneReportsMutableDataDeletion])).toContain(
-      `${pruneScenario.id}: prune must preserve mutable stack data`,
-    );
-
-    const directCreateScenario = findScenario("api-boundary.direct-create-stack-is-ephemeral");
-    const directCreateReportsManagedSideEffects = {
-      ...directCreateScenario,
-      expected: {
-        ...directCreateScenario.expected,
-        details: {
-          ...directCreateScenario.expected.details,
-          git_inspected: true,
-          identity_marker_created: true,
-          global_registry_mutated: true,
-        },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([directCreateReportsManagedSideEffects])).toContain(
-      `${directCreateScenario.id}: direct createStack must remain isolated from managed state`,
-    );
-
-    const contextualSelectionScenario = findScenario(
-      "identity.same-checkout-branch-and-name-reuses-stack",
-    );
-    const contextualSelectionWithoutCheckedOutBranch = {
-      ...contextualSelectionScenario,
-      given: contextualSelectionScenario.given.map((fact) =>
-        fact.kind === "branch" ? { ...fact, checkedOut: false } : fact,
-      ),
-    } satisfies ManagedStackContractScenario;
-    expect(
-      validateManagedStackContractFixtures([contextualSelectionWithoutCheckedOutBranch]),
-    ).toContain(
-      `${contextualSelectionScenario.id}: contextual Git selection requires exactly one checked-out branch`,
-    );
-
-    const failedBootstrapScenario = findScenario("bootstrap.failed-copy-rolls-back");
-    const failedBootstrapReportsPublishedPartialTarget = {
-      ...failedBootstrapScenario,
-      expected: {
-        ...failedBootstrapScenario.expected,
-        details: {
-          ...failedBootstrapScenario.expected.details,
-          active_target_exists: true,
-          registry_record_published: true,
-        },
-        output: {
-          ...failedBootstrapScenario.expected.output,
-          api: {
-            ...failedBootstrapScenario.expected.output.api,
-            activeTargetExists: true,
-            registryRecordPublished: true,
-          },
-        },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(
-      validateManagedStackContractFixtures([failedBootstrapReportsPublishedPartialTarget]),
-    ).toContain(
-      `${failedBootstrapScenario.id}: bootstrap rollback must leave no published partial target`,
-    );
-
-    const stableCredentialsScenario = findScenario(
-      "credentials.omitted-values-use-stable-defaults",
-    );
-    const omittedCredentialsRotatePerStart = {
-      ...stableCredentialsScenario,
-      expected: {
-        ...stableCredentialsScenario.expected,
-        details: { ...stableCredentialsScenario.expected.details, generated_per_start: true },
-        output: {
-          ...stableCredentialsScenario.expected.output,
-          json: { ...stableCredentialsScenario.expected.output.json, credentials_stable: false },
-        },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([omittedCredentialsRotatePerStart])).toContain(
-      `${stableCredentialsScenario.id}: omitted credentials must reuse stable local defaults`,
-    );
-
-    const trackedMarkerScenario = findScenario("identity.fresh-clone-ignores-tracked-marker");
-    const trackedMarkerReportedMutable = {
-      ...trackedMarkerScenario,
-      expected: {
-        ...trackedMarkerScenario.expected,
-        details: {
-          ...trackedMarkerScenario.expected.details,
-          tracked_marker_ignored: false,
-          tracked_marker_mutated: true,
-        },
-        output: {
-          ...trackedMarkerScenario.expected.output,
-          json: {
-            ...trackedMarkerScenario.expected.output.json,
-            tracked_marker_ignored: false,
-          },
-        },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([trackedMarkerReportedMutable])).toContain(
-      `${trackedMarkerScenario.id}: tracked identity marker must remain ignored and unmodified`,
-    );
-
-    const runtimeStatusScenario = findScenario("runtime.status-reports-one-stack-wide-runtime");
-    const runtimeStatusReportsMixedGraph = {
-      ...runtimeStatusScenario,
-      expected: {
-        ...runtimeStatusScenario.expected,
-        details: { ...runtimeStatusScenario.expected.details, mixed_runtime: true },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([runtimeStatusReportsMixedGraph])).toContain(
-      `${runtimeStatusScenario.id}: runtime drift must bind running stack, persisted runtime, config, and projections`,
-    );
-
-    const folderToGitAmbiguityScenario = findScenario(
-      "identity.folder-to-git-ambiguous-claim-fails",
-    );
-    const exactFolderClaimReportedAmbiguous = {
-      ...folderToGitAmbiguityScenario,
-      given: folderToGitAmbiguityScenario.given.map((fact) =>
-        fact.kind === "identity-claim" ? { ...fact, status: "exact" } : fact,
-      ),
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([exactFolderClaimReportedAmbiguous])).toContain(
-      `${folderToGitAmbiguityScenario.id}: folder-to-Git ambiguity error requires an ambiguous live project claim`,
-    );
-
-    const strictRuntimeScenario = findScenario("runtime.explicit-runtime-is-strict");
-    const strictRuntimeWithoutAvailableAlternative = {
-      ...strictRuntimeScenario,
-      given: strictRuntimeScenario.given.filter(
-        (fact) => fact.kind !== "runtime-availability" || fact.runtime !== "native",
-      ),
-    } satisfies ManagedStackContractScenario;
-    expect(
-      validateManagedStackContractFixtures([strictRuntimeWithoutAvailableAlternative]),
-    ).toContain(
-      `${strictRuntimeScenario.id}: explicit runtime error must bind an unavailable requested runtime`,
-    );
-
-    const repeatedDeletionScenario = findScenario("reclamation.delete-repeat-is-idempotent");
-    const repeatedDeletionHidesIdempotency = {
-      ...repeatedDeletionScenario,
-      expected: {
-        ...repeatedDeletionScenario.expected,
-        details: { ...repeatedDeletionScenario.expected.details, idempotent: false },
-        output: {
-          ...repeatedDeletionScenario.expected.output,
-          json: { ...repeatedDeletionScenario.expected.output.json, already_deleted: false },
-        },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([repeatedDeletionHidesIdempotency])).toContain(
-      `${repeatedDeletionScenario.id}: idempotent deletion requires a tombstoned target`,
-    );
-
-    const standaloneGitScenario = findScenario("identity.branch-create-and-switch-is-no-op");
-    const standaloneGitReportsManagedDeletion = {
-      ...standaloneGitScenario,
-      expected: {
-        ...standaloneGitScenario.expected,
-        outcome: "delete",
-        details: { ...standaloneGitScenario.expected.details, managed_command_ran: true },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([standaloneGitReportsManagedDeletion])).toContain(
-      `${standaloneGitScenario.id}: standalone Git action must remain outside managed lifecycle`,
-    );
-
-    const recreatedBranchScenario = findScenario("identity.branch-delete-recreate-creates-context");
-    const recreatedBranchOrphansUnrelatedContext = {
-      ...recreatedBranchScenario,
-      expected: {
-        ...recreatedBranchScenario.expected,
-        details: { ...recreatedBranchScenario.expected.details, orphaned_context_id: "context-x" },
-        output: {
-          ...recreatedBranchScenario.expected.output,
-          json: {
-            ...recreatedBranchScenario.expected.output.json,
-            orphaned_context_id: "context-x",
-          },
-        },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(
-      validateManagedStackContractFixtures([recreatedBranchOrphansUnrelatedContext]),
-    ).toContain(
-      `${recreatedBranchScenario.id}: branch recreation must orphan the displaced context`,
-    );
-
-    const incompatibleLegacyScenario = findScenario("bootstrap.incompatible-legacy-starts-fresh");
-    const incompatibleLegacyReportsMutation = {
-      ...incompatibleLegacyScenario,
-      expected: {
-        ...incompatibleLegacyScenario.expected,
-        details: { ...incompatibleLegacyScenario.expected.details, legacy_state_mutated: true },
-        output: {
-          ...incompatibleLegacyScenario.expected.output,
-          json: {
-            ...incompatibleLegacyScenario.expected.output.json,
-            legacy_state_mutated: true,
-          },
-        },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([incompatibleLegacyReportsMutation])).toContain(
-      `${incompatibleLegacyScenario.id}: fresh bootstrap must not copy or mutate legacy state`,
-    );
-
-    const restartPersistedCredentialsScenario = findScenario(
-      "credentials.unchanged-values-survive-restart",
-    );
-    const persistedCredentialsRotate = {
-      ...restartPersistedCredentialsScenario,
-      expected: {
-        ...restartPersistedCredentialsScenario.expected,
-        details: {
-          ...restartPersistedCredentialsScenario.expected.details,
-          credentials_rotated: true,
-        },
-        output: {
-          ...restartPersistedCredentialsScenario.expected.output,
-          json: {
-            ...restartPersistedCredentialsScenario.expected.output.json,
-            credentials_unchanged: false,
-          },
-        },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([persistedCredentialsRotate])).toContain(
-      `${restartPersistedCredentialsScenario.id}: persisted credentials must survive restart unchanged`,
-    );
-
-    const freshAutomaticRuntimeScenario = findScenario("runtime.auto-prefers-docker");
-    const freshAutomaticRuntimeIsNotPersisted = {
-      ...freshAutomaticRuntimeScenario,
-      expected: {
-        ...freshAutomaticRuntimeScenario.expected,
-        details: { ...freshAutomaticRuntimeScenario.expected.details, persisted: false },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([freshAutomaticRuntimeIsNotPersisted])).toContain(
-      `${freshAutomaticRuntimeScenario.id}: fresh automatic runtime selection must be persisted`,
-    );
-
-    const resolvedNamesScenario = findScenario(
-      "identity.valid-stack-names-resolve-deterministically",
-    );
-    if (resolvedNamesScenario.when.interface !== "managed-api") {
-      throw new Error(
-        "identity.valid-stack-names-resolve-deterministically managed API fixture is required",
-      );
-    }
-    const resolvedNamesIncludeInvalidName = {
-      ...resolvedNamesScenario,
-      given: resolvedNamesScenario.given.map((fact) =>
-        fact.kind === "stack-names" ? { ...fact, names: ["default", "Review_42"] } : fact,
-      ),
-      when: {
-        ...resolvedNamesScenario.when,
-        input: { ...resolvedNamesScenario.when.input, stackNames: ["default", "Review_42"] },
-      },
-      expected: {
-        ...resolvedNamesScenario.expected,
-        details: {
-          default: "stack-feat-default",
-          Review_42: "stack-feat-review-42",
-        },
-        output: {
-          ...resolvedNamesScenario.expected.output,
-          api: {
-            default: { contextId: "context-feat", stackId: "stack-feat-default" },
-            Review_42: { contextId: "context-feat", stackId: "stack-feat-review-42" },
-          },
-        },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([resolvedNamesIncludeInvalidName])).toContain(
-      `${resolvedNamesScenario.id}: requested stack names must match their fact and projected results`,
-    );
-
-    const runtimeInputScenario = findScenario("runtime.explicit-api-overrides-auto");
-    const authInputScenario = findScenario("credentials.configured-values-are-authoritative");
-    const portInputScenario = findScenario("ports.explicit-free-port-is-used");
-    if (
-      runtimeInputScenario.when.interface !== "managed-api" ||
-      authInputScenario.when.interface !== "managed-api" ||
-      portInputScenario.when.interface !== "managed-api"
-    ) {
-      throw new Error("startStack managed API fixtures are required");
-    }
-    const invalidStartInputs = [
       {
-        ...runtimeInputScenario,
-        when: {
-          ...runtimeInputScenario.when,
-          input: { ...runtimeInputScenario.when.input, runtime: 42 },
-        },
-      } satisfies ManagedStackContractScenario,
+        fixtures: [
+          {
+            ...reuse,
+            expected: {
+              ...reuse.expected,
+              selection: { ...reuse.expected.selection, stackId: "stack-undeclared" },
+            },
+          },
+        ],
+        expectedError: `${reuse.id}: selection references undeclared ID stack-undeclared`,
+      },
       {
-        ...authInputScenario,
-        when: {
-          ...authInputScenario.when,
-          input: { ...authInputScenario.when.input, auth: 42 },
-        },
-      } satisfies ManagedStackContractScenario,
+        fixtures: [
+          {
+            ...reuse,
+            expected: {
+              ...reuse.expected,
+              output: {
+                ...reuse.expected.output,
+                json: { ...reuse.expected.output.json, outcome: "create" },
+              },
+            },
+          },
+        ],
+        expectedError: `${reuse.id}: projected outcome disagrees with the managed result`,
+      },
       {
-        ...portInputScenario,
-        when: {
-          ...portInputScenario.when,
-          input: { ...portInputScenario.when.input, portIntents: "invalid" },
-        },
-      } satisfies ManagedStackContractScenario,
+        fixtures: [
+          {
+            ...portConflict,
+            expected: {
+              ...portConflict.expected,
+              error: { ...portConflict.expected.error, code: "exact_port_occupied" },
+              output: {
+                ...portConflict.expected.output,
+                json: { ...portConflict.expected.output.json, code: "exact_port_occupied" },
+              },
+            },
+          },
+        ],
+        expectedError: `${portConflict.id}: diagnostic code exact_port_occupied must use SCREAMING_SNAKE_CASE`,
+      },
+      {
+        fixtures: [
+          {
+            ...readOnly,
+            expected: {
+              ...readOnly.expected,
+              writes: [{ target: "registry", operation: "update", id: "context-main" }],
+            },
+          },
+        ],
+        expectedError: `${readOnly.id}: report outcome must not mutate state`,
+      },
+      {
+        fixtures: [reuse, reuse],
+        expectedError: `${reuse.id}: duplicate scenario ID`,
+      },
     ];
-    for (const invalidStartInput of invalidStartInputs) {
-      expect(validateManagedStackContractFixtures([invalidStartInput])).toContain(
-        `${invalidStartInput.id}: managed action must use a declared public method`,
+
+    for (const testCase of cases) {
+      expect(validateManagedStackContractFixtures(testCase.fixtures)).toContain(
+        testCase.expectedError,
       );
     }
-
-    const nativeAutomaticScenario = findScenario("runtime.auto-selects-fully-qualified-native");
-    const nativeAutomaticReportsPartialGraph = {
-      ...nativeAutomaticScenario,
-      expected: {
-        ...nativeAutomaticScenario.expected,
-        details: {
-          ...nativeAutomaticScenario.expected.details,
-          qualified_service_count: 1,
-          mixed_runtime: true,
-        },
-        output: {
-          ...nativeAutomaticScenario.expected.output,
-          api: { ...nativeAutomaticScenario.expected.output.api, qualifiedServiceCount: 1 },
-        },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([nativeAutomaticReportsPartialGraph])).toContain(
-      `${nativeAutomaticScenario.id}: automatic native selection must bind the full qualified graph`,
-    );
-
-    const symlinkAliasScenario = findScenario("identity.symlink-alias-reuses-checkout");
-    const symlinkAliasReportsUnrelatedPath = {
-      ...symlinkAliasScenario,
-      expected: {
-        ...symlinkAliasScenario.expected,
-        output: {
-          ...symlinkAliasScenario.expected.output,
-          json: { ...symlinkAliasScenario.expected.output.json, canonical_path: "/other/project" },
-        },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([symlinkAliasReportsUnrelatedPath])).toContain(
-      `${symlinkAliasScenario.id}: symlink alias must report its canonical checkout path`,
-    );
-
-    const firstDeletionScenario = findScenario("reclamation.delete-orphan-by-stack-id");
-    const firstDeletionHidesTombstone = {
-      ...firstDeletionScenario,
-      expected: {
-        ...firstDeletionScenario.expected,
-        details: { ...firstDeletionScenario.expected.details, tombstoned: false },
-        output: {
-          ...firstDeletionScenario.expected.output,
-          human: {
-            ...firstDeletionScenario.expected.output.human!,
-            fields: {
-              ...firstDeletionScenario.expected.output.human!.fields,
-              tombstoned: "false",
-            },
-          },
-          json: { ...firstDeletionScenario.expected.output.json, tombstoned: false },
-        },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([firstDeletionHidesTombstone])).toContain(
-      `${firstDeletionScenario.id}: registry tombstone must be reported by every projection`,
-    );
-
-    const injectedServiceScenario = findScenario(
-      "api-boundary.managed-api-accepts-injected-repository",
-    );
-    const injectedServiceRequiresCli = {
-      ...injectedServiceScenario,
-      expected: {
-        ...injectedServiceScenario.expected,
-        details: { ...injectedServiceScenario.expected.details, cli_required: true },
-        output: {
-          ...injectedServiceScenario.expected.output,
-          api: { ...injectedServiceScenario.expected.output.api, cliRequired: true },
-        },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([injectedServiceRequiresCli])).toContain(
-      `${injectedServiceScenario.id}: injected repository and state root must match the observed managed service`,
-    );
-
-    const sameCommitScenario = findScenario(
-      "identity.same-commit-different-branches-are-independent",
-    );
-    const sameCommitScenarioHasDifferentCheckedOutCommit = {
-      ...sameCommitScenario,
-      given: sameCommitScenario.given.map((fact) =>
-        fact.kind === "git-state" ? { ...fact, commit: "unrelated-commit" } : fact,
-      ),
-    } satisfies ManagedStackContractScenario;
-    expect(
-      validateManagedStackContractFixtures([sameCommitScenarioHasDifferentCheckedOutCommit]),
-    ).toContain(
-      `${sameCommitScenario.id}: branch comparison must prove both refs share the checked-out commit`,
-    );
-
-    const ordinaryFolderScenario = findScenario(
-      "identity.non-git-folder-first-start-persists-identity",
-    );
-    const ordinaryFolderTracksMarker = {
-      ...ordinaryFolderScenario,
-      expected: {
-        ...ordinaryFolderScenario.expected,
-        details: { ...ordinaryFolderScenario.expected.details, identity_marker_tracked: true },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([ordinaryFolderTracksMarker])).toContain(
-      `${ordinaryFolderScenario.id}: ordinary-folder identity marker must remain untracked`,
-    );
-
-    const runningLegacyNoMutationScenario = findScenario(
-      "bootstrap.running-legacy-source-fails-without-mutation",
-    );
-    const runningLegacyReportsPartialTarget = {
-      ...runningLegacyNoMutationScenario,
-      expected: {
-        ...runningLegacyNoMutationScenario.expected,
-        details: {
-          ...runningLegacyNoMutationScenario.expected.details,
-          managed_target_published: true,
-          partial_state: true,
-        },
-        output: {
-          ...runningLegacyNoMutationScenario.expected.output,
-          json: {
-            ...runningLegacyNoMutationScenario.expected.output.json,
-            managed_target_published: true,
-          },
-        },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([runningLegacyReportsPartialTarget])).toContain(
-      `${runningLegacyNoMutationScenario.id}: running legacy error must leave no partial managed target`,
-    );
-
-    const copiedBranchWarningScenario = findScenario(
-      "identity.branch-copy-read-only-does-not-write",
-    );
-    const copiedBranchWarningReportsUnrelatedOwners = {
-      ...copiedBranchWarningScenario,
-      expected: {
-        ...copiedBranchWarningScenario.expected,
-        output: {
-          ...copiedBranchWarningScenario.expected.output,
-          json: {
-            ...copiedBranchWarningScenario.expected.output.json,
-            branch: "unrelated-branch",
-            owner: "unrelated-owner",
-          },
-        },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(
-      validateManagedStackContractFixtures([copiedBranchWarningReportsUnrelatedOwners]),
-    ).toContain(
-      `${copiedBranchWarningScenario.id}: copied branch warning must bind observed branch ownership`,
-    );
-
-    const directDisposalScenario = findScenario(
-      "api-boundary.direct-dispose-removes-temporary-roots",
-    );
-    const directDisposalLeaksTemporaryRoots = {
-      ...directDisposalScenario,
-      expected: {
-        ...directDisposalScenario.expected,
-        writes: [],
-        details: { ...directDisposalScenario.expected.details, temporary_roots_removed: false },
-        output: {
-          ...directDisposalScenario.expected.output,
-          api: {
-            ...directDisposalScenario.expected.output.api,
-            temporaryRootsRemoved: false,
-          },
-        },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([directDisposalLeaksTemporaryRoots])).toContain(
-      `${directDisposalScenario.id}: direct stack disposal must remove omitted temporary roots`,
-    );
-
-    const partialDirectRootsScenario = findScenario(
-      "api-boundary.direct-create-stack-keeps-omitted-runtime-root-temporary",
-    );
-    const partialDirectRootsLeakRuntimeRoot = {
-      ...partialDirectRootsScenario,
-      expected: {
-        ...partialDirectRootsScenario.expected,
-        writes: [],
-        details: { ...partialDirectRootsScenario.expected.details, temporary_roots: [] },
-        output: {
-          ...partialDirectRootsScenario.expected.output,
-          api: { ...partialDirectRootsScenario.expected.output.api, temporaryRoots: [] },
-        },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([partialDirectRootsLeakRuntimeRoot])).toContain(
-      `${partialDirectRootsScenario.id}: direct stack root inputs must agree with temporary-state behavior`,
-    );
-
-    const copiedBranchCreateScenario = findScenario(
-      "identity.branch-copy-known-owner-creates-context-on-mutation",
-    );
-    const copiedBranchCreateReportsUnrelatedAncestry = {
-      ...copiedBranchCreateScenario,
-      expected: {
-        ...copiedBranchCreateScenario.expected,
-        details: {
-          ...copiedBranchCreateScenario.expected.details,
-          original_context_id: "context-unrelated",
-        },
-        output: {
-          ...copiedBranchCreateScenario.expected.output,
-          json: {
-            ...copiedBranchCreateScenario.expected.output.json,
-            original_context_id: "context-unrelated",
-          },
-        },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(
-      validateManagedStackContractFixtures([copiedBranchCreateReportsUnrelatedAncestry]),
-    ).toContain(
-      `${copiedBranchCreateScenario.id}: copied branch creation must bind its original context`,
-    );
-
-    const retryableBootstrapScenario = findScenario("bootstrap.failed-copy-rolls-back");
-    const failedBootstrapIsNotRetryable = {
-      ...retryableBootstrapScenario,
-      expected: {
-        ...retryableBootstrapScenario.expected,
-        output: {
-          ...retryableBootstrapScenario.expected.output,
-          api: { ...retryableBootstrapScenario.expected.output.api, retryable: false },
-        },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([failedBootstrapIsNotRetryable])).toContain(
-      `${retryableBootstrapScenario.id}: bootstrap rollback must leave no published partial target`,
-    );
-
-    const bareWorktreeScenario = findScenario(
-      "identity.bare-repository-linked-worktrees-share-project",
-    );
-    const bareWorktreeRequiresPrimary = {
-      ...bareWorktreeScenario,
-      expected: {
-        ...bareWorktreeScenario.expected,
-        output: {
-          ...bareWorktreeScenario.expected.output,
-          api: { ...bareWorktreeScenario.expected.output.api, primaryWorktreeRequired: true },
-        },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([bareWorktreeRequiresPrimary])).toContain(
-      `${bareWorktreeScenario.id}: bare-repository worktree must not require a primary worktree`,
-    );
-
-    const missingOriginalScenario = findScenario("identity.original-gone-turns-copy-into-rename");
-    const missingOriginalHidesRename = {
-      ...missingOriginalScenario,
-      expected: {
-        ...missingOriginalScenario.expected,
-        output: {
-          ...missingOriginalScenario.expected.output,
-          json: { ...missingOriginalScenario.expected.output.json, rename_detected: false },
-        },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([missingOriginalHidesRename])).toContain(
-      `${missingOriginalScenario.id}: missing original branch must be reported as a rename`,
-    );
-
-    const divergedTimelineScenario = findScenario(
-      "bootstrap.managed-and-legacy-diverge-after-copy",
-    );
-    const divergedTimelineReportsSynchronization = {
-      ...divergedTimelineScenario,
-      expected: {
-        ...divergedTimelineScenario.expected,
-        details: { ...divergedTimelineScenario.expected.details, timelines_diverged: false },
-        output: {
-          ...divergedTimelineScenario.expected.output,
-          json: { ...divergedTimelineScenario.expected.output.json, timelines_diverged: false },
-        },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(
-      validateManagedStackContractFixtures([divergedTimelineReportsSynchronization]),
-    ).toContain(
-      `${divergedTimelineScenario.id}: managed restart must report legacy timeline divergence`,
-    );
-
-    const cliProjectionScenario = findScenario("api-boundary.cli-projects-shared-managed-results");
-    const cliReimplementsManagedDecisions = {
-      ...cliProjectionScenario,
-      expected: {
-        ...cliProjectionScenario.expected,
-        details: {
-          ...cliProjectionScenario.expected.details,
-          managed_result_projected: false,
-          identity_decisions_in_cli: 1,
-        },
-        output: {
-          ...cliProjectionScenario.expected.output,
-          human: {
-            ...cliProjectionScenario.expected.output.human!,
-            fields: { stackId: "stack-main-default", runtime: "native" },
-          },
-          json: { ...cliProjectionScenario.expected.output.json, runtime: "native" },
-        },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([cliReimplementsManagedDecisions])).toContain(
-      `${cliProjectionScenario.id}: projected managed status requires an active running record and persisted runtime`,
-    );
-
-    const folderConversionScenario = findScenario(
-      "identity.folder-to-git-exact-claim-preserves-identity",
-    );
-    const folderConversionLosesTransition = {
-      ...folderConversionScenario,
-      given: folderConversionScenario.given.filter(
-        (fact) => fact.kind !== "identity-transition" || fact.operation !== "folder-to-git",
-      ),
-      expected: {
-        ...folderConversionScenario.expected,
-        output: {
-          ...folderConversionScenario.expected.output,
-          json: { ...folderConversionScenario.expected.output.json, converted_to_git: false },
-        },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([folderConversionLosesTransition])).toContain(
-      `${folderConversionScenario.id}: folder-to-Git result must bind the workspace transition`,
-    );
-
-    const replacedRefOrphanScenario = findScenario(
-      "identity.manual-ref-replacement-orphans-context",
-    );
-    const replacedRefOrphansUnrelatedContext = {
-      ...replacedRefOrphanScenario,
-      expected: {
-        ...replacedRefOrphanScenario.expected,
-        details: {
-          ...replacedRefOrphanScenario.expected.details,
-          orphaned_context_id: "context-unrelated",
-        },
-        output: {
-          ...replacedRefOrphanScenario.expected.output,
-          json: {
-            ...replacedRefOrphanScenario.expected.output.json,
-            orphaned_context_id: "context-unrelated",
-          },
-        },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([replacedRefOrphansUnrelatedContext])).toContain(
-      `${replacedRefOrphanScenario.id}: ref replacement must orphan the displaced branch context`,
-    );
-
-    const managedReuseLegacyScenario = findScenario(
-      "bootstrap.managed-and-legacy-diverge-after-copy",
-    );
-    const managedReuseMutatesLegacy = {
-      ...managedReuseLegacyScenario,
-      expected: {
-        ...managedReuseLegacyScenario.expected,
-        details: {
-          ...managedReuseLegacyScenario.expected.details,
-          legacy_state_mutated: true,
-        },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([managedReuseMutatesLegacy])).toContain(
-      `${managedReuseLegacyScenario.id}: managed target reuse must not mutate legacy state`,
-    );
-
-    const readOnlyUnregisteredScenario = findScenario(
-      "identity.read-only-unregistered-checkout-does-not-write",
-    );
-    const readOnlyDiscoveryReportsRegistration = {
-      ...readOnlyUnregisteredScenario,
-      expected: {
-        ...readOnlyUnregisteredScenario.expected,
-        details: {
-          ...readOnlyUnregisteredScenario.expected.details,
-          registered: true,
-          identity_marker_created: true,
-        },
-        output: {
-          ...readOnlyUnregisteredScenario.expected.output,
-          json: { ...readOnlyUnregisteredScenario.expected.output.json, registered: true },
-        },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([readOnlyDiscoveryReportsRegistration])).toContain(
-      `${readOnlyUnregisteredScenario.id}: read-only unregistered status must not create identity state`,
-    );
-
-    const stickyReturnScenario = findScenario("ports.sticky-ports-reuse-on-return");
-    const stickyReturnHidesPersistence = {
-      ...stickyReturnScenario,
-      expected: {
-        ...stickyReturnScenario.expected,
-        output: {
-          ...stickyReturnScenario.expected.output,
-          json: { ...stickyReturnScenario.expected.output.json, sticky: false },
-        },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([stickyReturnHidesPersistence])).toContain(
-      `${stickyReturnScenario.id}: sticky port reuse must bind automatic config, assignment, and projections`,
-    );
-
-    const unavailablePersistedRuntimeScenario = findScenario(
-      "runtime.missing-persisted-prerequisite-fails",
-    );
-    const unavailablePersistedRuntimeReportsSwitch = {
-      ...unavailablePersistedRuntimeScenario,
-      expected: {
-        ...unavailablePersistedRuntimeScenario.expected,
-        details: {
-          ...unavailablePersistedRuntimeScenario.expected.details,
-          switched_to_docker: true,
-        },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(
-      validateManagedStackContractFixtures([unavailablePersistedRuntimeReportsSwitch]),
-    ).toContain(
-      `${unavailablePersistedRuntimeScenario.id}: unavailable persisted runtime must not report a switch`,
-    );
-
-    const unnamedFreshCloneScenario = findScenario(
-      "identity.fresh-clone-creates-project-and-checkout",
-    );
-    if (unnamedFreshCloneScenario.expected.selection === undefined) {
-      throw new Error("fresh clone selection is required");
-    }
-    const unnamedFreshCloneSelectsReview = {
-      ...unnamedFreshCloneScenario,
-      expected: {
-        ...unnamedFreshCloneScenario.expected,
-        selection: { ...unnamedFreshCloneScenario.expected.selection, stackName: "review" },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([unnamedFreshCloneSelectsReview])).toContain(
-      `${unnamedFreshCloneScenario.id}: unnamed CLI start must select the default stack`,
-    );
-
-    const freshCloneMutatesIndex = {
-      ...unnamedFreshCloneScenario,
-      expected: {
-        ...unnamedFreshCloneScenario.expected,
-        details: { ...unnamedFreshCloneScenario.expected.details, git_index_mutated: true },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([freshCloneMutatesIndex])).toContain(
-      `${unnamedFreshCloneScenario.id}: fresh clone identity creation must not mutate the Git index`,
-    );
-
-    const exactPortProjectionScenario = findScenario("ports.explicit-free-port-is-used");
-    const exactPortProjectsAutomaticIntent = {
-      ...exactPortProjectionScenario,
-      expected: {
-        ...exactPortProjectionScenario.expected,
-        output: {
-          ...exactPortProjectionScenario.expected.output,
-          api: { ...exactPortProjectionScenario.expected.output.api, intent: "automatic" },
-        },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([exactPortProjectsAutomaticIntent])).toContain(
-      `${exactPortProjectionScenario.id}: exact port request api.port must project exact intent`,
-    );
-
-    const metadataPruneScenario = findScenario("reclamation.prune-removes-metadata-only");
-    const metadataPruneReportsNoRemoval = {
-      ...metadataPruneScenario,
-      expected: {
-        ...metadataPruneScenario.expected,
-        details: { ...metadataPruneScenario.expected.details, metadata_removed: false },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([metadataPruneReportsNoRemoval])).toContain(
-      `${metadataPruneScenario.id}: registry deletion must report removed metadata`,
-    );
-
-    const checkoutIndependentDeletionScenario = findScenario(
-      "reclamation.delete-orphan-by-stack-id",
-    );
-    const globalDeletionRequiresCheckout = {
-      ...checkoutIndependentDeletionScenario,
-      expected: {
-        ...checkoutIndependentDeletionScenario.expected,
-        details: {
-          ...checkoutIndependentDeletionScenario.expected.details,
-          checkout_required: true,
-        },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([globalDeletionRequiresCheckout])).toContain(
-      `${checkoutIndependentDeletionScenario.id}: global orphan deletion requires an orphaned target`,
-    );
-
-    const branchHistoryEvidenceScenario = findScenario("identity.branch-commit-preserves-context");
-    const branchHistoryLosesEvidence = {
-      ...branchHistoryEvidenceScenario,
-      given: branchHistoryEvidenceScenario.given.filter(
-        (fact) => fact.kind !== "branch" && fact.kind !== "identity-transition",
-      ),
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([branchHistoryLosesEvidence])).toContain(
-      `${branchHistoryEvidenceScenario.id}: branch history preservation requires matching branch evidence`,
-    );
-
-    const configuredCredentialUpdateScenario = findScenario(
-      "credentials.explicit-change-applies-after-stop",
-    );
-    const configuredCredentialUpdateLosesState = {
-      ...configuredCredentialUpdateScenario,
-      given: configuredCredentialUpdateScenario.given.filter(
-        (fact) => fact.kind !== "credential-state",
-      ),
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([configuredCredentialUpdateLosesState])).toContain(
-      `${configuredCredentialUpdateScenario.id}: credential change requires configured old and new values`,
-    );
-
-    const branchHistoryLosesIntent = {
-      ...branchHistoryEvidenceScenario,
-      given: branchHistoryEvidenceScenario.given.filter((fact) => fact.kind !== "branch-history"),
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([branchHistoryLosesIntent])).toContain(
-      `${branchHistoryEvidenceScenario.id}: branch history preservation requires matching branch evidence`,
-    );
-
-    const freshAutomaticPortsScenario = findScenario(
-      "ports.new-target-allocates-and-persists-omitted-ports",
-    );
-    const freshAutomaticPortsHidePersistence = {
-      ...freshAutomaticPortsScenario,
-      expected: {
-        ...freshAutomaticPortsScenario.expected,
-        details: {
-          ...freshAutomaticPortsScenario.expected.details,
-          host_wide: false,
-          sticky: false,
-        },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([freshAutomaticPortsHidePersistence])).toContain(
-      `${freshAutomaticPortsScenario.id}: fresh automatic ports must be host-wide sticky assignments`,
-    );
-
-    const absentLegacyScenario = findScenario("bootstrap.absent-legacy-starts-fresh");
-    const absentLegacyReportsMutation = {
-      ...absentLegacyScenario,
-      expected: {
-        ...absentLegacyScenario.expected,
-        details: { ...absentLegacyScenario.expected.details, legacy_state_mutated: true },
-        output: {
-          ...absentLegacyScenario.expected.output,
-          json: {
-            ...absentLegacyScenario.expected.output.json,
-            legacy_state_mutated: true,
-          },
-        },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([absentLegacyReportsMutation])).toContain(
-      `${absentLegacyScenario.id}: absent legacy bootstrap must not report mutation`,
-    );
-
-    const stableDefaultCredentialsScenario = findScenario(
-      "credentials.omitted-values-use-stable-defaults",
-    );
-    const stableDefaultCredentialsLoseEvidence = {
-      ...stableDefaultCredentialsScenario,
-      given: stableDefaultCredentialsScenario.given.filter(
-        (fact) => fact.kind !== "credential-state",
-      ),
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([stableDefaultCredentialsLoseEvidence])).toContain(
-      `${stableDefaultCredentialsScenario.id}: omitted credentials must reuse stable local defaults`,
-    );
-
-    const dockerPrecedenceScenario = findScenario("runtime.auto-prefers-docker");
-    const dockerPrecedenceLosesNativeCandidate = {
-      ...dockerPrecedenceScenario,
-      given: dockerPrecedenceScenario.given.filter(
-        (fact) => fact.kind !== "runtime-availability" || fact.runtime !== "native",
-      ),
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([dockerPrecedenceLosesNativeCandidate])).toContain(
-      `${dockerPrecedenceScenario.id}: fresh automatic runtime selection requires Docker and native availability evidence`,
-    );
-
-    const resolveStackScenario = findScenario(
-      "identity.same-checkout-branch-and-name-reuses-stack",
-    );
-    if (resolveStackScenario.when.interface !== "managed-api") {
-      throw new Error("same-checkout reuse must use the managed API");
-    }
-    const resolveStackUsesMalformedStateRoot = {
-      ...resolveStackScenario,
-      when: {
-        ...resolveStackScenario.when,
-        input: { ...resolveStackScenario.when.input, stateRoot: 42 },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([resolveStackUsesMalformedStateRoot])).toContain(
-      `${resolveStackScenario.id}: managed action must use a declared public method`,
-    );
-
-    const runningPortDriftScenario = findScenario(
-      "ports.config-change-on-running-stack-reports-drift",
-    );
-    const runningPortDriftHuman = runningPortDriftScenario.expected.output.human;
-    if (runningPortDriftHuman === undefined) {
-      throw new Error("running port drift human output is required");
-    }
-    const runningPortDriftReportsFalse = {
-      ...runningPortDriftScenario,
-      expected: {
-        ...runningPortDriftScenario.expected,
-        output: {
-          ...runningPortDriftScenario.expected.output,
-          human: {
-            ...runningPortDriftHuman,
-            fields: { ...runningPortDriftHuman.fields, drift: "false" },
-          },
-          json: { ...runningPortDriftScenario.expected.output.json, drift: false },
-        },
-      },
-    } satisfies ManagedStackContractScenario;
-    expect(validateManagedStackContractFixtures([runningPortDriftReportsFalse])).toContain(
-      `${runningPortDriftScenario.id}: exact port change must bind previous assignment and requested value`,
-    );
   });
 
   it("covers the approved identity journeys through public commands and APIs", () => {
@@ -4063,6 +349,7 @@ describe("managed stack acceptance contract", () => {
         "credentials.configured-values-are-authoritative",
         "credentials.explicit-change-applies-after-stop",
         "credentials.omitted-values-use-stable-defaults",
+        "credentials.plaintext-secrets-stay-out-of-global-state",
         "credentials.running-change-reports-drift",
         "credentials.unchanged-values-survive-restart",
       ].sort(),
