+++ packages/stack/src/managed-stack-contract.ts
@@ -192,6 +192,7 @@ export type ManagedStackContractFact =
       readonly source: "configured" | "legacy" | "local-default" | "persisted";
       readonly valuesId: string;
       readonly previousValuesId?: string;
+      readonly plaintextPresentInGlobalState?: boolean;
     }
   | {
       readonly kind: "identity-marker";
@@ -224,11 +225,6 @@ export type ManagedStackContractFact =
       readonly repositoryId?: string;
       readonly runtime: "bun" | "node";
     };
-
-type ManagedStackContractPortAssignmentFact = Extract<
-  ManagedStackContractFact,
-  { readonly kind: "port-assignment" }
->;
 
 export interface ManagedStackContractOutput {
   readonly human?: {
@@ -284,11 +280,6 @@ type ManagedStackContractWrite =
       readonly operation: "delete" | "start" | "update";
       readonly id: string;
     };
-
-type ManagedStackContractTemporaryRootWrite = Extract<
-  ManagedStackContractWrite,
-  { readonly target: "temporary-root" }
->;
 
 export interface ManagedStackContractEffects {
   readonly writes: ReadonlyArray<ManagedStackContractWrite>;
@@ -372,3917 +363,6 @@ const defineManagedStackContractFixtures = <
 >(
   fixtures: Fixtures,
 ): Fixtures => fixtures;
-
-const isManagedStackContractRecord = (
-  value: ManagedStackContractJson | undefined,
-): value is Readonly<Record<string, ManagedStackContractJson>> =>
-  typeof value === "object" && value !== null && !Array.isArray(value);
-
-const managedStackContractJsonEquals = (
-  left: ManagedStackContractJson,
-  right: ManagedStackContractJson,
-): boolean => {
-  if (Object.is(left, right)) {
-    return true;
-  }
-  if (Array.isArray(left) && Array.isArray(right)) {
-    return (
-      left.length === right.length &&
-      left.every((value, index) => managedStackContractJsonEquals(value, right[index] ?? null))
-    );
-  }
-  if (isManagedStackContractRecord(left) && isManagedStackContractRecord(right)) {
-    const leftKeys = Object.keys(left).sort();
-    const rightKeys = Object.keys(right).sort();
-    return (
-      leftKeys.length === rightKeys.length &&
-      leftKeys.every(
-        (key, index) =>
-          key === rightKeys[index] &&
-          right[key] !== undefined &&
-          managedStackContractJsonEquals(left[key] ?? null, right[key]),
-      )
-    );
-  }
-  return false;
-};
-
-const managedStackContractDecision = (
-  scenario: ManagedStackContractScenario,
-): Readonly<Record<string, ManagedStackContractJson>> =>
-  scenario.expected.selection === undefined
-    ? { outcome: scenario.expected.outcome }
-    : {
-        outcome: scenario.expected.outcome,
-        projectId: scenario.expected.selection.projectId,
-        checkoutId: scenario.expected.selection.checkoutId,
-        contextId: scenario.expected.selection.contextId,
-        stackId: scenario.expected.selection.stackId,
-        stackName: scenario.expected.selection.stackName,
-      };
-
-const managedStackContractStringSetEquals = (
-  left: ReadonlyArray<string>,
-  right: ReadonlyArray<string>,
-): boolean => {
-  const leftSet = new Set(left);
-  const rightSet = new Set(right);
-  return (
-    leftSet.size === left.length &&
-    rightSet.size === right.length &&
-    leftSet.size === rightSet.size &&
-    [...leftSet].every((value) => rightSet.has(value))
-  );
-};
-
-const managedStackNamePattern = /^(?=.{1,63}$)[a-z0-9](?:[a-z0-9-]*[a-z0-9])?$/;
-
-const requiredBranchHistoryByScenarioId: Readonly<
-  Partial<Record<string, "commit" | "rebase" | "reset">>
-> = {
-  "identity.branch-commit-preserves-context": "commit",
-  "identity.branch-rebase-preserves-context": "rebase",
-  "identity.branch-reset-preserves-context": "reset",
-};
-
-const isManagedStartAction = (action: ManagedStackContractAction): boolean =>
-  (action.interface === "cli" && action.argv[0] === "start") ||
-  (action.interface === "managed-api" &&
-    (action.method === "startStack" ||
-      action.method === "startConcurrently" ||
-      action.input.operation === "start"));
-
-export const validateManagedStackContractFixtures = (
-  fixtures: ReadonlyArray<ManagedStackContractScenario>,
-): ReadonlyArray<string> => {
-  const errors: Array<string> = [];
-  const ids = new Set<string>();
-  const nativeServices = managedNativeServiceMatrix.services.map(([service]) => service);
-  const nativeServiceSet = new Set<ServiceName>(nativeServices);
-  const fixturesById = new Map(fixtures.map((scenario) => [scenario.id, scenario]));
-
-  for (const scenario of fixtures) {
-    if (ids.has(scenario.id)) {
-      errors.push(`${scenario.id}: duplicate scenario ID`);
-    }
-    ids.add(scenario.id);
-
-    if (!scenario.id.startsWith(`${scenario.area}.`)) {
-      errors.push(`${scenario.id}: ID must start with ${scenario.area}.`);
-    }
-    if (scenario.title.trim().length === 0) {
-      errors.push(`${scenario.id}: title is required`);
-    }
-    if (scenario.given.length === 0) {
-      errors.push(`${scenario.id}: at least one given fact is required`);
-    }
-
-    if (scenario.when.interface === "cli" || scenario.when.interface === "git") {
-      if (scenario.when.argv.length === 0) {
-        errors.push(`${scenario.id}: argv must contain a public command`);
-      }
-      if (scenario.when.cwd.trim().length === 0) {
-        errors.push(`${scenario.id}: cwd is required for command scenarios`);
-      }
-    } else {
-      if (scenario.when.method.trim().length === 0) {
-        errors.push(`${scenario.id}: public API method is required`);
-      }
-    }
-
-    if (scenario.when.interface === "managed-api") {
-      const managedMethods = new Set([
-        "createManagedStackService",
-        "preflightNative",
-        "resolvePortIntents",
-        "resolveStack",
-        "resolveStackNames",
-        "runPortableContract",
-        "runRepositoryContract",
-        "startConcurrently",
-        "startStack",
-      ]);
-      const { input, method } = scenario.when;
-      const allowedInputKeys: Readonly<Record<string, ReadonlySet<string>>> = {
-        createManagedStackService: new Set(["repository", "stateRoot"]),
-        preflightNative: new Set(["platform"]),
-        resolvePortIntents: new Set(["config", "decodedDefaults", "effectiveConfig"]),
-        resolveStack: new Set(["cwd", "operation", "stackName", "stateRoot"]),
-        resolveStackNames: new Set(["cwd", "stackNames"]),
-        runPortableContract: new Set(["runtimes", "scenarioId"]),
-        runRepositoryContract: new Set(["adapters", "scenarioId"]),
-        startConcurrently: new Set(["contenders", "cwd", "stackName"]),
-        startStack: new Set(["auth", "injectCopyFailure", "portIntents", "runtime", "stackId"]),
-      };
-      const inputMatchesMethod =
-        (method === "createManagedStackService" &&
-          typeof input.repository === "string" &&
-          typeof input.stateRoot === "string") ||
-        (method === "preflightNative" && typeof input.platform === "string") ||
-        (method === "resolvePortIntents" &&
-          (isManagedStackContractRecord(input.config) ||
-            isManagedStackContractRecord(input.effectiveConfig))) ||
-        (method === "resolveStack" &&
-          typeof input.cwd === "string" &&
-          typeof input.stackName === "string" &&
-          (input.stateRoot === undefined || typeof input.stateRoot === "string")) ||
-        (method === "resolveStackNames" &&
-          typeof input.cwd === "string" &&
-          Array.isArray(input.stackNames)) ||
-        (method === "runPortableContract" &&
-          Array.isArray(input.runtimes) &&
-          typeof input.scenarioId === "string") ||
-        (method === "runRepositoryContract" &&
-          Array.isArray(input.adapters) &&
-          typeof input.scenarioId === "string") ||
-        (method === "startConcurrently" &&
-          typeof input.cwd === "string" &&
-          typeof input.stackName === "string" &&
-          typeof input.contenders === "number") ||
-        (method === "startStack" &&
-          typeof input.stackId === "string" &&
-          (input.auth === undefined || typeof input.auth === "string") &&
-          (input.injectCopyFailure === undefined || typeof input.injectCopyFailure === "boolean") &&
-          (input.portIntents === undefined || isManagedStackContractRecord(input.portIntents)) &&
-          (input.runtime === undefined ||
-            input.runtime === "auto" ||
-            input.runtime === "docker" ||
-            input.runtime === "native"));
-      const inputUsesOnlyDeclaredKeys = Object.keys(input).every((key) =>
-        allowedInputKeys[method]?.has(key),
-      );
-      if (!managedMethods.has(method) || !inputMatchesMethod || !inputUsesOnlyDeclaredKeys) {
-        errors.push(`${scenario.id}: managed action must use a declared public method`);
-      }
-    } else if (
-      scenario.when.interface === "stack-api" &&
-      scenario.when.method !== "createStack" &&
-      scenario.when.method !== "dispose"
-    ) {
-      errors.push(`${scenario.id}: direct stack action must use createStack or dispose`);
-    }
-
-    if (scenario.when.interface === "cli") {
-      const [command, subcommand] = scenario.when.argv;
-      const valueFlagsByCommand: Readonly<Record<string, ReadonlySet<string>>> = {
-        start: new Set(["--runtime", "--stack"]),
-        status: new Set(["--output"]),
-        stop: new Set(["--stack", "--stack-id"]),
-      };
-      const switchFlagsByCommand: Readonly<Record<string, ReadonlySet<string>>> = {
-        start: new Set(["--experimental"]),
-        status: new Set(["--experimental"]),
-        stop: new Set(["--all", "--experimental", "--no-backup"]),
-      };
-      let validShape = command === "start" || command === "status" || command === "stop";
-      if (command === "stack") {
-        validShape = subcommand === "prune";
-      }
-      let index = command === "stack" ? 2 : 1;
-      while (validShape && index < scenario.when.argv.length) {
-        const argument = scenario.when.argv[index];
-        if (argument === undefined) {
-          validShape = false;
-          break;
-        }
-        if (command === "stack" && argument === "--experimental") {
-          index += 1;
-          continue;
-        }
-        if (command !== undefined && switchFlagsByCommand[command]?.has(argument)) {
-          index += 1;
-          continue;
-        }
-        if (command !== undefined && valueFlagsByCommand[command]?.has(argument)) {
-          const value = scenario.when.argv[index + 1];
-          validShape = value !== undefined;
-          if (
-            (argument === "--output" && value !== "json") ||
-            (argument === "--runtime" && value !== "docker" && value !== "native")
-          ) {
-            validShape = false;
-          }
-          index += 2;
-          continue;
-        }
-        validShape = false;
-      }
-      if (!scenario.when.argv.includes("--experimental") || !validShape) {
-        errors.push(`${scenario.id}: CLI action must use a declared experimental command shape`);
-      }
-    }
-
-    if (scenario.when.interface === "git") {
-      const [command, option, branchName] = scenario.when.argv;
-      const validShape =
-        typeof branchName === "string" &&
-        ((command === "switch" && option === "-c") ||
-          (command === "branch" && (option === "-D" || option === "-d")));
-      if (!validShape) {
-        errors.push(`${scenario.id}: Git action must use a declared branch command shape`);
-      } else if (
-        command === "switch" &&
-        scenario.expected.output.human?.summary !== `Switched to a new branch '${branchName}'`
-      ) {
-        errors.push(`${scenario.id}: Git switch output must match its requested branch`);
-      }
-      if (
-        scenario.expected.writes.length === 0 &&
-        scenario.expected.runtimeEffects.length === 0 &&
-        (scenario.expected.outcome !== "no-op" ||
-          scenario.expected.details?.managed_command_ran !== false)
-      ) {
-        errors.push(`${scenario.id}: standalone Git action must remain outside managed lifecycle`);
-      }
-    }
-
-    const isStatusOperation =
-      (scenario.when.interface === "cli" && scenario.when.argv[0] === "status") ||
-      ((scenario.when.interface === "managed-api" || scenario.when.interface === "stack-api") &&
-        scenario.when.input.operation === "status");
-    if (
-      isStatusOperation &&
-      scenario.expected.outcome !== "error" &&
-      scenario.expected.outcome !== "report"
-    ) {
-      errors.push(`${scenario.id}: successful status operations must report`);
-    }
-    if (
-      isStatusOperation &&
-      (scenario.expected.writes.length > 0 || scenario.expected.runtimeEffects.length > 0)
-    ) {
-      errors.push(`${scenario.id}: status operations must not mutate state`);
-    }
-
-    const { output } = scenario.expected;
-    if (output.human === undefined && output.json === undefined && output.api === undefined) {
-      errors.push(`${scenario.id}: at least one observable output is required`);
-    }
-    if (scenario.expected.error?.code === "MUTUALLY_EXCLUSIVE_STACK_SELECTORS") {
-      const stopArgv =
-        scenario.when.interface === "cli" && scenario.when.argv[0] === "stop"
-          ? scenario.when.argv
-          : [];
-      const selectorModes = ["--stack", "--stack-id", "--all"].filter((selector) =>
-        stopArgv.includes(selector),
-      );
-      const jsonSelectors = output.json?.selectors;
-      const projectedJsonSelectors =
-        Array.isArray(jsonSelectors) &&
-        jsonSelectors.every((selector): selector is string => typeof selector === "string")
-          ? jsonSelectors
-          : undefined;
-      const projectedHumanSelectors = output.human?.fields.selectors
-        ?.split(",")
-        .map((selector) => selector.trim())
-        .filter((selector) => selector.length > 0);
-      if (
-        selectorModes.length < 2 ||
-        projectedJsonSelectors === undefined ||
-        !managedStackContractStringSetEquals(selectorModes, projectedJsonSelectors) ||
-        projectedHumanSelectors === undefined ||
-        !managedStackContractStringSetEquals(selectorModes, projectedHumanSelectors)
-      ) {
-        errors.push(
-          `${scenario.id}: selector conflict must bind at least two requested selection modes`,
-        );
-      }
-    }
-
-    if (scenario.when.interface === "stack-api" && scenario.when.method === "createStack") {
-      const directOptions = scenario.given.filter((fact) => fact.kind === "direct-stack-options");
-      const directInput = scenario.when.input;
-      const expectedStackRoot = typeof directInput.stackRoot === "string" ? "explicit" : "omitted";
-      const expectedRuntimeRoot =
-        typeof directInput.runtimeRoot === "string" ? "explicit" : "omitted";
-      const omittedRoots: Array<"stack" | "runtime"> = [];
-      if (expectedStackRoot === "omitted") {
-        omittedRoots.push("stack");
-      }
-      if (expectedRuntimeRoot === "omitted") {
-        omittedRoots.push("runtime");
-      }
-      const temporaryRootWrites = scenario.expected.writes.filter(
-        (write): write is ManagedStackContractTemporaryRootWrite =>
-          write.target === "temporary-root" && write.operation === "create",
-      );
-      const detailRoots = scenario.expected.details?.temporary_roots;
-      const projectedRoots = output.api?.temporaryRoots;
-      const validDetailRoots =
-        Array.isArray(detailRoots) &&
-        detailRoots.every(
-          (root): root is "stack" | "runtime" => root === "stack" || root === "runtime",
-        )
-          ? detailRoots
-          : undefined;
-      const validProjectedRoots =
-        Array.isArray(projectedRoots) &&
-        projectedRoots.every(
-          (root): root is "stack" | "runtime" => root === "stack" || root === "runtime",
-        )
-          ? projectedRoots
-          : undefined;
-      if (
-        directOptions.length !== 1 ||
-        directOptions[0]?.stackRoot !== expectedStackRoot ||
-        directOptions[0]?.runtimeRoot !== expectedRuntimeRoot ||
-        temporaryRootWrites.length !== omittedRoots.length ||
-        !managedStackContractStringSetEquals(
-          temporaryRootWrites.map((write) => write.root),
-          omittedRoots,
-        ) ||
-        validDetailRoots === undefined ||
-        validDetailRoots.length !== omittedRoots.length ||
-        !managedStackContractStringSetEquals(validDetailRoots, omittedRoots) ||
-        validProjectedRoots === undefined ||
-        validProjectedRoots.length !== omittedRoots.length ||
-        !managedStackContractStringSetEquals(validProjectedRoots, omittedRoots)
-      ) {
-        errors.push(
-          `${scenario.id}: direct stack root inputs must agree with temporary-state behavior`,
-        );
-      }
-      if (
-        scenario.expected.details?.git_inspected !== false ||
-        scenario.expected.details.identity_marker_created !== false ||
-        scenario.expected.details.global_registry_mutated !== false ||
-        scenario.expected.writes.some((write) => write.target !== "temporary-root") ||
-        scenario.expected.runtimeEffects.length > 0
-      ) {
-        errors.push(`${scenario.id}: direct createStack must remain isolated from managed state`);
-      }
-    }
-    if (scenario.when.interface === "stack-api" && scenario.when.method === "dispose") {
-      const handle = scenario.when.input.handle;
-      const directState = scenario.given.find(
-        (fact) => fact.kind === "direct-stack-state" && fact.handle === handle,
-      );
-      const deletedTemporaryRoots = scenario.expected.writes.filter(
-        (write): write is ManagedStackContractTemporaryRootWrite =>
-          write.target === "temporary-root" && write.operation === "delete",
-      );
-      const removedDetailRoots = scenario.expected.details?.removed_temporary_roots;
-      const removedProjectedRoots = output.api?.removedTemporaryRoots;
-      const validRemovedDetailRoots =
-        Array.isArray(removedDetailRoots) &&
-        removedDetailRoots.every(
-          (root): root is "stack" | "runtime" => root === "stack" || root === "runtime",
-        )
-          ? removedDetailRoots
-          : undefined;
-      const validRemovedProjectedRoots =
-        Array.isArray(removedProjectedRoots) &&
-        removedProjectedRoots.every(
-          (root): root is "stack" | "runtime" => root === "stack" || root === "runtime",
-        )
-          ? removedProjectedRoots
-          : undefined;
-      const declaredTemporaryRoots =
-        directState?.kind === "direct-stack-state" ? directState.temporaryRoots : [];
-      if (
-        typeof handle !== "string" ||
-        directState?.kind !== "direct-stack-state" ||
-        directState.lifecycle !== "created" ||
-        new Set(declaredTemporaryRoots.map(({ root }) => root)).size !==
-          declaredTemporaryRoots.length ||
-        scenario.expected.outcome !== "delete" ||
-        deletedTemporaryRoots.length !== declaredTemporaryRoots.length ||
-        declaredTemporaryRoots.some(
-          (root) =>
-            !deletedTemporaryRoots.some(
-              (write) => write.root === root.root && write.id === root.stateId,
-            ),
-        ) ||
-        deletedTemporaryRoots.some(
-          (write) =>
-            !declaredTemporaryRoots.some(
-              (root) => root.root === write.root && root.stateId === write.id,
-            ),
-        ) ||
-        scenario.expected.writes.some((write) => write.target !== "temporary-root") ||
-        scenario.expected.runtimeEffects.length > 0 ||
-        scenario.expected.details?.temporary_roots_removed !== true ||
-        validRemovedDetailRoots === undefined ||
-        validRemovedDetailRoots.length !== declaredTemporaryRoots.length ||
-        !managedStackContractStringSetEquals(
-          validRemovedDetailRoots,
-          declaredTemporaryRoots.map(({ root }) => root),
-        ) ||
-        output.api?.handle !== handle ||
-        output.api.disposed !== true ||
-        output.api.temporaryRootsRemoved !== true ||
-        validRemovedProjectedRoots === undefined ||
-        validRemovedProjectedRoots.length !== declaredTemporaryRoots.length ||
-        !managedStackContractStringSetEquals(
-          validRemovedProjectedRoots,
-          declaredTemporaryRoots.map(({ root }) => root),
-        )
-      ) {
-        errors.push(`${scenario.id}: direct stack disposal must remove omitted temporary roots`);
-      }
-    }
-    for (const write of scenario.expected.writes) {
-      if (write.id.trim().length === 0) {
-        errors.push(`${scenario.id}: ${write.target} write requires a target ID`);
-      }
-    }
-    for (const effect of scenario.expected.runtimeEffects) {
-      if (effect.stackId.trim().length === 0) {
-        errors.push(`${scenario.id}: ${effect.operation} runtime effect requires a stack ID`);
-      }
-    }
-
-    const cliStackIdIndex =
-      scenario.when.interface === "cli" ? scenario.when.argv.indexOf("--stack-id") : -1;
-    const explicitActionStackId =
-      scenario.when.interface === "cli" && cliStackIdIndex >= 0
-        ? scenario.when.argv[cliStackIdIndex + 1]
-        : (scenario.when.interface === "managed-api" || scenario.when.interface === "stack-api") &&
-            typeof scenario.when.input.stackId === "string"
-          ? scenario.when.input.stackId
-          : undefined;
-    if (explicitActionStackId !== undefined) {
-      const expectedStackIds = new Set<string>();
-      if (scenario.expected.selection !== undefined) {
-        expectedStackIds.add(scenario.expected.selection.stackId);
-      }
-      for (const write of scenario.expected.writes) {
-        if (
-          write.target === "managed-state" ||
-          write.target === "registry" ||
-          write.target === "runtime-state"
-        ) {
-          expectedStackIds.add(write.id);
-        }
-      }
-      for (const effect of scenario.expected.runtimeEffects) {
-        expectedStackIds.add(effect.stackId);
-      }
-      for (const projectedStackId of [
-        output.api?.stackId,
-        output.json?.stack_id,
-        output.human?.fields.stackId,
-      ]) {
-        if (typeof projectedStackId === "string") {
-          expectedStackIds.add(projectedStackId);
-        }
-      }
-      for (const expectedStackId of expectedStackIds) {
-        if (expectedStackId !== explicitActionStackId) {
-          errors.push(
-            `${scenario.id}: explicit action target ${explicitActionStackId} disagrees with expected stack ${expectedStackId}`,
-          );
-        }
-      }
-    }
-
-    const cliStackNameIndex =
-      scenario.when.interface === "cli" ? scenario.when.argv.indexOf("--stack") : -1;
-    const explicitActionStackName =
-      scenario.when.interface === "cli" && cliStackNameIndex >= 0
-        ? scenario.when.argv[cliStackNameIndex + 1]
-        : (scenario.when.interface === "managed-api" || scenario.when.interface === "stack-api") &&
-            typeof scenario.when.input.stackName === "string"
-          ? scenario.when.input.stackName
-          : undefined;
-    if (
-      explicitActionStackName !== undefined &&
-      scenario.expected.selection !== undefined &&
-      explicitActionStackName !== scenario.expected.selection.stackName
-    ) {
-      errors.push(
-        `${scenario.id}: explicit stack name ${explicitActionStackName} disagrees with selected stack ${scenario.expected.selection.stackName}`,
-      );
-    }
-    if (
-      scenario.when.interface === "cli" &&
-      scenario.when.argv[0] === "start" &&
-      cliStackNameIndex < 0 &&
-      scenario.expected.selection !== undefined &&
-      scenario.expected.selection.stackName !== "default"
-    ) {
-      errors.push(`${scenario.id}: unnamed CLI start must select the default stack`);
-    }
-    if (scenario.expected.error?.code === "INVALID_STACK_NAME") {
-      const declaredNames = scenario.given.find((fact) => fact.kind === "stack-names");
-      if (
-        explicitActionStackName === undefined ||
-        managedStackNamePattern.test(explicitActionStackName) ||
-        declaredNames?.kind !== "stack-names" ||
-        declaredNames.names.length !== 1 ||
-        declaredNames.names[0] !== explicitActionStackName ||
-        output.human?.fields.stack !== explicitActionStackName ||
-        output.json?.stack_name !== explicitActionStackName
-      ) {
-        errors.push(
-          `${scenario.id}: invalid stack name error must bind a requested name outside the supported grammar`,
-        );
-      }
-    }
-
-    const cliRuntimeIndex =
-      scenario.when.interface === "cli" ? scenario.when.argv.indexOf("--runtime") : -1;
-    const explicitRuntime =
-      scenario.when.interface === "cli" && cliRuntimeIndex >= 0
-        ? scenario.when.argv[cliRuntimeIndex + 1]
-        : scenario.when.interface === "managed-api" &&
-            typeof scenario.when.input.runtime === "string"
-          ? scenario.when.input.runtime
-          : undefined;
-    if (explicitRuntime !== undefined) {
-      const actionRequestSource = scenario.when.interface === "cli" ? "cli" : "managed-api";
-      const runtimeRequest = scenario.given.find(
-        (fact) =>
-          fact.kind === "runtime-request" &&
-          fact.runtime === explicitRuntime &&
-          (explicitRuntime === "auto" || fact.source === actionRequestSource),
-      );
-      if (runtimeRequest?.kind !== "runtime-request") {
-        errors.push(
-          `${scenario.id}: explicit runtime ${explicitRuntime} must match its ${actionRequestSource} request fact`,
-        );
-      }
-      if (explicitRuntime !== "auto") {
-        for (const projectedRuntime of [
-          scenario.expected.details?.resolved_runtime,
-          output.api?.runtime,
-          output.json?.runtime,
-          output.human?.fields.runtime,
-        ]) {
-          if (projectedRuntime !== undefined && projectedRuntime !== explicitRuntime) {
-            errors.push(
-              `${scenario.id}: resolved runtime must match explicit request ${explicitRuntime}`,
-            );
-          }
-        }
-        if (isManagedStartAction(scenario.when) && scenario.expected.outcome !== "error") {
-          const availability = scenario.given.find(
-            (fact) => fact.kind === "runtime-availability" && fact.runtime === explicitRuntime,
-          );
-          if (availability?.kind !== "runtime-availability" || availability.available !== true) {
-            errors.push(
-              `${scenario.id}: successful explicit runtime requires matching availability`,
-            );
-          }
-        }
-      }
-    }
-
-    const runtimeRequests = scenario.given.filter((fact) => fact.kind === "runtime-request");
-    const effectiveRuntimeRequest =
-      runtimeRequests.find((fact) => fact.source === "cli" || fact.source === "managed-api") ??
-      runtimeRequests.find((fact) => fact.source === "config") ??
-      runtimeRequests.find((fact) => fact.source === "default");
-    if (
-      isManagedStartAction(scenario.when) &&
-      scenario.expected.outcome !== "error" &&
-      effectiveRuntimeRequest?.kind === "runtime-request" &&
-      effectiveRuntimeRequest.runtime !== "auto"
-    ) {
-      const requestedRuntime = effectiveRuntimeRequest.runtime;
-      const availability = scenario.given.find(
-        (fact) => fact.kind === "runtime-availability" && fact.runtime === requestedRuntime,
-      );
-      const projectedRuntimes = [
-        scenario.expected.details?.resolved_runtime,
-        output.api?.runtime,
-        output.json?.runtime,
-        output.human?.fields.runtime,
-      ].filter((runtime) => runtime !== undefined);
-      const projectedSources = [
-        scenario.expected.details?.source,
-        output.api?.runtimeSource,
-        output.json?.runtime_source,
-      ].filter((source) => source !== undefined);
-      if (
-        availability?.kind !== "runtime-availability" ||
-        availability.available !== true ||
-        projectedRuntimes.length === 0 ||
-        projectedRuntimes.some((runtime) => runtime !== requestedRuntime) ||
-        projectedSources.some((source) => source !== effectiveRuntimeRequest.source)
-      ) {
-        errors.push(
-          `${scenario.id}: effective runtime request must match availability and successful projections`,
-        );
-      }
-    }
-    if (
-      isManagedStartAction(scenario.when) &&
-      effectiveRuntimeRequest?.kind === "runtime-request" &&
-      effectiveRuntimeRequest.runtime === "auto"
-    ) {
-      const runtimeTargetStackId = scenario.expected.selection?.stackId ?? explicitActionStackId;
-      const persistedRuntime = scenario.given.find(
-        (fact) =>
-          fact.kind === "persisted-runtime" &&
-          runtimeTargetStackId !== undefined &&
-          fact.stackId === runtimeTargetStackId,
-      );
-      const dockerAvailability = scenario.given.find(
-        (fact) => fact.kind === "runtime-availability" && fact.runtime === "docker",
-      );
-      const nativeAvailability = scenario.given.find(
-        (fact) => fact.kind === "runtime-availability" && fact.runtime === "native",
-      );
-      const nativeQualification = scenario.given.find(
-        (fact) => fact.kind === "native-qualification",
-      );
-      const nativeQualified =
-        nativeQualification?.kind === "native-qualification" &&
-        nativeQualification.failedServices.length === 0 &&
-        nativeQualification.qualifiedServices.length === nativeServices.length;
-      const persistedAvailability =
-        persistedRuntime?.kind === "persisted-runtime"
-          ? scenario.given.find(
-              (fact) =>
-                fact.kind === "runtime-availability" && fact.runtime === persistedRuntime.runtime,
-            )
-          : undefined;
-      if (
-        persistedRuntime?.kind === "persisted-runtime" &&
-        persistedAvailability?.kind !== "runtime-availability"
-      ) {
-        errors.push(
-          `${scenario.id}: persisted automatic runtime requires matching availability evidence`,
-        );
-      }
-      if (
-        persistedRuntime === undefined &&
-        (dockerAvailability?.kind !== "runtime-availability" ||
-          nativeAvailability?.kind !== "runtime-availability")
-      ) {
-        errors.push(
-          `${scenario.id}: fresh automatic runtime selection requires Docker and native availability evidence`,
-        );
-      }
-      const resolvedRuntime =
-        persistedRuntime?.kind === "persisted-runtime" &&
-        persistedAvailability?.kind === "runtime-availability"
-          ? persistedAvailability.available
-            ? persistedRuntime.runtime
-            : undefined
-          : persistedRuntime?.kind === "persisted-runtime"
-            ? undefined
-            : dockerAvailability?.kind === "runtime-availability" && dockerAvailability.available
-              ? "docker"
-              : nativeAvailability?.kind === "runtime-availability" &&
-                  nativeAvailability.available &&
-                  nativeQualified
-                ? "native"
-                : undefined;
-      if (
-        persistedRuntime === undefined &&
-        resolvedRuntime !== undefined &&
-        scenario.expected.outcome !== "error" &&
-        scenario.expected.details?.persisted !== true
-      ) {
-        errors.push(`${scenario.id}: fresh automatic runtime selection must be persisted`);
-      }
-      if (
-        resolvedRuntime === "native" &&
-        nativeQualification?.kind === "native-qualification" &&
-        (scenario.expected.details?.qualified_service_count !==
-          nativeQualification.qualifiedServices.length ||
-          scenario.expected.details.mixed_runtime !== false ||
-          output.api?.qualifiedServiceCount !== nativeQualification.qualifiedServices.length)
-      ) {
-        errors.push(
-          `${scenario.id}: automatic native selection must bind the full qualified graph`,
-        );
-      }
-      if (
-        persistedRuntime?.kind === "persisted-runtime" &&
-        persistedAvailability?.kind === "runtime-availability" &&
-        persistedAvailability.available &&
-        (scenario.expected.details?.runtime !== persistedRuntime.runtime ||
-          scenario.expected.details.auto_re_evaluated !== false ||
-          output.json?.persisted !== true)
-      ) {
-        errors.push(
-          `${scenario.id}: automatic persisted runtime reuse must report persisted provenance`,
-        );
-      }
-      if (
-        persistedRuntime === undefined &&
-        resolvedRuntime === "native" &&
-        (dockerAvailability?.kind !== "runtime-availability" ||
-          dockerAvailability.available ||
-          typeof dockerAvailability.reason !== "string")
-      ) {
-        errors.push(
-          `${scenario.id}: automatic native fallback requires explicit Docker unavailability`,
-        );
-      }
-      if (
-        persistedRuntime?.kind === "persisted-runtime" &&
-        persistedAvailability?.kind === "runtime-availability" &&
-        !persistedAvailability.available
-      ) {
-        if (
-          scenario.expected.outcome !== "error" ||
-          scenario.expected.error?.code !== "PERSISTED_RUNTIME_UNAVAILABLE" ||
-          typeof persistedAvailability.reason !== "string" ||
-          output.json?.runtime !== persistedRuntime.runtime ||
-          output.json.reason !== persistedAvailability.reason ||
-          scenario.expected.runtimeEffects.some((effect) => effect.operation === "start")
-        ) {
-          errors.push(`${scenario.id}: unavailable persisted runtime must fail without switching`);
-        }
-      } else if (resolvedRuntime === undefined) {
-        const unavailableReasonsMatch =
-          dockerAvailability?.kind === "runtime-availability" &&
-          dockerAvailability.available === false &&
-          typeof dockerAvailability.reason === "string" &&
-          nativeAvailability?.kind === "runtime-availability" &&
-          nativeAvailability.available === false &&
-          typeof nativeAvailability.reason === "string" &&
-          output.human?.fields.docker === dockerAvailability.reason &&
-          output.human.fields.native === nativeAvailability.reason &&
-          output.json?.docker_reason === dockerAvailability.reason &&
-          output.json.native_reason === nativeAvailability.reason;
-        if (
-          scenario.expected.outcome !== "error" ||
-          scenario.expected.error?.code !== "NO_RUNTIME_AVAILABLE" ||
-          scenario.expected.runtimeEffects.some((effect) => effect.operation === "start") ||
-          !unavailableReasonsMatch
-        ) {
-          errors.push(
-            `${scenario.id}: automatic runtime failure must bind both unavailability reasons`,
-          );
-        }
-      } else {
-        const projectedRuntimes = [
-          scenario.expected.details?.resolved_runtime,
-          output.api?.runtime,
-          output.json?.runtime,
-          output.human?.fields.runtime,
-        ].filter((runtime) => runtime !== undefined);
-        if (
-          scenario.expected.outcome === "error" ||
-          projectedRuntimes.length === 0 ||
-          projectedRuntimes.some((runtime) => runtime !== resolvedRuntime)
-        ) {
-          errors.push(
-            `${scenario.id}: automatic runtime must resolve from persisted state or declared availability`,
-          );
-        }
-      }
-    }
-
-    if (scenario.when.interface === "managed-api" && typeof scenario.when.input.auth === "string") {
-      const authReference = scenario.when.input.auth;
-      const configuredCredentials = scenario.given.find(
-        (fact) => fact.kind === "credential-state" && fact.source === "configured",
-      );
-      if (
-        configuredCredentials?.kind !== "credential-state" ||
-        configuredCredentials.valuesId !== authReference ||
-        scenario.expected.details?.credential_values_id !== authReference ||
-        scenario.expected.details.global_credentials_reference !== authReference ||
-        output.api?.credentialsValuesId !== authReference
-      ) {
-        errors.push(
-          `${scenario.id}: configured credential input ${authReference} must match persisted references`,
-        );
-      }
-    }
-
-    if (scenario.when.interface === "managed-api" && scenario.when.method === "resolveStack") {
-      const stateRoot = scenario.when.input.stateRoot;
-      const managedOptions = scenario.given.filter((fact) => fact.kind === "managed-api-options");
-      const isolatedOptions = managedOptions.find((fact) => fact.stateRoot === "isolated");
-      if (typeof stateRoot === "string" || isolatedOptions !== undefined) {
-        if (
-          typeof stateRoot !== "string" ||
-          managedOptions.length !== 1 ||
-          isolatedOptions?.kind !== "managed-api-options" ||
-          isolatedOptions.stateRootPath !== stateRoot ||
-          scenario.expected.details?.state_root !== stateRoot ||
-          scenario.expected.details.default_system_state_mutated !== false
-        ) {
-          errors.push(
-            `${scenario.id}: isolated state root input must match its options and observed boundary`,
-          );
-        }
-      }
-    }
-
-    if (
-      scenario.when.interface === "managed-api" &&
-      scenario.when.method === "createManagedStackService"
-    ) {
-      const repositoryId = scenario.when.input.repository;
-      const stateRootPath = scenario.when.input.stateRoot;
-      const injectedOptions = scenario.given.find(
-        (fact) => fact.kind === "managed-api-options" && fact.repository === "injected",
-      );
-      if (
-        typeof repositoryId !== "string" ||
-        typeof stateRootPath !== "string" ||
-        injectedOptions?.kind !== "managed-api-options" ||
-        injectedOptions.repositoryId !== repositoryId ||
-        injectedOptions.stateRoot !== "isolated" ||
-        injectedOptions.stateRootPath !== stateRootPath ||
-        !scenario.expected.writes.some(
-          (write) =>
-            write.target === "ephemeral-state" &&
-            write.operation === "create" &&
-            write.id === repositoryId,
-        ) ||
-        output.api?.repository !== repositoryId ||
-        scenario.expected.details?.cli_required !== false ||
-        output.api.cliRequired !== false
-      ) {
-        errors.push(
-          `${scenario.id}: injected repository and state root must match the observed managed service`,
-        );
-      }
-    }
-
-    if (scenario.when.interface === "managed-api" && scenario.when.method === "resolveStackNames") {
-      const requestedNames = scenario.when.input.stackNames;
-      const declaredNames = scenario.given.find((fact) => fact.kind === "stack-names");
-      const activeContext = scenario.given.find(
-        (fact) => fact.kind === "branch" && fact.checkedOut,
-      );
-      const detailKeys = Object.keys(scenario.expected.details ?? {});
-      const apiKeys = Object.keys(output.api ?? {});
-      if (
-        !Array.isArray(requestedNames) ||
-        !requestedNames.every((name) => typeof name === "string") ||
-        !requestedNames.every(
-          (name) => typeof name === "string" && managedStackNamePattern.test(name),
-        ) ||
-        declaredNames?.kind !== "stack-names" ||
-        !managedStackContractStringSetEquals(requestedNames, declaredNames.names) ||
-        !managedStackContractStringSetEquals(requestedNames, detailKeys) ||
-        !managedStackContractStringSetEquals(requestedNames, apiKeys)
-      ) {
-        errors.push(
-          `${scenario.id}: requested stack names must match their fact and projected results`,
-        );
-      }
-      if (Array.isArray(requestedNames)) {
-        for (const name of requestedNames) {
-          if (typeof name !== "string") {
-            continue;
-          }
-          const detailStackId = scenario.expected.details?.[name];
-          const apiResult = output.api?.[name];
-          if (
-            typeof detailStackId !== "string" ||
-            !isManagedStackContractRecord(apiResult) ||
-            activeContext?.kind !== "branch" ||
-            apiResult.contextId !== activeContext.contextId ||
-            apiResult.stackId !== detailStackId
-          ) {
-            errors.push(
-              `${scenario.id}: resolved stack name ${name} must bind its context and stack ID`,
-            );
-          }
-        }
-      }
-    }
-
-    if (
-      scenario.when.interface === "managed-api" &&
-      scenario.when.method === "resolvePortIntents"
-    ) {
-      const configFacts = scenario.given.filter((fact) => fact.kind === "config-port");
-      const projectedKeys = Object.keys(output.api ?? {});
-      if (
-        !managedStackContractStringSetEquals(
-          configFacts.map(({ key }) => key),
-          projectedKeys,
-        )
-      ) {
-        errors.push(`${scenario.id}: resolved port keys must match their config facts`);
-      }
-      if (isManagedStackContractRecord(scenario.when.input.config)) {
-        const decodedDefaults = scenario.when.input.decodedDefaults;
-        if (
-          !isManagedStackContractRecord(decodedDefaults) ||
-          !managedStackContractStringSetEquals(
-            configFacts.map(({ key }) => key),
-            Object.keys(decodedDefaults),
-          )
-        ) {
-          errors.push(`${scenario.id}: decoded default keys must cover resolved port facts`);
-        }
-      }
-      for (const fact of configFacts) {
-        const projection = output.api?.[fact.key];
-        const effectiveConfig = isManagedStackContractRecord(scenario.when.input.effectiveConfig)
-          ? scenario.when.input.effectiveConfig
-          : undefined;
-        const localConfig = isManagedStackContractRecord(scenario.when.input.config)
-          ? scenario.when.input.config
-          : undefined;
-        const actionValue = effectiveConfig?.[fact.key] ?? localConfig?.[fact.key];
-        if (
-          !isManagedStackContractRecord(projection) ||
-          projection.intent !== fact.intent ||
-          projection.source !== fact.source ||
-          (fact.intent === "exact" &&
-            (actionValue !== fact.value || projection.port !== fact.value)) ||
-          (fact.intent === "automatic" && localConfig?.[fact.key] !== undefined)
-        ) {
-          errors.push(`${scenario.id}: resolved port ${fact.key} must match its input and fact`);
-        }
-      }
-    }
-
-    if (
-      scenario.when.interface === "cli" &&
-      (scenario.when.argv[0] === "start" ||
-        scenario.when.argv[0] === "status" ||
-        scenario.when.argv[0] === "stop") &&
-      !scenario.when.argv.includes("--stack-id") &&
-      typeof output.json?.stack_id === "string" &&
-      scenario.expected.selection === undefined
-    ) {
-      errors.push(`${scenario.id}: contextual CLI stack result requires a selected target`);
-    }
-
-    if (
-      scenario.when.interface === "managed-api" &&
-      scenario.when.method === "runPortableContract"
-    ) {
-      const referencedId = scenario.when.input.scenarioId;
-      const referencedScenario =
-        typeof referencedId === "string" ? fixturesById.get(referencedId) : undefined;
-      if (referencedScenario === undefined) {
-        errors.push(`${scenario.id}: portable contract must reference a declared scenario`);
-      } else {
-        const referencedDecision = managedStackContractDecision(referencedScenario);
-        const runtimes = scenario.when.input.runtimes;
-        if (
-          !Array.isArray(runtimes) ||
-          runtimes.length === 0 ||
-          !runtimes.every((runtime) => typeof runtime === "string" && runtime.length > 0)
-        ) {
-          errors.push(`${scenario.id}: portable contract must declare its runtimes`);
-        } else {
-          if (new Set(runtimes).size !== runtimes.length) {
-            errors.push(`${scenario.id}: portable contract runtimes must be unique`);
-          }
-          const runtimeFacts = scenario.given.flatMap((fact) =>
-            fact.kind === "managed-api-options" ? [fact.runtime] : [],
-          );
-          const declaredRuntimeSet = new Set(runtimes);
-          const runtimeFactSet = new Set(runtimeFacts);
-          if (
-            declaredRuntimeSet.size !== runtimeFactSet.size ||
-            [...declaredRuntimeSet].some((runtime) => !runtimeFactSet.has(runtime))
-          ) {
-            errors.push(`${scenario.id}: portable runtimes must match declared runtime facts`);
-          }
-          const runtimeOptions = scenario.given.filter(
-            (fact) => fact.kind === "managed-api-options",
-          );
-          const firstRuntimeOptions = runtimeOptions[0];
-          if (
-            firstRuntimeOptions === undefined ||
-            runtimeOptions.length !== runtimes.length ||
-            runtimes.some(
-              (runtime) =>
-                runtimeOptions.filter((options) => options.runtime === runtime).length !== 1,
-            ) ||
-            runtimeOptions.some(
-              (options) =>
-                options.repository !== firstRuntimeOptions.repository ||
-                options.repositoryId !== firstRuntimeOptions.repositoryId ||
-                options.stateRoot !== firstRuntimeOptions.stateRoot ||
-                options.stateRootPath !== firstRuntimeOptions.stateRootPath,
-            )
-          ) {
-            errors.push(
-              `${scenario.id}: portable comparison must hold repository and state root constant`,
-            );
-          }
-
-          let firstRuntimeResult: Readonly<Record<string, ManagedStackContractJson>> | undefined;
-          let runtimeResultsEqual = true;
-          for (const runtime of runtimes) {
-            const runtimeResult = output.api?.[runtime];
-            if (
-              !isManagedStackContractRecord(runtimeResult) ||
-              runtimeResult.outcome !== referencedScenario.expected.outcome
-            ) {
-              runtimeResultsEqual = false;
-              errors.push(
-                `${scenario.id}: portable ${runtime} outcome must match ${referencedScenario.id}`,
-              );
-              continue;
-            }
-            if (
-              referencedScenario.expected.selection !== undefined &&
-              runtimeResult.stackId !== referencedScenario.expected.selection.stackId
-            ) {
-              errors.push(
-                `${scenario.id}: portable ${runtime} stackId must match ${referencedScenario.id}`,
-              );
-            }
-            if (!managedStackContractJsonEquals(runtimeResult, referencedDecision)) {
-              errors.push(
-                `${scenario.id}: portable ${runtime} decision must completely match ${referencedScenario.id}`,
-              );
-            }
-            if (firstRuntimeResult === undefined) {
-              firstRuntimeResult = runtimeResult;
-            } else if (!managedStackContractJsonEquals(firstRuntimeResult, runtimeResult)) {
-              runtimeResultsEqual = false;
-              errors.push(`${scenario.id}: portable runtime decisions must be identical`);
-            }
-          }
-          if (
-            scenario.expected.details?.results_equal !== runtimeResultsEqual ||
-            output.api?.equal !== runtimeResultsEqual
-          ) {
-            errors.push(`${scenario.id}: portable equality flags must match compared results`);
-          }
-        }
-      }
-    }
-
-    if (
-      scenario.when.interface === "managed-api" &&
-      scenario.when.method === "runRepositoryContract"
-    ) {
-      const referencedId = scenario.when.input.scenarioId;
-      const referencedScenario =
-        typeof referencedId === "string" ? fixturesById.get(referencedId) : undefined;
-      if (referencedScenario === undefined) {
-        errors.push(`${scenario.id}: repository contract must reference a declared scenario`);
-      } else {
-        const referencedDecision = managedStackContractDecision(referencedScenario);
-        const adapters = scenario.when.input.adapters;
-        if (
-          !Array.isArray(adapters) ||
-          adapters.length === 0 ||
-          !adapters.every((adapter) => typeof adapter === "string" && adapter.length > 0)
-        ) {
-          errors.push(`${scenario.id}: repository contract must declare its adapters`);
-        } else {
-          if (new Set(adapters).size !== adapters.length) {
-            errors.push(`${scenario.id}: repository contract adapters must be unique`);
-          }
-          const repositoryFacts = scenario.given.flatMap((fact) =>
-            fact.kind === "managed-api-options" ? [fact.repository] : [],
-          );
-          const declaredRepositorySet = new Set(adapters);
-          const repositoryFactSet = new Set(repositoryFacts);
-          if (
-            declaredRepositorySet.size !== repositoryFactSet.size ||
-            [...declaredRepositorySet].some((adapter) => !repositoryFactSet.has(adapter))
-          ) {
-            errors.push(`${scenario.id}: repository adapters must match declared repository facts`);
-          }
-          const repositoryOptions = scenario.given.filter(
-            (fact) => fact.kind === "managed-api-options",
-          );
-          const firstRepositoryOptions = repositoryOptions[0];
-          if (
-            firstRepositoryOptions === undefined ||
-            repositoryOptions.length !== adapters.length ||
-            adapters.some(
-              (adapter) =>
-                repositoryOptions.filter((options) => options.repository === adapter).length !== 1,
-            ) ||
-            repositoryOptions.some(
-              (options) =>
-                options.runtime !== firstRepositoryOptions.runtime ||
-                options.stateRoot !== firstRepositoryOptions.stateRoot ||
-                options.stateRootPath !== firstRepositoryOptions.stateRootPath,
-            )
-          ) {
-            errors.push(
-              `${scenario.id}: repository comparison must hold runtime and state root constant`,
-            );
-          }
-
-          let firstAdapterResult: Readonly<Record<string, ManagedStackContractJson>> | undefined;
-          let adapterResultsEqual = true;
-          for (const adapter of adapters) {
-            const adapterResult = output.api?.[adapter];
-            if (
-              !isManagedStackContractRecord(adapterResult) ||
-              adapterResult.outcome !== referencedScenario.expected.outcome
-            ) {
-              adapterResultsEqual = false;
-              errors.push(
-                `${scenario.id}: repository ${adapter} outcome must match ${referencedScenario.id}`,
-              );
-              continue;
-            }
-            if (
-              referencedScenario.expected.selection !== undefined &&
-              adapterResult.stackId !== referencedScenario.expected.selection.stackId
-            ) {
-              errors.push(
-                `${scenario.id}: repository ${adapter} stackId must match ${referencedScenario.id}`,
-              );
-            }
-            if (!managedStackContractJsonEquals(adapterResult, referencedDecision)) {
-              errors.push(
-                `${scenario.id}: repository ${adapter} decision must completely match ${referencedScenario.id}`,
-              );
-            }
-            if (firstAdapterResult === undefined) {
-              firstAdapterResult = adapterResult;
-            } else if (!managedStackContractJsonEquals(firstAdapterResult, adapterResult)) {
-              adapterResultsEqual = false;
-              errors.push(`${scenario.id}: repository adapter decisions must be identical`);
-            }
-          }
-          if (
-            scenario.expected.details?.decisions_equal !== adapterResultsEqual ||
-            output.api?.equal !== adapterResultsEqual
-          ) {
-            errors.push(`${scenario.id}: repository equality flags must match compared decisions`);
-          }
-        }
-      }
-    }
-
-    if (scenario.expected.outcome === "error") {
-      if (scenario.expected.error === undefined) {
-        errors.push(`${scenario.id}: error outcome requires structured error metadata`);
-      } else if (scenario.expected.error.recovery.length === 0) {
-        errors.push(`${scenario.id}: error outcome requires recovery guidance`);
-      }
-    } else if (scenario.expected.error !== undefined) {
-      errors.push(`${scenario.id}: non-error outcome cannot include error metadata`);
-    }
-
-    for (const diagnostic of [scenario.expected.error, scenario.expected.warning]) {
-      if (diagnostic !== undefined && !/^[A-Z][A-Z0-9]*(?:_[A-Z0-9]+)*$/.test(diagnostic.code)) {
-        errors.push(
-          `${scenario.id}: diagnostic code ${diagnostic.code} must use SCREAMING_SNAKE_CASE`,
-        );
-      }
-    }
-
-    const hasMutation =
-      scenario.expected.writes.length > 0 || scenario.expected.runtimeEffects.length > 0;
-    if (scenario.expected.outcome === "report" && hasMutation) {
-      errors.push(`${scenario.id}: report outcome must not mutate state`);
-    }
-    if (scenario.expected.outcome === "error" && hasMutation) {
-      const isBootstrapRollback = scenario.expected.error?.code === "LEGACY_BOOTSTRAP_FAILED";
-      const containsOnlyRollbackCleanup =
-        scenario.expected.writes.every(
-          (write) => write.target === "managed-state" && write.operation === "delete",
-        ) && scenario.expected.runtimeEffects.every((effect) => effect.operation === "delete");
-      if (!isBootstrapRollback || !containsOnlyRollbackCleanup) {
-        errors.push(`${scenario.id}: error outcome must not mutate state outside rollback cleanup`);
-      }
-    }
-
-    if (scenario.expected.warning !== undefined) {
-      if (scenario.expected.outcome === "error") {
-        errors.push(`${scenario.id}: error outcome cannot also include warning metadata`);
-      }
-      if (scenario.expected.warning.recovery.length === 0) {
-        errors.push(`${scenario.id}: warning metadata requires recovery guidance`);
-      }
-    }
-
-    const declaredIds = new Set<string>();
-    const projectIds = new Set<string>();
-    const checkoutIds = new Set<string>();
-    const contextIds = new Set<string>();
-    const existingCheckoutIdentityIds = new Set<string>();
-    for (const fact of scenario.given) {
-      switch (fact.kind) {
-        case "branch":
-          declaredIds.add(fact.contextId);
-          contextIds.add(fact.contextId);
-          break;
-        case "checkout":
-          declaredIds.add(fact.projectId);
-          declaredIds.add(fact.checkoutId);
-          projectIds.add(fact.projectId);
-          checkoutIds.add(fact.checkoutId);
-          existingCheckoutIdentityIds.add(fact.projectId);
-          existingCheckoutIdentityIds.add(fact.checkoutId);
-          break;
-        case "credential-state":
-          declaredIds.add(fact.valuesId);
-          if (fact.previousValuesId !== undefined) {
-            declaredIds.add(fact.previousValuesId);
-          }
-          break;
-        case "direct-stack-state":
-          for (const root of fact.temporaryRoots) {
-            declaredIds.add(root.stateId);
-          }
-          break;
-        case "identity-claim":
-          if (fact.status !== "absent") {
-            declaredIds.add(fact.id);
-          }
-          switch (fact.scope) {
-            case "checkout":
-              checkoutIds.add(fact.id);
-              break;
-            case "context":
-              contextIds.add(fact.id);
-              break;
-            case "project":
-              projectIds.add(fact.id);
-              break;
-          }
-          break;
-        case "identity-marker":
-          declaredIds.add(fact.markerId);
-          declaredIds.add(fact.projectId);
-          declaredIds.add(fact.checkoutId);
-          declaredIds.add(fact.contextId);
-          projectIds.add(fact.projectId);
-          checkoutIds.add(fact.checkoutId);
-          contextIds.add(fact.contextId);
-          existingCheckoutIdentityIds.add(fact.projectId);
-          existingCheckoutIdentityIds.add(fact.checkoutId);
-          break;
-        case "managed-record":
-        case "managed-target":
-        case "persisted-runtime":
-          declaredIds.add(fact.stackId);
-          break;
-        case "operation-result":
-          declaredIds.add(fact.stackId);
-          break;
-        case "occupied-port":
-          if (fact.ownerId !== undefined) {
-            declaredIds.add(fact.ownerId);
-          }
-          break;
-        case "port-assignment":
-          declaredIds.add(fact.stackId);
-          break;
-        case "stack":
-          declaredIds.add(fact.contextId);
-          declaredIds.add(fact.stackId);
-          contextIds.add(fact.contextId);
-          break;
-        default:
-          break;
-      }
-    }
-
-    const actionCwd =
-      scenario.when.interface === "cli" || scenario.when.interface === "git"
-        ? scenario.when.cwd
-        : typeof scenario.when.input.cwd === "string"
-          ? scenario.when.input.cwd
-          : undefined;
-    if (
-      scenario.when.interface === "managed-api" &&
-      scenario.when.input.cwd !== undefined &&
-      (actionCwd === undefined ||
-        !scenario.given.some(
-          (fact) =>
-            (fact.kind === "checkout" && fact.path === actionCwd) ||
-            (fact.kind === "workspace" &&
-              (fact.path === actionCwd || fact.canonicalPath === actionCwd)) ||
-            (fact.kind === "git-state" && fact.workspacePath === actionCwd),
-        ))
-    ) {
-      errors.push(`${scenario.id}: managed action cwd must match a declared workspace`);
-    }
-    if (scenario.when.interface === "cli") {
-      const declaredActionPaths = scenario.given.flatMap((fact) => {
-        if (fact.kind === "checkout" || fact.kind === "git-state") {
-          return [fact.kind === "checkout" ? fact.path : fact.workspacePath];
-        }
-        if (fact.kind === "workspace") {
-          return fact.canonicalPath === undefined ? [fact.path] : [fact.path, fact.canonicalPath];
-        }
-        return [];
-      });
-      if (declaredActionPaths.length > 0 && !declaredActionPaths.includes(scenario.when.cwd)) {
-        errors.push(`${scenario.id}: CLI action cwd must match a declared workspace`);
-      }
-    }
-    const workspaceHistory = scenario.given.find(
-      (fact) => fact.kind === "workspace-history" && fact.path === actionCwd,
-    );
-    if (workspaceHistory?.kind === "workspace-history") {
-      const currentWorkspace = scenario.given.find(
-        (fact) => fact.kind === "workspace" && fact.path === actionCwd,
-      );
-      const folderToGitTransition = scenario.given.find(
-        (fact) => fact.kind === "identity-transition" && fact.operation === "folder-to-git",
-      );
-      if (
-        workspaceHistory.previousMode !== "ordinary-folder" ||
-        currentWorkspace?.kind !== "workspace" ||
-        currentWorkspace.mode !== "git" ||
-        folderToGitTransition?.kind !== "identity-transition" ||
-        folderToGitTransition.from !== "ordinary-folder" ||
-        folderToGitTransition.to !== "git" ||
-        (scenario.expected.outcome !== "error" && output.json?.converted_to_git !== true)
-      ) {
-        errors.push(`${scenario.id}: folder-to-Git result must bind the workspace transition`);
-      }
-    }
-    const absentCheckoutClaim = scenario.given.find(
-      (fact) =>
-        fact.kind === "identity-claim" && fact.scope === "checkout" && fact.status === "absent",
-    );
-    if (
-      isStatusOperation &&
-      absentCheckoutClaim?.kind === "identity-claim" &&
-      scenario.expected.writes.length === 0 &&
-      (scenario.expected.details?.registered !== false ||
-        scenario.expected.details.identity_marker_created !== false ||
-        output.json?.registered !== false)
-    ) {
-      errors.push(`${scenario.id}: read-only unregistered status must not create identity state`);
-    }
-    if (
-      isStatusOperation &&
-      (scenario.expected.details?.registered === false || output.json?.registered === false) &&
-      absentCheckoutClaim?.kind !== "identity-claim"
-    ) {
-      errors.push(`${scenario.id}: unregistered status requires an absent checkout claim`);
-    }
-
-    const branchCopy = scenario.given.find(
-      (fact) => fact.kind === "identity-transition" && fact.operation === "branch-copy",
-    );
-    if (branchCopy?.kind === "identity-transition") {
-      const originalBranchExists = scenario.given.some(
-        (fact) => fact.kind === "branch" && fact.name === branchCopy.from,
-      );
-      const copiedBranch = scenario.given.find(
-        (fact) => fact.kind === "branch" && fact.name === branchCopy.to,
-      );
-      const originalClaim = scenario.given.find(
-        (fact) =>
-          fact.kind === "identity-claim" &&
-          fact.scope === "context" &&
-          fact.owner === branchCopy.from &&
-          fact.status === (branchCopy.originalExists ? "exact" : "absent"),
-      );
-      if (
-        typeof branchCopy.from !== "string" ||
-        typeof branchCopy.to !== "string" ||
-        branchCopy.originalExists !== originalBranchExists ||
-        copiedBranch?.kind !== "branch" ||
-        copiedBranch.checkedOut !== true
-      ) {
-        errors.push(
-          `${scenario.id}: copied branch transition must match live original and checked-out branch facts`,
-        );
-      }
-      if (scenario.expected.warning?.code === "COPIED_BRANCH_CONTEXT_CONFLICT") {
-        if (
-          typeof branchCopy.from !== "string" ||
-          typeof branchCopy.to !== "string" ||
-          originalClaim?.kind !== "identity-claim" ||
-          scenario.expected.warning.message !==
-            `${branchCopy.to} copied ${originalClaim.id} from ${branchCopy.from}` ||
-          output.json?.branch !== branchCopy.to ||
-          output.json.owner !== branchCopy.from ||
-          output.json.context_id !== originalClaim.id
-        ) {
-          errors.push(`${scenario.id}: copied branch warning must bind observed branch ownership`);
-        }
-      }
-      if (
-        branchCopy.originalExists === true &&
-        scenario.expected.outcome === "create" &&
-        (originalClaim?.kind !== "identity-claim" ||
-          scenario.expected.details?.original_context_id !== originalClaim.id ||
-          scenario.expected.details.original_owner !== branchCopy.from ||
-          output.json?.original_context_id !== originalClaim.id)
-      ) {
-        errors.push(`${scenario.id}: copied branch creation must bind its original context`);
-      }
-      if (
-        branchCopy.originalExists === false &&
-        scenario.expected.outcome === "reuse" &&
-        (originalClaim?.kind !== "identity-claim" || output.json?.rename_detected !== true)
-      ) {
-        errors.push(`${scenario.id}: missing original branch must be reported as a rename`);
-      }
-    }
-    if (
-      scenario.when.interface === "managed-api" &&
-      scenario.when.method === "resolveStack" &&
-      scenario.when.input.operation !== undefined &&
-      scenario.when.input.operation !== "start" &&
-      scenario.when.input.operation !== "status"
-    ) {
-      errors.push(`${scenario.id}: resolveStack operation must be start or status`);
-    }
-    const actionGitState = scenario.given.find(
-      (fact) => fact.kind === "git-state" && fact.workspacePath === actionCwd,
-    );
-    const actionWorkspace = scenario.given.find(
-      (fact) => fact.kind === "workspace" && fact.path === actionCwd,
-    );
-    if (actionWorkspace?.kind === "workspace" && actionWorkspace.mode === "bare-worktree") {
-      if (
-        actionGitState?.kind !== "git-state" ||
-        actionGitState.commonDirectory === actionGitState.gitDirectory ||
-        scenario.expected.details?.project_identity_location !== actionGitState.commonDirectory ||
-        scenario.expected.details.checkout_identity_location !== actionGitState.gitDirectory ||
-        output.api?.primaryWorktreeRequired !== false
-      ) {
-        errors.push(`${scenario.id}: bare-repository worktree must not require a primary worktree`);
-      }
-    }
-    const checkedOutBranch = scenario.given.find(
-      (fact) => fact.kind === "branch" && fact.checkedOut,
-    );
-    const branchRename = scenario.given.find(
-      (fact) => fact.kind === "identity-transition" && fact.operation === "branch-rename",
-    );
-    const branchRecreation = scenario.given.find(
-      (fact) => fact.kind === "identity-transition" && fact.operation === "branch-delete-recreate",
-    );
-    if (branchRecreation?.kind === "identity-transition") {
-      const displacedContext = scenario.given.find(
-        (fact) =>
-          fact.kind === "identity-claim" &&
-          fact.scope === "context" &&
-          fact.status === "absent" &&
-          fact.owner === branchRecreation.from,
-      );
-      if (
-        typeof branchRecreation.from !== "string" ||
-        typeof branchRecreation.to !== "string" ||
-        branchRecreation.from !== branchRecreation.to ||
-        displacedContext?.kind !== "identity-claim" ||
-        scenario.expected.details?.orphaned_context_id !== displacedContext.id ||
-        output.json?.orphaned_context_id !== displacedContext.id
-      ) {
-        errors.push(`${scenario.id}: branch recreation must orphan the displaced context`);
-      }
-    }
-    if (branchRename?.kind === "identity-transition") {
-      const renamedContextWrite = scenario.expected.writes.find(
-        (write) =>
-          write.target === "git-config" &&
-          write.operation === "update" &&
-          checkedOutBranch?.kind === "branch" &&
-          write.id === checkedOutBranch.contextId,
-      );
-      if (
-        typeof branchRename.from !== "string" ||
-        typeof branchRename.to !== "string" ||
-        branchRename.from === branchRename.to ||
-        checkedOutBranch?.kind !== "branch" ||
-        checkedOutBranch.name !== branchRename.to ||
-        renamedContextWrite?.target !== "git-config" ||
-        renamedContextWrite.owner !== branchRename.to
-      ) {
-        errors.push(
-          `${scenario.id}: branch rename must match the checked-out branch and updated context owner`,
-        );
-      }
-    }
-    const refReplacement = scenario.given.find(
-      (fact) => fact.kind === "identity-transition" && fact.operation === "ref-replacement",
-    );
-    if (
-      refReplacement?.kind === "identity-transition" &&
-      (typeof refReplacement.from !== "string" ||
-        typeof refReplacement.to !== "string" ||
-        refReplacement.from === refReplacement.to ||
-        actionGitState?.kind !== "git-state" ||
-        refReplacement.to !== actionGitState.commit)
-    ) {
-      errors.push(`${scenario.id}: ref replacement target must match the action workspace commit`);
-    }
-    if (refReplacement?.kind === "identity-transition") {
-      const displacedContext = scenario.given.find(
-        (fact) =>
-          fact.kind === "identity-claim" &&
-          fact.scope === "context" &&
-          fact.status === "absent" &&
-          actionGitState?.kind === "git-state" &&
-          fact.owner === actionGitState.branch,
-      );
-      if (
-        displacedContext?.kind !== "identity-claim" ||
-        scenario.expected.details?.orphaned_context_id !== displacedContext.id ||
-        output.json?.orphaned_context_id !== displacedContext.id
-      ) {
-        errors.push(`${scenario.id}: ref replacement must orphan the displaced branch context`);
-      }
-    }
-    const branchRefs = scenario.given.filter((fact) => fact.kind === "branch-ref");
-    if (branchRefs.length > 0) {
-      const comparedBranches = scenario.given.filter(
-        (fact) =>
-          fact.kind === "branch" &&
-          !fact.checkedOut &&
-          branchRefs.some((ref) => ref.name === fact.name),
-      );
-      const comparedBranch = comparedBranches.length === 1 ? comparedBranches[0] : undefined;
-      const checkedOutRef =
-        checkedOutBranch?.kind === "branch"
-          ? branchRefs.find((fact) => fact.name === checkedOutBranch.name)
-          : undefined;
-      const comparedRef =
-        comparedBranch?.kind === "branch"
-          ? branchRefs.find((fact) => fact.name === comparedBranch.name)
-          : undefined;
-      if (
-        branchRefs.length !== 2 ||
-        comparedBranches.length !== 1 ||
-        actionGitState?.kind !== "git-state" ||
-        checkedOutBranch?.kind !== "branch" ||
-        comparedBranch?.kind !== "branch" ||
-        checkedOutRef?.kind !== "branch-ref" ||
-        comparedRef?.kind !== "branch-ref" ||
-        checkedOutRef.commit !== actionGitState.commit ||
-        comparedRef.commit !== actionGitState.commit ||
-        output.api?.otherContextId !== comparedBranch.contextId
-      ) {
-        errors.push(
-          `${scenario.id}: branch comparison must prove both refs share the checked-out commit`,
-        );
-      }
-    }
-    const requiredBranchHistory = requiredBranchHistoryByScenarioId[scenario.id];
-    const branchHistory = scenario.given.find((fact) => fact.kind === "branch-history");
-    const branchHistoryOperation = requiredBranchHistory ?? branchHistory?.operation;
-    if (branchHistoryOperation !== undefined) {
-      const expectedTransitionOperation =
-        branchHistoryOperation === "commit"
-          ? "branch-commit"
-          : branchHistoryOperation === "rebase"
-            ? "branch-rebase"
-            : "branch-reset";
-      const historyTransition = scenario.given.find(
-        (fact) =>
-          fact.kind === "identity-transition" && fact.operation === expectedTransitionOperation,
-      );
-      const historyBranch = scenario.given.find(
-        (fact) =>
-          fact.kind === "branch" &&
-          branchHistory?.kind === "branch-history" &&
-          fact.name === branchHistory.branch &&
-          fact.checkedOut,
-      );
-      if (
-        branchHistory?.kind !== "branch-history" ||
-        (requiredBranchHistory !== undefined &&
-          branchHistory.operation !== requiredBranchHistory) ||
-        historyTransition?.kind !== "identity-transition" ||
-        historyTransition.from !== branchHistory.fromCommit ||
-        historyTransition.to !== branchHistory.toCommit ||
-        historyBranch?.kind !== "branch" ||
-        scenario.expected.selection?.contextId !== historyBranch.contextId ||
-        scenario.expected.outcome !== "reuse"
-      ) {
-        errors.push(
-          `${scenario.id}: branch history preservation requires matching branch evidence`,
-        );
-      }
-    }
-    const symlinkAlias = scenario.given.find(
-      (fact) => fact.kind === "identity-transition" && fact.operation === "symlink-alias",
-    );
-    if (symlinkAlias?.kind === "identity-transition") {
-      const aliasWorkspace = scenario.given.find(
-        (fact) => fact.kind === "workspace" && fact.path === symlinkAlias.to,
-      );
-      const canonicalClaim = scenario.given.find(
-        (fact) =>
-          fact.kind === "identity-claim" &&
-          fact.scope === "checkout" &&
-          fact.status === "exact" &&
-          fact.path === symlinkAlias.from,
-      );
-      if (
-        aliasWorkspace?.kind !== "workspace" ||
-        aliasWorkspace.canonicalPath !== symlinkAlias.from ||
-        canonicalClaim?.kind !== "identity-claim" ||
-        output.json?.canonical_path !== symlinkAlias.from
-      ) {
-        errors.push(`${scenario.id}: symlink alias must report its canonical checkout path`);
-      }
-    }
-    if (scenario.expected.error?.code === "AMBIGUOUS_CONTEXT_OWNER") {
-      const projectedContextId = output.json?.context_id;
-      const projectedBranches = output.json?.branches;
-      const projectedBranchNames =
-        Array.isArray(projectedBranches) &&
-        projectedBranches.every((branch): branch is string => typeof branch === "string")
-          ? projectedBranches
-          : undefined;
-      const claimingBranchNames =
-        typeof projectedContextId === "string"
-          ? scenario.given.flatMap((fact) =>
-              fact.kind === "branch" && fact.contextId === projectedContextId ? [fact.name] : [],
-            )
-          : [];
-      const humanBranchNames = output.human?.fields.branches
-        ?.split(",")
-        .map((branch) => branch.trim())
-        .filter((branch) => branch.length > 0);
-      const ambiguousContextClaim = scenario.given.find(
-        (fact) =>
-          fact.kind === "identity-claim" &&
-          fact.scope === "context" &&
-          fact.id === projectedContextId &&
-          fact.status === "ambiguous",
-      );
-      const copiedBranchTransition = scenario.given.find(
-        (fact) =>
-          fact.kind === "identity-transition" &&
-          fact.operation === "branch-copy" &&
-          typeof fact.from === "string" &&
-          claimingBranchNames.includes(fact.from) &&
-          typeof fact.to === "string" &&
-          claimingBranchNames.includes(fact.to) &&
-          fact.originalExists === true,
-      );
-      if (
-        typeof projectedContextId !== "string" ||
-        new Set(claimingBranchNames).size < 2 ||
-        ambiguousContextClaim?.kind !== "identity-claim" ||
-        copiedBranchTransition?.kind !== "identity-transition" ||
-        projectedBranchNames === undefined ||
-        !managedStackContractStringSetEquals(claimingBranchNames, projectedBranchNames) ||
-        humanBranchNames === undefined ||
-        !managedStackContractStringSetEquals(claimingBranchNames, humanBranchNames)
-      ) {
-        errors.push(
-          `${scenario.id}: ambiguous context must bind at least two claiming branches to its projections`,
-        );
-      }
-    }
-    if (
-      scenario.when.interface === "git" &&
-      scenario.when.argv[0] === "branch" &&
-      (scenario.when.argv[1] === "-D" || scenario.when.argv[1] === "-d")
-    ) {
-      if (
-        scenario.expected.details?.stack_orphaned !== true ||
-        scenario.expected.details.stack_data_preserved !== true
-      ) {
-        errors.push(`${scenario.id}: branch deletion must preserve and orphan managed stack data`);
-      }
-      const deletedBranchName = scenario.when.argv[2];
-      const deletedBranch = scenario.given.find(
-        (fact) => fact.kind === "branch" && fact.name === deletedBranchName,
-      );
-      const affectedStack =
-        deletedBranch?.kind === "branch"
-          ? scenario.given.find(
-              (fact) => fact.kind === "stack" && fact.contextId === deletedBranch.contextId,
-            )
-          : undefined;
-      if (
-        deletedBranchName === undefined ||
-        deletedBranch?.kind !== "branch" ||
-        affectedStack?.kind !== "stack"
-      ) {
-        errors.push(
-          `${scenario.id}: branch deletion must bind its branch to an affected managed stack`,
-        );
-      } else {
-        if (deletedBranch.checkedOut) {
-          errors.push(`${scenario.id}: deleted branch ${deletedBranchName} cannot be checked out`);
-        }
-        if (scenario.expected.details?.orphaned_stack_id !== affectedStack.stackId) {
-          errors.push(
-            `${scenario.id}: orphaned stack must be ${affectedStack.stackId} for branch ${deletedBranchName}`,
-          );
-        }
-      }
-      if (
-        actionCwd === undefined ||
-        !scenario.given.some((fact) => fact.kind === "checkout" && fact.path === actionCwd) ||
-        !scenario.given.some(
-          (fact) => fact.kind === "git-state" && fact.workspacePath === actionCwd,
-        )
-      ) {
-        errors.push(`${scenario.id}: branch deletion must declare checkout Git state`);
-      }
-    }
-    if (
-      actionCwd !== undefined &&
-      scenario.given.some(
-        (fact) =>
-          fact.kind === "workspace" &&
-          fact.path === actionCwd &&
-          (fact.mode === "bare-worktree" || fact.mode === "linked-worktree"),
-      ) &&
-      !scenario.given.some((fact) => fact.kind === "git-state" && fact.workspacePath === actionCwd)
-    ) {
-      errors.push(`${scenario.id}: resolving worktree ${actionCwd} requires its Git state`);
-    }
-
-    const detachedGitState = scenario.given.find(
-      (fact) =>
-        fact.kind === "git-state" && fact.workspacePath === actionCwd && fact.head === "detached",
-    );
-    if (
-      detachedGitState?.kind === "git-state" &&
-      scenario.expected.outcome === "reuse" &&
-      !scenario.given.some(
-        (fact) =>
-          fact.kind === "identity-transition" &&
-          fact.operation === "detached-commit" &&
-          fact.to === detachedGitState.commit,
-      )
-    ) {
-      errors.push(`${scenario.id}: detached reuse must declare the commit transition`);
-    }
-
-    if (scenario.expected.error?.code === "DUPLICATE_CHECKOUT_CLAIM") {
-      const copiedWorkspace = scenario.given.find(
-        (fact) => fact.kind === "workspace" && fact.path === actionCwd,
-      );
-      const duplicateClaim = scenario.given.find(
-        (fact) => fact.kind === "identity-claim" && fact.scope === "checkout",
-      );
-      const copyTransition = scenario.given.find(
-        (fact) => fact.kind === "identity-transition" && fact.operation === "checkout-copy",
-      );
-      const projectedPaths = output.json?.paths;
-      if (
-        copiedWorkspace?.kind !== "workspace" ||
-        copiedWorkspace.copiedFrom === undefined ||
-        duplicateClaim?.kind !== "identity-claim" ||
-        duplicateClaim.status !== "duplicate" ||
-        duplicateClaim.path !== copiedWorkspace.copiedFrom ||
-        copyTransition?.kind !== "identity-transition" ||
-        copyTransition.from !== copiedWorkspace.copiedFrom ||
-        copyTransition.to !== copiedWorkspace.path ||
-        output.json?.checkout_id !== duplicateClaim.id ||
-        !Array.isArray(projectedPaths) ||
-        projectedPaths.length !== 2 ||
-        !projectedPaths.includes(copiedWorkspace.path) ||
-        !projectedPaths.includes(copiedWorkspace.copiedFrom)
-      ) {
-        errors.push(
-          `${scenario.id}: duplicate checkout error must bind both conflicting live paths`,
-        );
-      }
-    }
-
-    if (scenario.expected.error?.code === "CHECKOUT_PATH_INACCESSIBLE") {
-      const movedWorkspace = scenario.given.find(
-        (fact) => fact.kind === "workspace" && fact.path === actionCwd,
-      );
-      const ambiguousClaim = scenario.given.find(
-        (fact) => fact.kind === "identity-claim" && fact.scope === "checkout",
-      );
-      if (
-        movedWorkspace?.kind !== "workspace" ||
-        movedWorkspace.previousPathAccess !== "inaccessible" ||
-        movedWorkspace.previousPath === undefined ||
-        ambiguousClaim?.kind !== "identity-claim" ||
-        ambiguousClaim.status !== "ambiguous" ||
-        ambiguousClaim.path !== movedWorkspace.previousPath ||
-        output.human?.fields.previousPath !== movedWorkspace.previousPath ||
-        output.human?.fields.currentPath !== movedWorkspace.path ||
-        output.json?.checkout_id !== ambiguousClaim.id
-      ) {
-        errors.push(
-          `${scenario.id}: inaccessible checkout error must bind path access and ambiguous claim`,
-        );
-      }
-    }
-
-    const movedWorkspace = scenario.given.find(
-      (fact) => fact.kind === "workspace" && fact.path === actionCwd,
-    );
-    const exactCheckoutClaim = scenario.given.find(
-      (fact) => fact.kind === "identity-claim" && fact.scope === "checkout",
-    );
-    if (
-      movedWorkspace?.kind === "workspace" &&
-      movedWorkspace.previousPathAccess === "missing" &&
-      exactCheckoutClaim?.kind === "identity-claim" &&
-      exactCheckoutClaim.status === "exact" &&
-      scenario.expected.outcome === "reuse"
-    ) {
-      const apiRebindMatches =
-        output.api === undefined ||
-        (output.api.rebound === true && output.api.checkoutId === exactCheckoutClaim.id);
-      const jsonRebindMatches =
-        output.json === undefined ||
-        (output.json.rebound_from === movedWorkspace.previousPath &&
-          output.json.checkout_id === exactCheckoutClaim.id);
-      if (
-        movedWorkspace.previousPath === undefined ||
-        exactCheckoutClaim.path !== movedWorkspace.previousPath ||
-        !apiRebindMatches ||
-        !jsonRebindMatches
-      ) {
-        errors.push(`${scenario.id}: automatic checkout rebind requires a missing previous path`);
-      }
-      if (
-        !scenario.expected.writes.some(
-          (write) =>
-            write.target === "registry" &&
-            write.operation === "update" &&
-            write.id === exactCheckoutClaim.id,
-        )
-      ) {
-        errors.push(
-          `${scenario.id}: automatic checkout rebind must persist the checkout registry update`,
-        );
-      }
-    }
-
-    if (scenario.expected.error?.code === "AMBIGUOUS_FOLDER_TO_GIT_IDENTITY") {
-      const folderToGitTransition = scenario.given.find(
-        (fact) => fact.kind === "identity-transition" && fact.operation === "folder-to-git",
-      );
-      const actionWorkspace = scenario.given.find(
-        (fact) => fact.kind === "workspace" && fact.path === actionCwd,
-      );
-      const ambiguousProjectClaim = scenario.given.find(
-        (fact) =>
-          fact.kind === "identity-claim" &&
-          fact.scope === "project" &&
-          fact.status === "ambiguous" &&
-          fact.path === actionCwd,
-      );
-      if (
-        folderToGitTransition?.kind !== "identity-transition" ||
-        actionWorkspace?.kind !== "workspace" ||
-        actionWorkspace.mode !== "git" ||
-        ambiguousProjectClaim?.kind !== "identity-claim"
-      ) {
-        errors.push(
-          `${scenario.id}: folder-to-Git ambiguity error requires an ambiguous live project claim`,
-        );
-      }
-    }
-
-    for (const fact of scenario.given) {
-      if (fact.kind !== "native-qualification") {
-        continue;
-      }
-
-      if (
-        !managedNativeServiceMatrix.targetPlatforms.includes(fact.platform) &&
-        !managedNativeServiceMatrix.unsupportedPlatforms.includes(fact.platform)
-      ) {
-        errors.push(`${scenario.id}: native qualification uses unknown platform ${fact.platform}`);
-      }
-      if (
-        scenario.when.interface === "managed-api" &&
-        scenario.when.method === "preflightNative" &&
-        (typeof scenario.when.input.platform !== "string" ||
-          scenario.when.input.platform !== fact.platform ||
-          scenario.expected.output.api?.platform !== fact.platform)
-      ) {
-        errors.push(
-          `${scenario.id}: native qualification platform must match the preflight action`,
-        );
-      }
-
-      const qualified = new Set<ServiceName>();
-      const failed = new Set<ServiceName>();
-      for (const service of fact.qualifiedServices) {
-        if (!nativeServiceSet.has(service)) {
-          errors.push(`${scenario.id}: native qualification contains unknown service ${service}`);
-        }
-        if (qualified.has(service)) {
-          errors.push(`${scenario.id}: native qualification duplicates service ${service}`);
-        }
-        qualified.add(service);
-      }
-      for (const service of fact.failedServices) {
-        if (!nativeServiceSet.has(service)) {
-          errors.push(`${scenario.id}: native qualification contains unknown service ${service}`);
-        }
-        if (failed.has(service)) {
-          errors.push(`${scenario.id}: native qualification duplicates service ${service}`);
-        }
-        if (qualified.has(service)) {
-          errors.push(`${scenario.id}: native qualification places ${service} in both partitions`);
-        }
-        failed.add(service);
-      }
-      for (const service of nativeServices) {
-        if (!qualified.has(service) && !failed.has(service)) {
-          errors.push(`${scenario.id}: native qualification omits service ${service}`);
-        }
-      }
-      if (scenario.when.interface === "managed-api" && scenario.when.method === "preflightNative") {
-        const platformSupported = managedNativeServiceMatrix.targetPlatforms.includes(
-          fact.platform,
-        );
-        if (!platformSupported && scenario.expected.error?.code !== "NATIVE_PLATFORM_UNSUPPORTED") {
-          errors.push(
-            `${scenario.id}: unsupported native platform must use the dedicated preflight error`,
-          );
-        }
-        const platformQualified = failed.size === 0 && qualified.size === nativeServices.length;
-        if (
-          platformSupported &&
-          !platformQualified &&
-          scenario.expected.error?.code !== "NATIVE_PLATFORM_NOT_QUALIFIED"
-        ) {
-          errors.push(
-            `${scenario.id}: supported unqualified native platform must use NATIVE_PLATFORM_NOT_QUALIFIED`,
-          );
-        }
-        if (
-          scenario.expected.outcome !== (platformQualified ? "report" : "error") ||
-          scenario.expected.details?.qualified !== platformQualified ||
-          scenario.expected.details.qualified_service_count !== qualified.size ||
-          scenario.expected.details.failed_service_count !== failed.size ||
-          scenario.expected.output.api?.qualified !== platformQualified
-        ) {
-          errors.push(
-            `${scenario.id}: native preflight decision must match its qualification partitions`,
-          );
-        }
-        const projectedServices = scenario.expected.output.api?.services;
-        if (
-          platformQualified &&
-          (!Array.isArray(projectedServices) ||
-            !managedStackContractJsonEquals(projectedServices, fact.qualifiedServices))
-        ) {
-          errors.push(`${scenario.id}: native preflight services must match qualified services`);
-        }
-        const projectedFailures = scenario.expected.output.api?.failedServices;
-        if (
-          !platformQualified &&
-          (!Array.isArray(projectedFailures) ||
-            !managedStackContractJsonEquals(projectedFailures, fact.failedServices))
-        ) {
-          errors.push(`${scenario.id}: native preflight failures must match failed services`);
-        }
-        const projectedAvailableServices = scenario.expected.output.api?.availableServices;
-        if (
-          !platformQualified &&
-          (!Array.isArray(projectedAvailableServices) ||
-            projectedAvailableServices.length !== 0 ||
-            scenario.expected.details?.reduced_graph !== false ||
-            scenario.expected.details.docker_fallback_per_service !== false)
-        ) {
-          errors.push(
-            `${scenario.id}: failed native qualification must expose no reduced service graph`,
-          );
-        }
-      }
-    }
-
-    if (scenario.expected.error?.code === "NATIVE_PLATFORM_UNSUPPORTED") {
-      const qualification = scenario.given.find((fact) => fact.kind === "native-qualification");
-      const projectedPlatforms = output.json?.supported_platforms;
-      if (
-        qualification?.kind !== "native-qualification" ||
-        !managedNativeServiceMatrix.unsupportedPlatforms.includes(qualification.platform) ||
-        output.json?.platform !== qualification.platform ||
-        !Array.isArray(projectedPlatforms) ||
-        projectedPlatforms.length !== managedNativeServiceMatrix.targetPlatforms.length ||
-        !managedNativeServiceMatrix.targetPlatforms.every((platform) =>
-          projectedPlatforms.includes(platform),
-        ) ||
-        explicitRuntime !== "native"
-      ) {
-        errors.push(`${scenario.id}: unsupported native error must bind an unsupported platform`);
-      }
-    }
-
-    const writesIdentityMarker = scenario.expected.writes.some(
-      (write) => write.target === "identity-marker",
-    );
-    const clonedWorkspace = scenario.given.find(
-      (fact) =>
-        fact.kind === "workspace" && fact.path === actionCwd && fact.clonedFrom !== undefined,
-    );
-    if (
-      clonedWorkspace?.kind === "workspace" &&
-      scenario.expected.outcome === "create" &&
-      scenario.expected.writes.some((write) => write.target === "git-config") &&
-      (scenario.expected.details?.git_index_mutated !== false || writesIdentityMarker)
-    ) {
-      errors.push(`${scenario.id}: fresh clone identity creation must not mutate the Git index`);
-    }
-    const trackedIdentityMarker = scenario.given.find(
-      (fact) => fact.kind === "git-state" && fact.trackedIdentityMarker === true,
-    );
-    if (
-      trackedIdentityMarker?.kind === "git-state" &&
-      (scenario.expected.details?.tracked_marker_ignored !== true ||
-        scenario.expected.details.tracked_marker_mutated !== false ||
-        scenario.expected.details.git_index_mutated !== false ||
-        output.json?.tracked_marker_ignored !== true ||
-        writesIdentityMarker)
-    ) {
-      errors.push(`${scenario.id}: tracked identity marker must remain ignored and unmodified`);
-    }
-    if (writesIdentityMarker) {
-      if (
-        scenario.given.some(
-          (fact) => fact.kind === "git-state" && fact.trackedIdentityMarker === true,
-        )
-      ) {
-        errors.push(`${scenario.id}: a tracked identity marker must remain untouched`);
-      } else if (
-        scenario.given.some((fact) => fact.kind === "workspace" && fact.mode !== "ordinary-folder")
-      ) {
-        errors.push(`${scenario.id}: Git workspace identity must use Git-local metadata`);
-      }
-      if (scenario.expected.details?.identity_marker_tracked !== false) {
-        errors.push(`${scenario.id}: ordinary-folder identity marker must remain untracked`);
-      }
-    }
-
-    if (scenario.expected.outcome !== "create") {
-      for (const effect of scenario.expected.runtimeEffects) {
-        if (effect.operation !== "start") {
-          continue;
-        }
-        const explicitlyStopped = scenario.given.some(
-          (fact) =>
-            fact.kind === "stack" &&
-            fact.stackId === effect.stackId &&
-            fact.lifecycle === "stopped",
-        );
-        if (!explicitlyStopped) {
-          errors.push(
-            `${scenario.id}: starting existing stack ${effect.stackId} requires an explicit stopped lifecycle`,
-          );
-        }
-      }
-    }
-
-    for (const target of scenario.given) {
-      if (
-        target.kind === "managed-target" &&
-        !target.exists &&
-        scenario.given.some((fact) => fact.kind === "stack" && fact.stackId === target.stackId)
-      ) {
-        errors.push(
-          `${scenario.id}: absent managed target ${target.stackId} contradicts an existing stack`,
-        );
-      }
-    }
-    if (isManagedStartAction(scenario.when) && scenario.expected.outcome === "reuse") {
-      const existingStartedTarget = scenario.given.find(
-        (fact) =>
-          fact.kind === "managed-target" &&
-          fact.exists &&
-          scenario.expected.runtimeEffects.some(
-            (effect) => effect.operation === "start" && effect.stackId === fact.stackId,
-          ),
-      );
-      if (
-        existingStartedTarget?.kind === "managed-target" &&
-        (scenario.expected.details?.legacy_state_read !== false ||
-          (output.api !== undefined &&
-            (output.api.bootstrap !== "not-attempted" || output.api.legacyStateRead !== false)) ||
-          (output.json !== undefined && output.json.bootstrap !== "not-attempted"))
-      ) {
-        errors.push(`${scenario.id}: existing managed target must not report legacy bootstrap`);
-      }
-      const legacyState = scenario.given.find((fact) => fact.kind === "legacy-state");
-      if (
-        existingStartedTarget?.kind === "managed-target" &&
-        legacyState?.kind === "legacy-state" &&
-        (scenario.expected.details?.legacy_state_mutated !== false ||
-          (output.json?.legacy_state_mutated !== undefined &&
-            output.json.legacy_state_mutated !== false))
-      ) {
-        errors.push(`${scenario.id}: managed target reuse must not mutate legacy state`);
-      }
-      if (
-        existingStartedTarget?.kind === "managed-target" &&
-        legacyState?.kind === "legacy-state" &&
-        (legacyState.database === "incompatible" ||
-          legacyState.storage === "incompatible" ||
-          legacyState.credentials === "incompatible") &&
-        (scenario.expected.details?.timelines_diverged !== true ||
-          output.json?.timelines_diverged !== true)
-      ) {
-        errors.push(`${scenario.id}: managed restart must report legacy timeline divergence`);
-      }
-    }
-
-    const createdStackIds = scenario.expected.writes.flatMap((write) =>
-      write.target === "managed-state" && write.operation === "create" ? [write.id] : [],
-    );
-    for (const stackId of createdStackIds) {
-      if (
-        !scenario.given.some(
-          (fact) => fact.kind === "managed-target" && fact.stackId === stackId && !fact.exists,
-        )
-      ) {
-        errors.push(`${scenario.id}: managed creation must declare absent target ${stackId}`);
-      }
-    }
-    if (
-      isManagedStartAction(scenario.when) &&
-      scenario.expected.outcome === "create" &&
-      createdStackIds.length > 0
-    ) {
-      const legacyState = scenario.given.find((fact) => fact.kind === "legacy-state");
-      const legacyAllowsFreshCreation =
-        legacyState?.kind === "legacy-state" &&
-        ((legacyState.lifecycle === "absent" &&
-          legacyState.database === "absent" &&
-          legacyState.storage === "absent" &&
-          legacyState.credentials === "absent") ||
-          (legacyState.lifecycle === "stopped" &&
-            (legacyState.database === "incompatible" ||
-              legacyState.storage === "incompatible" ||
-              legacyState.credentials === "incompatible")));
-      if (!legacyAllowsFreshCreation) {
-        errors.push(
-          `${scenario.id}: managed creation must declare legacy state absent or incompatible`,
-        );
-      }
-      const legacyIsFullyAbsent =
-        legacyState?.kind === "legacy-state" &&
-        legacyState.lifecycle === "absent" &&
-        legacyState.database === "absent" &&
-        legacyState.storage === "absent" &&
-        legacyState.credentials === "absent";
-      if (
-        legacyIsFullyAbsent &&
-        ((scenario.expected.details?.legacy_state_mutated !== undefined &&
-          scenario.expected.details.legacy_state_mutated !== false) ||
-          (output.json?.legacy_state_mutated !== undefined &&
-            output.json.legacy_state_mutated !== false))
-      ) {
-        errors.push(`${scenario.id}: absent legacy bootstrap must not report mutation`);
-      }
-      if (
-        legacyState?.kind === "legacy-state" &&
-        legacyState.lifecycle === "stopped" &&
-        (legacyState.database === "incompatible" ||
-          legacyState.storage === "incompatible" ||
-          legacyState.credentials === "incompatible") &&
-        (scenario.expected.details?.legacy_state_mutated !== false ||
-          output.json?.legacy_state_mutated !== false ||
-          scenario.expected.writes.some(
-            (write) => write.target === "managed-state" && write.operation === "copy",
-          ) ||
-          scenario.expected.runtimeEffects.some((effect) => effect.operation === "copy"))
-      ) {
-        errors.push(`${scenario.id}: fresh bootstrap must not copy or mutate legacy state`);
-      }
-    }
-
-    const copiedStackIds = new Set(
-      scenario.expected.writes.flatMap((write) =>
-        write.target === "managed-state" && write.operation === "copy" ? [write.id] : [],
-      ),
-    );
-    for (const effect of scenario.expected.runtimeEffects) {
-      if (effect.operation === "copy") {
-        copiedStackIds.add(effect.stackId);
-      }
-    }
-    if (copiedStackIds.size > 0) {
-      const legacyState = scenario.given.find((fact) => fact.kind === "legacy-state");
-      const legacyIsCopyable =
-        legacyState?.kind === "legacy-state" &&
-        legacyState.lifecycle === "stopped" &&
-        legacyState.database === "compatible" &&
-        legacyState.storage === "compatible" &&
-        legacyState.credentials === "compatible";
-      for (const stackId of copiedStackIds) {
-        if (
-          !scenario.given.some(
-            (fact) => fact.kind === "managed-target" && fact.stackId === stackId && !fact.exists,
-          ) ||
-          !legacyIsCopyable
-        ) {
-          errors.push(
-            `${scenario.id}: bootstrap copy requires absent target ${stackId} and compatible stopped legacy state`,
-          );
-        }
-      }
-      if (
-        scenario.expected.details?.legacy_state_mutated !== false ||
-        (output.json !== undefined && output.json.legacy_state_mutated !== false)
-      ) {
-        errors.push(`${scenario.id}: bootstrap copy must not mutate legacy state`);
-      }
-    }
-
-    for (const effect of scenario.expected.runtimeEffects) {
-      if (effect.operation !== "start") {
-        continue;
-      }
-      const createsTarget = scenario.expected.writes.some(
-        (write) =>
-          write.target === "managed-state" &&
-          (write.operation === "create" || write.operation === "copy") &&
-          write.id === effect.stackId,
-      );
-      const targetExists = scenario.given.some(
-        (fact) =>
-          (fact.kind === "managed-target" && fact.stackId === effect.stackId && fact.exists) ||
-          (fact.kind === "stack" && fact.stackId === effect.stackId),
-      );
-      if (!createsTarget && !targetExists) {
-        errors.push(
-          `${scenario.id}: starting existing stack ${effect.stackId} requires an existing managed target`,
-        );
-      }
-    }
-
-    if (scenario.expected.error?.code === "PERSISTED_RUNTIME_UNAVAILABLE") {
-      if (scenario.expected.details?.switched_to_docker !== false) {
-        errors.push(`${scenario.id}: unavailable persisted runtime must not report a switch`);
-      }
-      for (const fact of scenario.given) {
-        if (fact.kind !== "persisted-runtime") {
-          continue;
-        }
-        const explicitlyStopped = scenario.given.some(
-          (candidate) =>
-            candidate.kind === "stack" &&
-            candidate.stackId === fact.stackId &&
-            candidate.lifecycle === "stopped",
-        );
-        if (!explicitlyStopped) {
-          errors.push(
-            `${scenario.id}: persisted runtime failure for ${fact.stackId} requires an explicit stopped lifecycle`,
-          );
-        }
-      }
-    }
-
-    for (const write of scenario.expected.writes) {
-      if (
-        write.operation === "copy" ||
-        write.operation === "create" ||
-        write.operation === "publish"
-      ) {
-        declaredIds.add(write.id);
-      }
-      if (write.target === "identity-marker") {
-        declaredIds.add(write.projectId);
-        declaredIds.add(write.checkoutId);
-        declaredIds.add(write.contextId);
-        projectIds.add(write.projectId);
-        checkoutIds.add(write.checkoutId);
-        contextIds.add(write.contextId);
-      }
-    }
-    for (const write of scenario.expected.writes) {
-      if (
-        write.operation !== "copy" &&
-        write.operation !== "create" &&
-        write.operation !== "publish" &&
-        !declaredIds.has(write.id)
-      ) {
-        errors.push(
-          `${scenario.id}: ${write.target} ${write.operation} references undeclared ID ${write.id}`,
-        );
-      }
-    }
-
-    const selection = scenario.expected.selection;
-    if (selection !== undefined) {
-      projectIds.add(selection.projectId);
-      checkoutIds.add(selection.checkoutId);
-      contextIds.add(selection.contextId);
-      for (const id of [
-        selection.projectId,
-        selection.checkoutId,
-        selection.contextId,
-        selection.stackId,
-      ]) {
-        if (!declaredIds.has(id)) {
-          errors.push(`${scenario.id}: selection references undeclared ID ${id}`);
-        }
-      }
-
-      const selectedStack = scenario.given.find(
-        (fact) => fact.kind === "stack" && fact.stackId === selection.stackId,
-      );
-      if (selectedStack?.kind === "stack") {
-        if (selectedStack.contextId !== selection.contextId) {
-          errors.push(
-            `${scenario.id}: selected stack ${selection.stackId} belongs to context ${selectedStack.contextId}, not ${selection.contextId}`,
-          );
-        }
-        if (selectedStack.name !== selection.stackName) {
-          errors.push(
-            `${scenario.id}: selected stack ${selection.stackId} is named ${selectedStack.name}, not ${selection.stackName}`,
-          );
-        }
-      }
-
-      const actionCheckout = scenario.given.find(
-        (fact) => fact.kind === "checkout" && fact.path === actionCwd,
-      );
-      if (
-        actionCheckout?.kind === "checkout" &&
-        (selection.projectId !== actionCheckout.projectId ||
-          selection.checkoutId !== actionCheckout.checkoutId)
-      ) {
-        errors.push(
-          `${scenario.id}: selection must use checkout ${actionCheckout.checkoutId} for ${actionCwd}`,
-        );
-      }
-
-      const declaredBranches = scenario.given.filter((fact) => fact.kind === "branch");
-      const checkedOutBranches = declaredBranches.filter((fact) => fact.checkedOut);
-      if (declaredBranches.length > 0 && checkedOutBranches.length !== 1) {
-        errors.push(
-          `${scenario.id}: contextual Git selection requires exactly one checked-out branch`,
-        );
-      }
-
-      if (checkedOutBranch?.kind === "branch") {
-        const createsSelectedContext = scenario.expected.writes.some(
-          (write) =>
-            write.target === "git-config" &&
-            write.operation === "create" &&
-            write.id === selection.contextId,
-        );
-        const hasCheckoutScopedContextClaim = scenario.given.some(
-          (fact) =>
-            fact.kind === "identity-claim" &&
-            fact.scope === "context" &&
-            fact.id === selection.contextId &&
-            fact.status === "exact" &&
-            (fact.owner === checkedOutBranch.name ||
-              fact.owner === `${selection.checkoutId}/${checkedOutBranch.name}`),
-        );
-        if (
-          (actionGitState?.kind === "git-state" &&
-            actionGitState.head === "branch" &&
-            actionGitState.branch !== checkedOutBranch.name) ||
-          (selection.contextId !== checkedOutBranch.contextId &&
-            !createsSelectedContext &&
-            !hasCheckoutScopedContextClaim)
-        ) {
-          errors.push(
-            `${scenario.id}: selected context must match the active Git branch and checked-out branch fact`,
-          );
-        }
-      }
-
-      if (scenario.when.interface === "managed-api" && scenario.when.method === "resolveStack") {
-        for (const write of scenario.expected.writes) {
-          if (write.target !== "git-config" || write.operation !== "create") {
-            continue;
-          }
-          const scope =
-            write.id === selection.projectId
-              ? "project"
-              : write.id === selection.checkoutId
-                ? "checkout"
-                : write.id === selection.contextId
-                  ? "context"
-                  : undefined;
-          if (
-            scope !== undefined &&
-            !scenario.given.some(
-              (fact) =>
-                fact.kind === "identity-claim" &&
-                fact.scope === scope &&
-                fact.id === write.id &&
-                fact.status === "absent",
-            )
-          ) {
-            errors.push(
-              `${scenario.id}: creating Git identity ${write.id} requires an absent ${scope} claim`,
-            );
-          }
-        }
-      }
-
-      if (
-        scenario.given.some(
-          (fact) => fact.kind === "identity-transition" && fact.operation === "folder-to-git",
-        )
-      ) {
-        for (const id of [selection.projectId, selection.checkoutId, selection.contextId]) {
-          if (
-            !scenario.expected.writes.some(
-              (write) => write.target === "git-config" && write.id === id,
-            )
-          ) {
-            errors.push(
-              `${scenario.id}: folder-to-Git identity ${id} must be persisted in Git-local metadata`,
-            );
-          }
-        }
-        if (
-          scenario.expected.outcome === "reuse" &&
-          (!scenario.given.some(
-            (fact) =>
-              fact.kind === "identity-claim" &&
-              fact.scope === "project" &&
-              fact.id === selection.projectId &&
-              fact.path === actionCwd &&
-              fact.status === "exact",
-          ) ||
-            !scenario.given.some(
-              (fact) =>
-                fact.kind === "identity-claim" &&
-                fact.scope === "checkout" &&
-                fact.id === selection.checkoutId &&
-                fact.path === actionCwd &&
-                fact.status === "exact",
-            ))
-        ) {
-          errors.push(
-            `${scenario.id}: folder-to-Git reuse requires exact project and checkout claims`,
-          );
-        }
-      }
-
-      const createsSelectedContext = scenario.expected.writes.some(
-        (write) =>
-          write.target === "git-config" &&
-          write.operation === "create" &&
-          write.id === selection.contextId,
-      );
-      const derivesContextFromGit = scenario.given.some(
-        (fact) =>
-          fact.kind === "identity-transition" &&
-          (fact.operation === "clone" ||
-            fact.operation === "branch-delete-recreate" ||
-            fact.operation === "folder-to-git" ||
-            fact.operation === "ref-replacement"),
-      );
-      const contextAlreadyDeclaredByBranch = scenario.given.some(
-        (fact) =>
-          fact.kind === "branch" && fact.contextId === selection.contextId && fact.checkedOut,
-      );
-      if (
-        createsSelectedContext &&
-        derivesContextFromGit &&
-        !contextAlreadyDeclaredByBranch &&
-        (actionCwd === undefined ||
-          !scenario.given.some(
-            (fact) => fact.kind === "git-state" && fact.workspacePath === actionCwd,
-          ))
-      ) {
-        errors.push(`${scenario.id}: creating a Git context requires Git state for the workspace`);
-      }
-
-      const recoveredIdentityMarker = scenario.given.find(
-        (fact) =>
-          fact.kind === "identity-marker" &&
-          fact.workspacePath === actionCwd &&
-          fact.projectId === selection.projectId &&
-          fact.checkoutId === selection.checkoutId &&
-          fact.contextId === selection.contextId,
-      );
-      if (
-        recoveredIdentityMarker?.kind === "identity-marker" &&
-        !scenario.given.some(
-          (fact) =>
-            fact.kind === "workspace" &&
-            fact.mode === "ordinary-folder" &&
-            (fact.path === actionCwd || fact.canonicalPath === actionCwd),
-        )
-      ) {
-        errors.push(`${scenario.id}: local identity marker recovery requires an ordinary folder`);
-      }
-
-      const ordinaryWorkspace = scenario.given.find(
-        (fact) =>
-          fact.kind === "workspace" &&
-          fact.mode === "ordinary-folder" &&
-          (fact.path === actionCwd || fact.canonicalPath === actionCwd),
-      );
-      if (ordinaryWorkspace?.kind === "workspace") {
-        if (scenario.expected.outcome === "create") {
-          if (
-            !scenario.expected.writes.some(
-              (write) =>
-                write.target === "identity-marker" &&
-                write.workspacePath === actionCwd &&
-                write.projectId === selection.projectId &&
-                write.checkoutId === selection.checkoutId &&
-                write.contextId === selection.contextId,
-            )
-          ) {
-            errors.push(
-              `${scenario.id}: ordinary-folder creation must persist its identity marker`,
-            );
-          }
-        } else if (
-          !scenario.given.some(
-            (fact) =>
-              fact.kind === "identity-marker" &&
-              fact.workspacePath === actionCwd &&
-              fact.projectId === selection.projectId &&
-              fact.checkoutId === selection.checkoutId &&
-              fact.contextId === selection.contextId,
-          )
-        ) {
-          errors.push(`${scenario.id}: ordinary-folder reuse must resolve its identity marker`);
-        }
-      }
-    }
-
-    if (scenario.when.interface === "managed-api" && scenario.when.method === "startConcurrently") {
-      const concurrencyFacts = scenario.given.filter(
-        (fact) => fact.kind === "concurrent-operation" && fact.operation === "create-stack",
-      );
-      const concurrencyFact = concurrencyFacts.length === 1 ? concurrencyFacts[0] : undefined;
-      if (concurrencyFact?.kind !== "concurrent-operation") {
-        errors.push(`${scenario.id}: concurrent start requires exactly one create-stack fact`);
-      } else {
-        const actionContenders = scenario.when.input.contenders;
-        if (
-          typeof actionContenders !== "number" ||
-          !Number.isInteger(actionContenders) ||
-          actionContenders < 2 ||
-          actionContenders !== concurrencyFact.contenders
-        ) {
-          errors.push(
-            `${scenario.id}: concurrent action contenders must match the declared race of ${concurrencyFact.contenders}`,
-          );
-        }
-
-        const actionStackName = scenario.when.input.stackName;
-        const actionTarget =
-          selection !== undefined && typeof actionStackName === "string"
-            ? `${selection.contextId}/${actionStackName}`
-            : undefined;
-        if (actionTarget !== concurrencyFact.target) {
-          errors.push(
-            `${scenario.id}: concurrent action target must match ${concurrencyFact.target}`,
-          );
-        }
-
-        const contenderResults = [
-          {
-            projection: "details",
-            results: scenario.expected.details?.contender_results,
-          },
-          {
-            projection: "API",
-            results: scenario.expected.output.api?.contenderResults,
-          },
-        ];
-        for (const { projection, results } of contenderResults) {
-          if (!Array.isArray(results) || results.length !== concurrencyFact.contenders) {
-            errors.push(
-              `${scenario.id}: concurrent ${projection} results must cover ${concurrencyFact.contenders} contenders`,
-            );
-          }
-        }
-        const detailResults = scenario.expected.details?.contender_results;
-        const apiResults = scenario.expected.output.api?.contenderResults;
-        if (
-          Array.isArray(detailResults) &&
-          Array.isArray(apiResults) &&
-          !managedStackContractJsonEquals(detailResults, apiResults)
-        ) {
-          errors.push(`${scenario.id}: concurrent result projections must agree`);
-        }
-        if (
-          Array.isArray(detailResults) &&
-          (detailResults.filter((result) => result === "create").length !== 1 ||
-            detailResults.some((result) => result !== "create" && result !== "reuse"))
-        ) {
-          errors.push(`${scenario.id}: concurrent race must create once and reuse thereafter`);
-        }
-        if (
-          scenario.expected.details?.published_stack_count !== 1 ||
-          scenario.expected.output.api?.publishedStackCount !== 1 ||
-          scenario.expected.details?.alias_count !== 0 ||
-          scenario.expected.output.api?.aliasCount !== 0
-        ) {
-          errors.push(`${scenario.id}: concurrent race must publish one stack without aliases`);
-        }
-      }
-    }
-
-    if (scenario.expected.error?.code === "EXACT_PORT_OCCUPIED") {
-      const configuredPort = scenario.given.find(
-        (fact) => fact.kind === "config-port" && fact.intent === "exact",
-      );
-      const occupiedPort =
-        configuredPort?.kind === "config-port"
-          ? scenario.given.find(
-              (fact) => fact.kind === "occupied-port" && fact.port === configuredPort.value,
-            )
-          : undefined;
-      if (
-        configuredPort?.kind !== "config-port" ||
-        typeof configuredPort.value !== "number" ||
-        occupiedPort?.kind !== "occupied-port" ||
-        output.human?.fields.port !== String(configuredPort.value) ||
-        output.human?.fields.configKey !== configuredPort.key ||
-        output.human?.fields.owner !== occupiedPort.owner ||
-        output.json?.port !== configuredPort.value ||
-        output.json?.config_key !== configuredPort.key ||
-        output.json?.owner !== occupiedPort.owner
-      ) {
-        errors.push(
-          `${scenario.id}: exact port conflict must bind config, occupancy, and projections`,
-        );
-      }
-    }
-
-    const managedPortOwners = scenario.given.flatMap((fact) =>
-      fact.kind === "occupied-port" && fact.owner === "managed-stack" ? [fact] : [],
-    );
-    if (scenario.expected.error?.code === "EXACT_PORT_OCCUPIED" && managedPortOwners.length > 0) {
-      if (selection === undefined) {
-        errors.push(`${scenario.id}: managed sibling port conflict requires a selected target`);
-      }
-      for (const owner of managedPortOwners) {
-        if (owner.ownerId.trim().length === 0) {
-          errors.push(`${scenario.id}: managed sibling port owner requires a stack ID`);
-        }
-        if (owner.ownerId === selection?.stackId) {
-          errors.push(
-            `${scenario.id}: managed sibling port owner must differ from the selected target`,
-          );
-        }
-        if (
-          (output.json !== undefined && output.json.owner_stack_id !== owner.ownerId) ||
-          (output.api !== undefined && output.api.ownerStackId !== owner.ownerId) ||
-          (output.human !== undefined && output.human.fields.ownerStackId !== owner.ownerId)
-        ) {
-          errors.push(`${scenario.id}: projected managed port owner must match ${owner.ownerId}`);
-        }
-      }
-    }
-
-    if (scenario.expected.error?.code === "STICKY_PORT_OCCUPIED") {
-      if (selection === undefined) {
-        errors.push(`${scenario.id}: sticky port conflict requires a selected target`);
-      } else {
-        const assignment = scenario.given.find(
-          (fact) =>
-            fact.kind === "port-assignment" &&
-            fact.stackId === selection.stackId &&
-            fact.intent === "automatic",
-        );
-        const occupiedPort =
-          assignment?.kind === "port-assignment"
-            ? scenario.given.find(
-                (fact) => fact.kind === "occupied-port" && fact.port === assignment.port,
-              )
-            : undefined;
-        if (
-          assignment?.kind !== "port-assignment" ||
-          occupiedPort?.kind !== "occupied-port" ||
-          output.json?.port !== assignment.port ||
-          output.json?.config_key !== assignment.key ||
-          output.json?.relocated !== false
-        ) {
-          errors.push(
-            `${scenario.id}: sticky port conflict must bind assignment, occupancy, and projections`,
-          );
-        }
-        if (
-          !scenario.given.some(
-            (fact) =>
-              fact.kind === "stack" &&
-              fact.stackId === selection.stackId &&
-              fact.lifecycle === "stopped",
-          )
-        ) {
-          errors.push(`${scenario.id}: sticky port conflict requires a stopped selected stack`);
-        }
-      }
-    }
-
-    const changedExactPort = scenario.given.find(
-      (fact) =>
-        fact.kind === "config-port" &&
-        fact.intent === "exact" &&
-        typeof fact.value === "number" &&
-        typeof fact.previousValue === "number",
-    );
-    const expectsExactPortChange =
-      (scenario.expected.outcome === "update" &&
-        typeof output.json?.previous_port === "number" &&
-        typeof output.json?.port === "number") ||
-      scenario.expected.warning?.code === "RUNNING_STACK_CONFIG_DRIFT";
-    if (expectsExactPortChange) {
-      let exactPortChangeMatches = false;
-      if (
-        changedExactPort?.kind === "config-port" &&
-        typeof changedExactPort.value === "number" &&
-        typeof changedExactPort.previousValue === "number"
-      ) {
-        const previousAssignment = scenario.given.find(
-          (fact) =>
-            fact.kind === "port-assignment" &&
-            fact.stackId === selection?.stackId &&
-            fact.key === changedExactPort.key &&
-            fact.port === changedExactPort.previousValue,
-        );
-        const updateProjectionMatches =
-          scenario.expected.outcome !== "update" ||
-          (output.json?.previous_port === changedExactPort.previousValue &&
-            output.json?.port === changedExactPort.value &&
-            output.human?.fields.apiUrl === `http://127.0.0.1:${changedExactPort.value}`);
-        const driftProjectionMatches =
-          scenario.expected.warning?.code !== "RUNNING_STACK_CONFIG_DRIFT" ||
-          (scenario.given.some(
-            (fact) =>
-              fact.kind === "stack" &&
-              fact.stackId === selection?.stackId &&
-              fact.lifecycle === "running",
-          ) &&
-            output.json?.config_key === changedExactPort.key &&
-            output.json?.running_port === changedExactPort.previousValue &&
-            output.json?.requested_port === changedExactPort.value &&
-            output.json?.drift === true &&
-            output.human?.fields.configKey === changedExactPort.key &&
-            output.human?.fields.runningPort === String(changedExactPort.previousValue) &&
-            output.human?.fields.configuredPort === String(changedExactPort.value) &&
-            output.human?.fields.drift === "true");
-        exactPortChangeMatches =
-          previousAssignment?.kind === "port-assignment" &&
-          previousAssignment.intent === "exact" &&
-          changedExactPort.value !== changedExactPort.previousValue &&
-          updateProjectionMatches &&
-          driftProjectionMatches;
-      }
-      if (!exactPortChangeMatches) {
-        errors.push(
-          `${scenario.id}: exact port change must bind previous assignment and requested value`,
-        );
-      }
-      if (scenario.expected.outcome === "update" && selection !== undefined) {
-        const persistenceIndex = scenario.expected.writes.findIndex(
-          (write) =>
-            write.target === "managed-state" &&
-            write.operation === "update" &&
-            write.id === selection.stackId,
-        );
-        const runtimeStartIndex = scenario.expected.writes.findIndex(
-          (write) =>
-            write.target === "runtime-state" &&
-            write.operation === "start" &&
-            write.id === selection.stackId,
-        );
-        if (persistenceIndex < 0 || runtimeStartIndex <= persistenceIndex) {
-          errors.push(
-            `${scenario.id}: exact port change must persist assignment before runtime start`,
-          );
-        }
-      }
-    }
-
-    const stickyAssignments =
-      scenario.expected.outcome === "reuse" && selection !== undefined
-        ? scenario.given.filter(
-            (fact): fact is ManagedStackContractPortAssignmentFact =>
-              fact.kind === "port-assignment" &&
-              fact.stackId === selection.stackId &&
-              fact.intent === "automatic" &&
-              scenario.given.some(
-                (config) =>
-                  config.kind === "config-port" &&
-                  config.key === fact.key &&
-                  config.intent === "automatic" &&
-                  config.source === "omitted",
-              ),
-          )
-        : [];
-    if (output.json?.sticky === true || stickyAssignments.length > 0) {
-      const projectedPorts = scenario.expected.output.json?.ports;
-      if (selection === undefined) {
-        errors.push(`${scenario.id}: sticky port reuse requires a selected target`);
-      } else if (
-        stickyAssignments.length === 0 ||
-        output.json?.sticky !== true ||
-        !isManagedStackContractRecord(projectedPorts) ||
-        stickyAssignments.some((assignment) => {
-          const service = assignment.key.endsWith(".port")
-            ? assignment.key.slice(0, -".port".length)
-            : assignment.key;
-          return (
-            projectedPorts?.[service] !== assignment.port ||
-            (service === "api" &&
-              output.human !== undefined &&
-              output.human.fields.apiUrl !== `http://127.0.0.1:${assignment.port}`)
-          );
-        })
-      ) {
-        errors.push(
-          `${scenario.id}: sticky port reuse must bind automatic config, assignment, and projections`,
-        );
-      }
-    }
-
-    if (output.api?.sticky === true && scenario.expected.outcome === "update") {
-      const targetStackId = selection?.stackId ?? explicitActionStackId;
-      const projectedPorts = output.api.ports;
-      const projectedIntents = output.api.intents;
-      if (
-        typeof targetStackId !== "string" ||
-        !isManagedStackContractRecord(projectedPorts) ||
-        !isManagedStackContractRecord(projectedIntents)
-      ) {
-        errors.push(`${scenario.id}: exact-to-automatic port transition requires projections`);
-      } else {
-        for (const [service, port] of Object.entries(projectedPorts)) {
-          const key = `${service}.port`;
-          const configPort = scenario.given.find(
-            (fact) => fact.kind === "config-port" && fact.key === key,
-          );
-          const previousAssignment = scenario.given.find(
-            (fact) =>
-              fact.kind === "port-assignment" && fact.stackId === targetStackId && fact.key === key,
-          );
-          if (
-            typeof port !== "number" ||
-            configPort?.kind !== "config-port" ||
-            configPort.intent !== "automatic" ||
-            configPort.source !== "omitted" ||
-            configPort.previousValue !== port ||
-            previousAssignment?.kind !== "port-assignment" ||
-            previousAssignment.intent !== "exact" ||
-            previousAssignment.port !== port ||
-            projectedIntents[service] !== "automatic"
-          ) {
-            errors.push(
-              `${scenario.id}: exact-to-automatic port transition must preserve the previous assignment`,
-            );
-          }
-        }
-      }
-    }
-
-    if (
-      scenario.area === "ports" &&
-      scenario.when.interface === "managed-api" &&
-      scenario.when.method === "startStack" &&
-      scenario.expected.outcome === "create"
-    ) {
-      const targetStackId = scenario.when.input.stackId;
-      const siblingAssignments = scenario.given.flatMap((fact) =>
-        fact.kind === "port-assignment" &&
-        typeof targetStackId === "string" &&
-        fact.stackId !== targetStackId
-          ? [fact]
-          : [],
-      );
-      const siblingStackIds = [...new Set(siblingAssignments.map(({ stackId }) => stackId))];
-      const projectedAvoidedStackIds = scenario.expected.details?.avoided_sibling_stack_ids;
-      const avoidedStackIds =
-        Array.isArray(projectedAvoidedStackIds) &&
-        projectedAvoidedStackIds.every((stackId): stackId is string => typeof stackId === "string")
-          ? projectedAvoidedStackIds
-          : undefined;
-      if (
-        (siblingStackIds.length > 0 || projectedAvoidedStackIds !== undefined) &&
-        (avoidedStackIds === undefined ||
-          !managedStackContractStringSetEquals(siblingStackIds, avoidedStackIds))
-      ) {
-        errors.push(`${scenario.id}: sibling allocation must bind all avoided stack IDs`);
-      }
-      const projectedPorts = scenario.expected.output.api?.ports;
-      if (!isManagedStackContractRecord(projectedPorts)) {
-        if (siblingAssignments.length > 0) {
-          errors.push(`${scenario.id}: sibling allocation must project its allocated ports`);
-        }
-      } else {
-        const occupiedSiblingPorts = new Set(siblingAssignments.map(({ port }) => port));
-        for (const port of Object.values(projectedPorts)) {
-          if (typeof port === "number") {
-            if (occupiedSiblingPorts.has(port)) {
-              errors.push(`${scenario.id}: allocated port ${port} conflicts with a sibling target`);
-            }
-          }
-        }
-      }
-    }
-
-    if (
-      scenario.when.interface === "managed-api" &&
-      scenario.when.method === "startStack" &&
-      isManagedStackContractRecord(scenario.when.input.portIntents)
-    ) {
-      const projectedPorts = scenario.expected.output.api?.ports;
-      const projectedIntents = scenario.expected.output.api?.intents;
-      const requestedEntries = Object.entries(scenario.when.input.portIntents);
-      const configPorts = scenario.given.filter((fact) => fact.kind === "config-port");
-      const actionStackId = scenario.when.input.stackId;
-      const createsManagedTarget =
-        typeof actionStackId === "string" &&
-        scenario.expected.writes.some(
-          (write) =>
-            write.target === "managed-state" &&
-            write.operation === "create" &&
-            write.id === actionStackId,
-        );
-      const requestsOmittedAutomaticPort = requestedEntries.some(
-        ([key, intent]) =>
-          intent === "automatic" &&
-          configPorts.some(
-            (fact) => fact.key === key && fact.intent === "automatic" && fact.source === "omitted",
-          ),
-      );
-      if (
-        createsManagedTarget &&
-        requestsOmittedAutomaticPort &&
-        (scenario.expected.details?.host_wide !== true || scenario.expected.details.sticky !== true)
-      ) {
-        errors.push(`${scenario.id}: fresh automatic ports must be host-wide sticky assignments`);
-      }
-      if (isManagedStackContractRecord(projectedPorts)) {
-        const allocatedPorts = new Set<number>();
-        for (const port of Object.values(projectedPorts)) {
-          if (typeof port !== "number") {
-            continue;
-          }
-          if (allocatedPorts.has(port)) {
-            errors.push(`${scenario.id}: allocated port ${port} is assigned more than once`);
-          }
-          allocatedPorts.add(port);
-        }
-      }
-      if (
-        !managedStackContractStringSetEquals(
-          requestedEntries.map(([key]) => key),
-          configPorts.map(({ key }) => key),
-        )
-      ) {
-        errors.push(`${scenario.id}: requested port keys must match their config facts`);
-      }
-      for (const [key, intent] of requestedEntries) {
-        const configPort = scenario.given.find(
-          (fact) => fact.kind === "config-port" && fact.key === key,
-        );
-        const service = key.endsWith(".port") ? key.slice(0, -".port".length) : key;
-        if (intent === "automatic") {
-          if (
-            configPort?.kind !== "config-port" ||
-            configPort.intent !== "automatic" ||
-            !isManagedStackContractRecord(projectedPorts) ||
-            typeof projectedPorts[service] !== "number" ||
-            !isManagedStackContractRecord(projectedIntents) ||
-            projectedIntents[service] !== "automatic"
-          ) {
-            errors.push(
-              `${scenario.id}: automatic port request ${key} must match its fact and projected allocation`,
-            );
-          }
-          continue;
-        }
-        if (!isManagedStackContractRecord(intent) || intent.intent !== "exact") {
-          errors.push(`${scenario.id}: port request ${key} has an invalid intent`);
-          continue;
-        }
-        const requestedPort = intent.port;
-        if (
-          typeof requestedPort !== "number" ||
-          configPort?.kind !== "config-port" ||
-          configPort.intent !== "exact" ||
-          configPort.value !== requestedPort
-        ) {
-          errors.push(`${scenario.id}: exact port request ${key} must match its config fact`);
-        } else if (
-          !isManagedStackContractRecord(projectedPorts) ||
-          projectedPorts[service] !== requestedPort
-        ) {
-          errors.push(
-            `${scenario.id}: projected exact port ${key} must match request ${requestedPort}`,
-          );
-        }
-        const projectedExactIntent =
-          requestedEntries.length === 1
-            ? output.api?.intent
-            : isManagedStackContractRecord(projectedIntents)
-              ? projectedIntents[service]
-              : undefined;
-        if (projectedExactIntent !== "exact") {
-          errors.push(`${scenario.id}: exact port request ${key} must project exact intent`);
-        }
-      }
-    }
-
-    if (scenario.expected.warning?.code === "RUNNING_STACK_RUNTIME_DRIFT") {
-      const persistedRuntime = scenario.given.find(
-        (fact) => fact.kind === "persisted-runtime" && fact.stackId === selection?.stackId,
-      );
-      const configuredRuntime = runtimeRequests.find(
-        (fact) => fact.source === "config" && fact.runtime !== "auto",
-      );
-      const projectedServices = output.json?.services;
-      if (
-        selection === undefined ||
-        !scenario.given.some(
-          (fact) =>
-            fact.kind === "stack" &&
-            fact.stackId === selection.stackId &&
-            fact.lifecycle === "running",
-        ) ||
-        persistedRuntime?.kind !== "persisted-runtime" ||
-        configuredRuntime?.kind !== "runtime-request" ||
-        persistedRuntime.runtime === configuredRuntime.runtime ||
-        output.human?.fields.runtime !== persistedRuntime.runtime ||
-        output.human.fields.configuredRuntime !== configuredRuntime.runtime ||
-        output.human.fields.drift !== "true" ||
-        output.json?.runtime !== persistedRuntime.runtime ||
-        output.json.configured_runtime !== configuredRuntime.runtime ||
-        output.json.drift !== true ||
-        scenario.expected.details?.mixed_runtime !== false ||
-        !isManagedStackContractRecord(projectedServices) ||
-        projectedServices.runtime !== persistedRuntime.runtime
-      ) {
-        errors.push(
-          `${scenario.id}: runtime drift must bind running stack, persisted runtime, config, and projections`,
-        );
-      }
-    }
-
-    const cliProjectsManagedResult =
-      scenario.when.interface === "cli" &&
-      scenario.when.argv[0] === "status" &&
-      scenario.given.some((fact) => fact.kind === "managed-api-options");
-    if (cliProjectsManagedResult) {
-      const managedRecord = scenario.given.find(
-        (fact) => fact.kind === "managed-record" && fact.stackId === selection?.stackId,
-      );
-      const selectedStack = scenario.given.find(
-        (fact) => fact.kind === "stack" && fact.stackId === selection?.stackId,
-      );
-      const persistedRuntime = scenario.given.find(
-        (fact) => fact.kind === "persisted-runtime" && fact.stackId === selection?.stackId,
-      );
-      if (
-        selection === undefined ||
-        managedRecord?.kind !== "managed-record" ||
-        managedRecord.status !== "active" ||
-        selectedStack?.kind !== "stack" ||
-        selectedStack.lifecycle !== "running" ||
-        persistedRuntime?.kind !== "persisted-runtime" ||
-        scenario.expected.details?.managed_result_projected !== true ||
-        scenario.expected.details.identity_decisions_in_cli !== 0 ||
-        output.human?.fields.stackId !== selection.stackId ||
-        output.human.fields.runtime !== persistedRuntime.runtime ||
-        output.json?.runtime !== persistedRuntime.runtime
-      ) {
-        errors.push(
-          `${scenario.id}: projected managed status requires an active running record and persisted runtime`,
-        );
-      }
-    }
-
-    if (scenario.expected.error?.code === "RUNTIME_CONFLICTS_WITH_PERSISTED_STACK") {
-      if (selection === undefined) {
-        errors.push(`${scenario.id}: persisted runtime conflict requires a selected target`);
-      } else {
-        const persistedRuntime = scenario.given.find(
-          (fact) => fact.kind === "persisted-runtime" && fact.stackId === selection.stackId,
-        );
-        const requestedRuntime =
-          explicitRuntime === "docker" || explicitRuntime === "native"
-            ? explicitRuntime
-            : runtimeRequests.find(
-                (fact) =>
-                  (fact.source === "cli" || fact.source === "managed-api") &&
-                  fact.runtime !== "auto",
-              )?.runtime;
-        if (persistedRuntime?.kind !== "persisted-runtime") {
-          errors.push(`${scenario.id}: persisted runtime must belong to the selected target`);
-        } else if (
-          requestedRuntime === undefined ||
-          persistedRuntime.runtime === requestedRuntime ||
-          output.human?.fields.persistedRuntime !== persistedRuntime.runtime ||
-          output.human?.fields.requestedRuntime !== requestedRuntime ||
-          output.json?.persisted_runtime !== persistedRuntime.runtime ||
-          output.json?.requested_runtime !== requestedRuntime ||
-          output.json?.code !== scenario.expected.error.code
-        ) {
-          errors.push(
-            `${scenario.id}: persisted runtime conflict must bind persisted and requested values`,
-          );
-        }
-      }
-    }
-
-    if (scenario.expected.error?.code === "RUNTIME_SELECTION_CONFLICT") {
-      const explicitRequest = runtimeRequests.find(
-        (fact) => fact.source === "cli" || fact.source === "managed-api",
-      );
-      const configRequest = runtimeRequests.find((fact) => fact.source === "config");
-      if (
-        explicitRequest?.kind !== "runtime-request" ||
-        configRequest?.kind !== "runtime-request" ||
-        explicitRequest.runtime === "auto" ||
-        configRequest.runtime === "auto" ||
-        explicitRequest.runtime === configRequest.runtime ||
-        output.json?.cli_runtime !== explicitRequest.runtime ||
-        output.json?.config_runtime !== configRequest.runtime ||
-        output.json?.code !== scenario.expected.error.code
-      ) {
-        errors.push(
-          `${scenario.id}: runtime conflict must bind different explicit and configured runtimes`,
-        );
-      }
-    }
-
-    if (
-      scenario.expected.error?.code === "DOCKER_UNAVAILABLE" ||
-      scenario.expected.error?.code === "NATIVE_UNAVAILABLE"
-    ) {
-      const unavailableRuntime = scenario.expected.error.code.startsWith("DOCKER")
-        ? "docker"
-        : "native";
-      const explicitRequest = runtimeRequests.find(
-        (fact) =>
-          (fact.source === "cli" || fact.source === "managed-api") &&
-          fact.runtime === unavailableRuntime,
-      );
-      const availability = scenario.given.find(
-        (fact) => fact.kind === "runtime-availability" && fact.runtime === unavailableRuntime,
-      );
-      const fallbackRuntime = unavailableRuntime === "docker" ? "native" : "docker";
-      const fallbackAvailability = scenario.given.find(
-        (fact) => fact.kind === "runtime-availability" && fact.runtime === fallbackRuntime,
-      );
-      if (
-        explicitRequest?.kind !== "runtime-request" ||
-        availability?.kind !== "runtime-availability" ||
-        availability.available ||
-        availability.reason === undefined ||
-        fallbackAvailability?.kind !== "runtime-availability" ||
-        !fallbackAvailability.available ||
-        output.json?.requested_runtime !== unavailableRuntime ||
-        output.json?.reason !== availability.reason ||
-        output.json?.fallback_attempted !== false ||
-        scenario.expected.details?.fallback_attempted !== false ||
-        scenario.expected.runtimeEffects.some((effect) => effect.operation === "start")
-      ) {
-        errors.push(
-          `${scenario.id}: explicit runtime error must bind an unavailable requested runtime`,
-        );
-      }
-    }
-
-    const legacySource = scenario.given.find((fact) => fact.kind === "legacy-state");
-    const absentManagedTarget = scenario.given.find(
-      (fact) => fact.kind === "managed-target" && fact.exists === false,
-    );
-    if (
-      scenario.expected.error?.code === "LEGACY_SOURCE_RUNNING" ||
-      (isManagedStartAction(scenario.when) &&
-        legacySource?.kind === "legacy-state" &&
-        legacySource.lifecycle === "running" &&
-        absentManagedTarget?.kind === "managed-target")
-    ) {
-      if (
-        legacySource?.kind !== "legacy-state" ||
-        legacySource.lifecycle !== "running" ||
-        absentManagedTarget?.kind !== "managed-target" ||
-        scenario.given.some(
-          (fact) =>
-            fact.kind === "managed-target" &&
-            absentManagedTarget?.kind === "managed-target" &&
-            fact.stackId === absentManagedTarget.stackId &&
-            fact.exists,
-        ) ||
-        scenario.expected.outcome !== "error" ||
-        scenario.expected.error?.code !== "LEGACY_SOURCE_RUNNING" ||
-        scenario.expected.writes.length > 0 ||
-        scenario.expected.runtimeEffects.length > 0
-      ) {
-        errors.push(
-          `${scenario.id}: running legacy error requires a running source and absent target`,
-        );
-      }
-      if (
-        scenario.expected.details?.legacy_source_stopped !== false ||
-        scenario.expected.details.managed_target_published !== false ||
-        scenario.expected.details.partial_state !== false ||
-        output.json?.legacy_source_stopped !== false ||
-        output.json.managed_target_published !== false
-      ) {
-        errors.push(`${scenario.id}: running legacy error must leave no partial managed target`);
-      }
-
-      const configuredPort = scenario.given.find(
-        (fact) => fact.kind === "config-port" && fact.intent === "exact",
-      );
-      const occupiedLegacyPort = scenario.given.find(
-        (fact) => fact.kind === "occupied-port" && fact.owner === "legacy-stack",
-      );
-      if (
-        (configuredPort !== undefined || occupiedLegacyPort !== undefined) &&
-        (configuredPort?.kind !== "config-port" ||
-          typeof configuredPort.value !== "number" ||
-          occupiedLegacyPort?.kind !== "occupied-port" ||
-          occupiedLegacyPort.port !== configuredPort.value ||
-          output.json?.port !== configuredPort.value ||
-          output.json.config_key !== configuredPort.key ||
-          scenario.expected.details?.allocation_attempted !== false ||
-          output.json.allocation_attempted !== false)
-      ) {
-        errors.push(
-          `${scenario.id}: running legacy port failure must bind config, occupancy, and projections`,
-        );
-      }
-    }
-
-    if (scenario.expected.error?.code === "LEGACY_BOOTSTRAP_FAILED") {
-      const rollbackStackId =
-        scenario.when.interface === "managed-api" && scenario.when.method === "startStack"
-          ? scenario.when.input.stackId
-          : undefined;
-      if (
-        scenario.when.interface !== "managed-api" ||
-        scenario.when.method !== "startStack" ||
-        scenario.when.input.injectCopyFailure !== true ||
-        !scenario.expected.writes.some(
-          (write) =>
-            write.target === "managed-state" &&
-            write.operation === "delete" &&
-            write.id === rollbackStackId,
-        ) ||
-        !scenario.expected.runtimeEffects.some(
-          (effect) => effect.operation === "delete" && effect.stackId === rollbackStackId,
-        )
-      ) {
-        errors.push(`${scenario.id}: bootstrap rollback requires enabled copy-failure injection`);
-      }
-      if (
-        typeof rollbackStackId !== "string" ||
-        !scenario.given.some(
-          (fact) =>
-            fact.kind === "managed-target" &&
-            fact.stackId === rollbackStackId &&
-            fact.exists === false,
-        ) ||
-        scenario.given.some((fact) => fact.kind === "stack" && fact.stackId === rollbackStackId)
-      ) {
-        errors.push(
-          `${scenario.id}: bootstrap rollback requires failure injection against an absent target`,
-        );
-      }
-      const legacySource = scenario.given.find((fact) => fact.kind === "legacy-state");
-      if (
-        legacySource?.kind !== "legacy-state" ||
-        legacySource.lifecycle !== "stopped" ||
-        legacySource.database !== "compatible" ||
-        legacySource.storage !== "compatible" ||
-        legacySource.credentials !== "compatible"
-      ) {
-        errors.push(
-          `${scenario.id}: bootstrap rollback requires a compatible stopped legacy source`,
-        );
-      }
-      if (
-        scenario.expected.details?.active_target_exists !== false ||
-        scenario.expected.details.registry_record_published !== false ||
-        output.api?.activeTargetExists !== false ||
-        output.api?.registryRecordPublished !== false ||
-        output.api?.retryable !== true ||
-        scenario.expected.writes.some(
-          (write) =>
-            (write.target === "managed-state" && write.operation !== "delete") ||
-            (write.target === "registry" && write.operation === "publish"),
-        )
-      ) {
-        errors.push(`${scenario.id}: bootstrap rollback must leave no published partial target`);
-      }
-    }
-
-    const destructivelyDeletesManagedState =
-      scenario.expected.writes.some(
-        (write) => write.target === "managed-state" && write.operation === "delete",
-      ) || scenario.expected.runtimeEffects.some((effect) => effect.operation === "delete");
-    const nonDestructiveStop =
-      scenario.when.interface === "cli" &&
-      scenario.when.argv[0] === "stop" &&
-      scenario.expected.outcome === "update" &&
-      scenario.expected.runtimeEffects.some((effect) => effect.operation === "stop") &&
-      !destructivelyDeletesManagedState &&
-      !scenario.expected.writes.some(
-        (write) =>
-          write.target === "registry" &&
-          (write.operation === "delete" || write.operation === "tombstone"),
-      );
-    if (
-      nonDestructiveStop &&
-      (scenario.expected.details?.data_preserved !== true ||
-        scenario.expected.details.registry_record_preserved !== true ||
-        output.human?.fields.dataPreserved !== "true" ||
-        output.json?.data_preserved !== true)
-    ) {
-      errors.push(`${scenario.id}: non-destructive stop must report preserved data and registry`);
-    }
-    if (
-      scenario.expected.details?.legacy_stack_stopped !== undefined ||
-      output.json?.legacy_stack_stopped !== undefined
-    ) {
-      const runningLegacyState = scenario.given.find(
-        (fact) => fact.kind === "legacy-state" && fact.lifecycle === "running",
-      );
-      if (
-        !nonDestructiveStop ||
-        runningLegacyState?.kind !== "legacy-state" ||
-        scenario.expected.details?.managed_stack_stopped !== true ||
-        scenario.expected.details.legacy_stack_stopped !== false ||
-        scenario.expected.details.legacy_state_mutated !== false ||
-        output.json?.managed_stack_stopped !== true ||
-        output.json.legacy_stack_stopped !== false
-      ) {
-        errors.push(
-          `${scenario.id}: engine-scoped stop requires a simultaneously running legacy stack`,
-        );
-      }
-    }
-    if (
-      destructivelyDeletesManagedState &&
-      scenario.when.interface === "cli" &&
-      scenario.when.argv[0] === "stop" &&
-      !scenario.when.argv.includes("--no-backup")
-    ) {
-      errors.push(`${scenario.id}: destructive stop requires --no-backup`);
-    }
-
-    const tombstonedTarget = scenario.given.find(
-      (fact) =>
-        fact.kind === "managed-record" &&
-        fact.status === "tombstoned" &&
-        fact.stackId === explicitActionStackId,
-    );
-    const claimsIdempotentDeletion =
-      scenario.expected.details?.idempotent === true || output.json?.already_deleted === true;
-    if (tombstonedTarget?.kind === "managed-record" || claimsIdempotentDeletion) {
-      if (
-        tombstonedTarget?.kind !== "managed-record" ||
-        scenario.expected.outcome !== "no-op" ||
-        scenario.expected.details?.tombstoned !== true ||
-        scenario.expected.details?.idempotent !== true ||
-        output.json?.tombstoned !== true ||
-        output.json?.already_deleted !== true ||
-        scenario.expected.writes.length > 0 ||
-        scenario.expected.runtimeEffects.length > 0
-      ) {
-        errors.push(`${scenario.id}: idempotent deletion requires a tombstoned target`);
-      }
-    }
-
-    const globallyTargetedStack = scenario.given.find(
-      (fact) => fact.kind === "stack" && fact.stackId === explicitActionStackId,
-    );
-    const isCheckoutIndependentStackDeletion =
-      scenario.when.interface === "cli" &&
-      scenario.when.argv[0] === "stop" &&
-      scenario.when.argv.includes("--stack-id") &&
-      scenario.expected.outcome === "delete" &&
-      !scenario.given.some((fact) => fact.kind === "checkout" && fact.path === actionCwd);
-    if (isCheckoutIndependentStackDeletion) {
-      if (
-        typeof explicitActionStackId !== "string" ||
-        globallyTargetedStack?.kind !== "stack" ||
-        globallyTargetedStack.orphaned !== true ||
-        output.json?.orphaned !== true ||
-        output.human?.fields.orphaned !== "true" ||
-        scenario.expected.details?.checkout_required !== false
-      ) {
-        errors.push(`${scenario.id}: global orphan deletion requires an orphaned target`);
-      }
-    }
-
-    if (
-      scenario.given.some((fact) => fact.kind === "credential-state") &&
-      scenario.expected.writes.some(
-        (write) =>
-          write.target === "managed-state" &&
-          (write.operation === "copy" ||
-            write.operation === "create" ||
-            write.operation === "update"),
-      ) &&
-      scenario.expected.details?.plaintext_secrets_in_global_state !== false
-    ) {
-      errors.push(`${scenario.id}: credential persistence must not expose plaintext globally`);
-    }
-
-    const changedCredentials = scenario.given.find(
-      (fact) => fact.kind === "credential-state" && fact.previousValuesId !== undefined,
-    );
-    const projectedCredentialReference = scenario.expected.details?.credential_values_id;
-    if (
-      typeof projectedCredentialReference === "string" &&
-      scenario.given.some((fact) => fact.kind === "credential-state")
-    ) {
-      const credentialState = scenario.given.find(
-        (fact) =>
-          fact.kind === "credential-state" && fact.valuesId === projectedCredentialReference,
-      );
-      const projectedSources = [
-        scenario.expected.details?.source,
-        output.api?.credentialsSource,
-        output.json?.credentials_source,
-      ].filter((source) => source !== undefined);
-      if (
-        credentialState?.kind !== "credential-state" ||
-        (output.api?.credentialsValuesId !== undefined &&
-          output.api.credentialsValuesId !== credentialState.valuesId) ||
-        (output.json?.credentials_values_id !== undefined &&
-          output.json.credentials_values_id !== credentialState.valuesId) ||
-        projectedSources.some((source) => source !== credentialState.source)
-      ) {
-        errors.push(
-          `${scenario.id}: persisted credential reference and source must match declared values`,
-        );
-      }
-    }
-    const stableLocalCredentials = scenario.given.find(
-      (fact) => fact.kind === "credential-state" && fact.source === "local-default",
-    );
-    const projectsStableLocalCredentials =
-      scenario.expected.details?.generated_per_start !== undefined ||
-      output.json?.credentials_source === "local-default" ||
-      output.json?.credentials_stable !== undefined;
-    if (
-      (stableLocalCredentials?.kind === "credential-state" || projectsStableLocalCredentials) &&
-      (stableLocalCredentials?.kind !== "credential-state" ||
-        scenario.expected.details?.credential_values_id !== stableLocalCredentials.valuesId ||
-        scenario.expected.details.generated_per_start !== false ||
-        output.json?.credentials_source !== "local-default" ||
-        output.json?.credentials_stable !== true)
-    ) {
-      errors.push(`${scenario.id}: omitted credentials must reuse stable local defaults`);
-    }
-    const persistedCredentials = scenario.given.find(
-      (fact) => fact.kind === "credential-state" && fact.source === "persisted",
-    );
-    if (
-      persistedCredentials?.kind === "credential-state" &&
-      (scenario.expected.details?.credential_values_id !== persistedCredentials.valuesId ||
-        scenario.expected.details.credentials_rotated !== false ||
-        output.json?.credentials_unchanged !== true)
-    ) {
-      errors.push(`${scenario.id}: persisted credentials must survive restart unchanged`);
-    }
-    if (
-      changedCredentials?.kind === "credential-state" &&
-      changedCredentials.previousValuesId === changedCredentials.valuesId
-    ) {
-      errors.push(`${scenario.id}: credential change requires different old and new values`);
-    }
-    const configuredCredentialChange =
-      changedCredentials?.kind === "credential-state" && changedCredentials.source === "configured"
-        ? changedCredentials
-        : undefined;
-    const selectedStoppedStack = scenario.given.find(
-      (fact) =>
-        fact.kind === "stack" &&
-        fact.stackId === selection?.stackId &&
-        fact.lifecycle === "stopped",
-    );
-    const projectsCredentialUpdate =
-      scenario.expected.outcome === "update" &&
-      (output.json?.previous_credentials_values_id !== undefined ||
-        output.json?.credentials_values_id !== undefined);
-    const appliesConfiguredCredentialChange =
-      configuredCredentialChange !== undefined && selectedStoppedStack?.kind === "stack";
-    if (projectsCredentialUpdate && configuredCredentialChange === undefined) {
-      errors.push(`${scenario.id}: credential change requires configured old and new values`);
-    }
-    if (
-      (projectsCredentialUpdate || appliesConfiguredCredentialChange) &&
-      configuredCredentialChange !== undefined &&
-      (selectedStoppedStack?.kind !== "stack" ||
-        scenario.expected.outcome !== "update" ||
-        output.json?.previous_credentials_values_id !==
-          configuredCredentialChange.previousValuesId ||
-        output.json?.credentials_values_id !== configuredCredentialChange.valuesId ||
-        !scenario.expected.writes.some(
-          (write) => write.target === "managed-state" && write.operation === "update",
-        ))
-    ) {
-      errors.push(`${scenario.id}: credential update must bind old and new persisted references`);
-    }
-
-    if (scenario.expected.warning?.code === "RUNNING_STACK_CREDENTIALS_DRIFT") {
-      if (
-        selection === undefined ||
-        changedCredentials?.kind !== "credential-state" ||
-        !scenario.given.some(
-          (fact) =>
-            fact.kind === "stack" &&
-            fact.stackId === selection.stackId &&
-            fact.lifecycle === "running",
-        ) ||
-        scenario.expected.outcome !== "report" ||
-        scenario.expected.writes.length > 0 ||
-        scenario.expected.runtimeEffects.length > 0 ||
-        output.json?.stack_id !== selection.stackId ||
-        output.json?.drift !== true ||
-        output.human?.fields.stackId !== selection.stackId ||
-        output.human?.fields.drift !== "true"
-      ) {
-        errors.push(`${scenario.id}: credential drift report requires a running selected stack`);
-      }
-    }
-
-    const legacyCredentials = scenario.given.find(
-      (fact) => fact.kind === "credential-state" && fact.source === "legacy",
-    );
-    const copiesManagedCredentials = scenario.expected.writes.some(
-      (write) => write.target === "managed-state" && write.operation === "copy",
-    );
-    if (
-      copiesManagedCredentials &&
-      (legacyCredentials?.kind === "credential-state" ||
-        typeof scenario.expected.details?.credential_values_id === "string" ||
-        typeof output.api?.credentialsValuesId === "string") &&
-      (legacyCredentials?.kind !== "credential-state" ||
-        scenario.expected.details?.credential_values_id !== legacyCredentials.valuesId ||
-        output.api?.credentialsValuesId !== legacyCredentials.valuesId)
-    ) {
-      errors.push(`${scenario.id}: copied legacy credentials must bind their persisted reference`);
-    }
-
-    if (scenario.expected.details?.retry_after_rollback === true) {
-      const retryStackId =
-        scenario.when.interface === "managed-api" && scenario.when.method === "startStack"
-          ? scenario.when.input.stackId
-          : undefined;
-      if (
-        typeof retryStackId !== "string" ||
-        !scenario.given.some(
-          (fact) =>
-            fact.kind === "operation-result" &&
-            fact.operation === "legacy-bootstrap" &&
-            fact.stackId === retryStackId &&
-            fact.outcome === "rolled-back",
-        )
-      ) {
-        errors.push(`${scenario.id}: bootstrap retry requires a rolled-back prior attempt`);
-      }
-    }
-
-    if (
-      scenario.when.interface === "cli" &&
-      scenario.when.argv[0] === "stack" &&
-      scenario.when.argv[1] === "prune"
-    ) {
-      if (
-        scenario.expected.details?.mutable_data_deleted !== false ||
-        output.human?.fields.dataDeleted !== "false" ||
-        output.json?.mutable_data_deleted !== false ||
-        scenario.expected.writes.some((write) => write.target === "managed-state") ||
-        scenario.expected.runtimeEffects.some((effect) => effect.operation === "delete")
-      ) {
-        errors.push(`${scenario.id}: prune must preserve mutable stack data`);
-      }
-      for (const write of scenario.expected.writes) {
-        if (write.target !== "registry" || write.operation !== "delete") {
-          continue;
-        }
-        const mutableDataExists = scenario.given.some(
-          (fact) =>
-            (fact.kind === "stack" && fact.stackId === write.id) ||
-            (fact.kind === "managed-target" && fact.stackId === write.id && fact.exists),
-        );
-        if (!mutableDataExists) {
-          errors.push(`${scenario.id}: data-preserving prune must declare mutable stack data`);
-        }
-        const orphanedRecord = scenario.given.some(
-          (fact) =>
-            fact.kind === "managed-record" &&
-            fact.stackId === write.id &&
-            fact.status === "orphaned",
-        );
-        const orphanedStack = scenario.given.some(
-          (fact) => fact.kind === "stack" && fact.stackId === write.id && fact.orphaned === true,
-        );
-        if (!orphanedRecord || !orphanedStack) {
-          errors.push(`${scenario.id}: prune may delete only orphaned registry metadata`);
-        }
-      }
-      const deletedRecordIds = scenario.expected.writes.flatMap((write) =>
-        write.target === "registry" && write.operation === "delete" ? [write.id] : [],
-      );
-      const projectedRecords = output.json?.pruned_records;
-      const projectedRecordIds =
-        Array.isArray(projectedRecords) &&
-        projectedRecords.every((record): record is string => typeof record === "string")
-          ? projectedRecords
-          : undefined;
-      const deletedRecordCount = deletedRecordIds.length;
-      const expectedSummary = `Pruned ${deletedRecordCount} orphaned metadata record${deletedRecordCount === 1 ? "" : "s"}`;
-      if (
-        deletedRecordCount === 0 ||
-        projectedRecordIds === undefined ||
-        !managedStackContractStringSetEquals(deletedRecordIds, projectedRecordIds) ||
-        output.json?.pruned_count !== deletedRecordCount ||
-        output.human?.summary !== expectedSummary
-      ) {
-        errors.push(`${scenario.id}: prune projections must match deleted registry records`);
-      }
-    }
-
-    const activeBranchName =
-      actionGitState?.kind === "git-state" && actionGitState.head === "branch"
-        ? actionGitState.branch
-        : checkedOutBranch?.kind === "branch"
-          ? checkedOutBranch.name
-          : undefined;
-
-    for (const write of scenario.expected.writes) {
-      if (write.target !== "git-config") {
-        continue;
-      }
-      if (write.operation === "create" && existingCheckoutIdentityIds.has(write.id)) {
-        errors.push(`${scenario.id}: Git identity ${write.id} is already declared`);
-      }
-      const expectedScope = projectIds.has(write.id)
-        ? "common"
-        : checkoutIds.has(write.id) || contextIds.has(write.id)
-          ? "worktree"
-          : undefined;
-      if (expectedScope !== undefined && write.scope !== expectedScope) {
-        errors.push(
-          `${scenario.id}: Git identity ${write.id} must use ${expectedScope} config scope`,
-        );
-      }
-      if (contextIds.has(write.id)) {
-        if (write.owner === undefined) {
-          errors.push(`${scenario.id}: Git context ${write.id} must declare its branch owner`);
-        } else if (activeBranchName !== undefined && write.owner !== activeBranchName) {
-          errors.push(
-            `${scenario.id}: Git context ${write.id} must belong to branch ${activeBranchName}`,
-          );
-        }
-      }
-    }
-
-    for (const effect of scenario.expected.runtimeEffects) {
-      if (!declaredIds.has(effect.stackId)) {
-        errors.push(`${scenario.id}: runtime effect references undeclared ID ${effect.stackId}`);
-      }
-
-      if (
-        effect.operation === "stop" &&
-        !scenario.given.some(
-          (fact) =>
-            fact.kind === "stack" &&
-            fact.stackId === effect.stackId &&
-            fact.lifecycle === "running",
-        )
-      ) {
-        errors.push(
-          `${scenario.id}: stopping stack ${effect.stackId} requires a running lifecycle`,
-        );
-      }
-
-      const hasWrite = scenario.expected.writes.some((write) => {
-        if (write.id !== effect.stackId) {
-          return false;
-        }
-        switch (effect.operation) {
-          case "copy":
-            return write.target === "managed-state" && write.operation === "copy";
-          case "delete":
-            return write.target === "managed-state" && write.operation === "delete";
-          case "start":
-            return write.target === "runtime-state" && write.operation === "start";
-          case "stop":
-            return (
-              write.target === "runtime-state" &&
-              (write.operation === "delete" || write.operation === "update")
-            );
-        }
-      });
-      if (!hasWrite) {
-        errors.push(
-          `${scenario.id}: ${effect.operation} runtime effect requires a matching state write`,
-        );
-      }
-    }
-
-    for (const write of scenario.expected.writes) {
-      if (
-        write.target === "managed-state" &&
-        (write.operation === "copy" || write.operation === "create") &&
-        !scenario.expected.writes.some(
-          (candidate) =>
-            candidate.target === "registry" &&
-            candidate.operation === "publish" &&
-            candidate.id === write.id,
-        )
-      ) {
-        errors.push(
-          `${scenario.id}: managed-state ${write.operation} requires registry publication`,
-        );
-      }
-
-      if (
-        write.target === "registry" &&
-        write.operation === "publish" &&
-        !scenario.expected.writes.some(
-          (candidate) =>
-            candidate.target === "managed-state" &&
-            (candidate.operation === "create" || candidate.operation === "copy") &&
-            candidate.id === write.id,
-        )
-      ) {
-        errors.push(`${scenario.id}: registry publication requires managed-state creation or copy`);
-      }
-
-      if (
-        scenario.expected.outcome === "delete" &&
-        write.target === "managed-state" &&
-        write.operation === "delete" &&
-        !scenario.expected.writes.some(
-          (candidate) =>
-            candidate.target === "registry" &&
-            candidate.operation === "tombstone" &&
-            candidate.id === write.id,
-        )
-      ) {
-        errors.push(`${scenario.id}: managed-state deletion requires a registry tombstone`);
-      }
-
-      if (
-        write.target === "registry" &&
-        write.operation === "tombstone" &&
-        !scenario.expected.writes.some(
-          (candidate) =>
-            candidate.target === "managed-state" &&
-            candidate.operation === "delete" &&
-            candidate.id === write.id,
-        )
-      ) {
-        errors.push(`${scenario.id}: registry tombstone requires managed-state deletion`);
-      }
-      if (
-        write.target === "registry" &&
-        write.operation === "tombstone" &&
-        (scenario.expected.details?.tombstoned !== true ||
-          (output.json !== undefined && output.json.tombstoned !== true) ||
-          (output.human !== undefined && output.human.fields.tombstoned !== "true"))
-      ) {
-        errors.push(`${scenario.id}: registry tombstone must be reported by every projection`);
-      }
-      if (
-        write.target === "registry" &&
-        write.operation === "delete" &&
-        scenario.expected.details?.metadata_removed !== true
-      ) {
-        errors.push(`${scenario.id}: registry deletion must report removed metadata`);
-      }
-
-      const requiredRuntimeOperation =
-        write.target === "runtime-state" && write.operation === "start"
-          ? "start"
-          : write.target === "runtime-state" &&
-              (write.operation === "delete" || write.operation === "update")
-            ? "stop"
-            : write.target === "managed-state" && write.operation === "copy"
-              ? "copy"
-              : write.target === "managed-state" && write.operation === "delete"
-                ? "delete"
-                : undefined;
-      if (
-        requiredRuntimeOperation !== undefined &&
-        !scenario.expected.runtimeEffects.some(
-          (effect) => effect.operation === requiredRuntimeOperation && effect.stackId === write.id,
-        )
-      ) {
-        errors.push(
-          `${scenario.id}: ${write.target} ${write.operation} requires a matching runtime effect`,
-        );
-      }
-    }
-
-    const checkProjection = (
-      projection: Readonly<Record<string, ManagedStackContractJson>> | undefined,
-      key: string,
-      expected: ManagedStackContractJson,
-    ): void => {
-      if (projection?.[key] !== undefined && projection[key] !== expected) {
-        errors.push(`${scenario.id}: projected ${key} disagrees with the managed result`);
-      }
-    };
-
-    if (
-      scenario.expected.output.json !== undefined &&
-      scenario.expected.output.json.outcome === undefined
-    ) {
-      errors.push(`${scenario.id}: JSON projection requires an outcome`);
-    }
-    const structuredCode = scenario.expected.error?.code ?? scenario.expected.warning?.code;
-    if (
-      scenario.expected.output.json !== undefined &&
-      structuredCode !== undefined &&
-      scenario.expected.output.json.code === undefined
-    ) {
-      errors.push(`${scenario.id}: JSON projection requires a code`);
-    }
-    for (const projection of [scenario.expected.output.json, scenario.expected.output.api]) {
-      checkProjection(projection, "outcome", scenario.expected.outcome);
-      if (scenario.expected.error !== undefined) {
-        checkProjection(projection, "code", scenario.expected.error.code);
-      }
-      if (scenario.expected.warning !== undefined) {
-        checkProjection(projection, "code", scenario.expected.warning.code);
-      }
-    }
-    if (selection !== undefined) {
-      checkProjection(scenario.expected.output.json, "project_id", selection.projectId);
-      checkProjection(scenario.expected.output.json, "checkout_id", selection.checkoutId);
-      checkProjection(scenario.expected.output.json, "context_id", selection.contextId);
-      checkProjection(scenario.expected.output.json, "stack_id", selection.stackId);
-      checkProjection(scenario.expected.output.json, "stack_name", selection.stackName);
-      checkProjection(scenario.expected.output.api, "projectId", selection.projectId);
-      checkProjection(scenario.expected.output.api, "checkoutId", selection.checkoutId);
-      checkProjection(scenario.expected.output.api, "contextId", selection.contextId);
-      checkProjection(scenario.expected.output.api, "stackId", selection.stackId);
-      checkProjection(scenario.expected.output.api, "stackName", selection.stackName);
-      checkProjection(scenario.expected.output.human?.fields, "projectId", selection.projectId);
-      checkProjection(scenario.expected.output.human?.fields, "checkoutId", selection.checkoutId);
-      checkProjection(scenario.expected.output.human?.fields, "contextId", selection.contextId);
-      checkProjection(scenario.expected.output.human?.fields, "stackId", selection.stackId);
-      checkProjection(scenario.expected.output.human?.fields, "stack", selection.stackName);
-      checkProjection(scenario.expected.output.human?.fields, "stackName", selection.stackName);
-
-      const selectedBranch = scenario.given.find(
-        (fact) =>
-          fact.kind === "branch" && fact.checkedOut && fact.contextId === selection.contextId,
-      );
-      if (selectedBranch?.kind === "branch") {
-        checkProjection(scenario.expected.output.human?.fields, "branch", selectedBranch.name);
-      }
-    }
-
-    const expectedRecovery =
-      scenario.expected.error?.recovery ?? scenario.expected.warning?.recovery;
-    const humanOutput = scenario.expected.output.human;
-    if (
-      humanOutput !== undefined &&
-      expectedRecovery !== undefined &&
-      (humanOutput.recovery === undefined ||
-        humanOutput.recovery.length !== expectedRecovery.length ||
-        humanOutput.recovery.some((step, index) => step !== expectedRecovery[index]))
-    ) {
-      errors.push(`${scenario.id}: human recovery disagrees with the managed result`);
-    }
-    const jsonOutput = scenario.expected.output.json;
-    const jsonRecovery = jsonOutput?.recovery;
-    if (
-      jsonOutput !== undefined &&
-      expectedRecovery !== undefined &&
-      (!Array.isArray(jsonRecovery) ||
-        jsonRecovery.length !== expectedRecovery.length ||
-        jsonRecovery.some((step, index) => step !== expectedRecovery[index]))
-    ) {
-      errors.push(`${scenario.id}: JSON recovery disagrees with the managed result`);
-    }
-  }
-
-  return errors;
-};
 
 const branchHistoryFixture = (
   label: "commit" | "rebase" | "reset",
@@ -7575,6 +3655,38 @@ const additionalLifecycleContractFixtures = defineManagedStackContractFixtures([
           stackId: "stack-main-default",
           bootstrap: "copied",
           credentialsValuesId: "legacy-auth-v1",
+        },
+      },
+    },
+  },
+  {
+    id: "credentials.plaintext-secrets-stay-out-of-global-state",
+    title: "Resolved plaintext secrets are absent from the global managed registry",
+    area: "credentials",
+    given: [
+      {
+        kind: "credential-state",
+        source: "persisted",
+        valuesId: "configured-auth-v1",
+        plaintextPresentInGlobalState: false,
+      },
+      { kind: "managed-record", stackId: "stack-main-default", status: "active" },
+    ],
+    when: {
+      interface: "managed-api",
+      method: "inspectGlobalRecord",
+      input: { stackId: "stack-main-default" },
+    },
+    expected: {
+      outcome: "report",
+      writes: [],
+      runtimeEffects: [],
+      details: { plaintext_secrets_present: false },
+      output: {
+        api: {
+          stackId: "stack-main-default",
+          credentialsReference: "configured-auth-v1",
+          plaintextSecrets: [],
         },
       },
     },
